Split stack-app.ts file into multiple files
