--- packages/template/src/lib/stack-app.ts
@@ -1,3744 +1,0 @@
-import { WebAuthnError, startAuthentication, startRegistration } from "@simplewebauthn/browser";
-import { KnownErrors, StackAdminInterface, StackClientInterface, StackServerInterface } from "@stackframe/stack-shared";
-import { ProductionModeError, getProductionModeErrors } from "@stackframe/stack-shared/dist/helpers/production-mode";
-import { ApiKeyCreateCrudRequest, ApiKeyCreateCrudResponse } from "@stackframe/stack-shared/dist/interface/adminInterface";
-import { ApiKeysCrud } from "@stackframe/stack-shared/dist/interface/crud/api-keys";
-import { ContactChannelsCrud } from "@stackframe/stack-shared/dist/interface/crud/contact-channels";
-import { CurrentUserCrud } from "@stackframe/stack-shared/dist/interface/crud/current-user";
-import { EmailTemplateCrud, EmailTemplateType } from "@stackframe/stack-shared/dist/interface/crud/email-templates";
-import { InternalProjectsCrud, ProjectsCrud } from "@stackframe/stack-shared/dist/interface/crud/projects";
-import { TeamInvitationCrud } from "@stackframe/stack-shared/dist/interface/crud/team-invitation";
-import { TeamMemberProfilesCrud } from "@stackframe/stack-shared/dist/interface/crud/team-member-profiles";
-import { TeamPermissionDefinitionsCrud, TeamPermissionsCrud } from "@stackframe/stack-shared/dist/interface/crud/team-permissions";
-import { TeamsCrud } from "@stackframe/stack-shared/dist/interface/crud/teams";
-import { UsersCrud } from "@stackframe/stack-shared/dist/interface/crud/users";
-import { InternalSession } from "@stackframe/stack-shared/dist/sessions";
-import { encodeBase64 } from "@stackframe/stack-shared/dist/utils/bytes";
-import { AsyncCache } from "@stackframe/stack-shared/dist/utils/caches";
-import { scrambleDuringCompileTime } from "@stackframe/stack-shared/dist/utils/compile-time";
-import { isBrowserLike } from "@stackframe/stack-shared/dist/utils/env";
-import { StackAssertionError, concatStacktraces, throwErr } from "@stackframe/stack-shared/dist/utils/errors";
-import { ReadonlyJson } from "@stackframe/stack-shared/dist/utils/json";
-import { DependenciesMap } from "@stackframe/stack-shared/dist/utils/maps";
-import { ProviderType } from "@stackframe/stack-shared/dist/utils/oauth";
-import { deepPlainEquals, filterUndefined, omit, pick } from "@stackframe/stack-shared/dist/utils/objects";
-import { ReactPromise, neverResolve, runAsynchronously, wait } from "@stackframe/stack-shared/dist/utils/promises";
-import { suspend, suspendIfSsr } from "@stackframe/stack-shared/dist/utils/react";
-import { Result } from "@stackframe/stack-shared/dist/utils/results";
-import { Store, storeLock } from "@stackframe/stack-shared/dist/utils/stores";
-import { mergeScopeStrings } from "@stackframe/stack-shared/dist/utils/strings";
-import { getRelativePart, isRelative } from "@stackframe/stack-shared/dist/utils/urls";
-import { generateUuid } from "@stackframe/stack-shared/dist/utils/uuids";
-import * as cookie from "cookie";
-// NEXT_LINE_PLATFORM next
-import * as NextNavigationUnscrambled from "next/navigation"; // import the entire module to get around some static compiler warnings emitted by Next.js in some cases
-// NEXT_LINE_PLATFORM react-like
-import React, { useCallback, useMemo } from "react";
-import { constructRedirectUrl } from "../utils/url";
-import { addNewOAuthProviderOrScope, callOAuthCallback, signInWithOAuth } from "./auth";
-import { CookieHelper, createBrowserCookieHelper, createCookieHelper, createPlaceholderCookieHelper, deleteCookieClient, getCookieClient, setOrDeleteCookie, setOrDeleteCookieClient } from "./cookie";
-
-let isReactServer = false;
-// IF_PLATFORM next
-import * as sc from "@stackframe/stack-sc";
-import { cookies } from '@stackframe/stack-sc';
-isReactServer = sc.isReactServer;
-// END_PLATFORM
-
-// IF_PLATFORM next
-// NextNavigation.useRouter does not exist in react-server environments and some bundlers try to be helpful and throw a warning. Ignore the warning.
-const NextNavigation = scrambleDuringCompileTime(NextNavigationUnscrambled);
-// END_PLATFORM
-
-const clientVersion = "STACK_COMPILE_TIME_CLIENT_PACKAGE_VERSION_SENTINEL";
-if (clientVersion.startsWith("STACK_COMPILE_TIME")) {
-  throw new StackAssertionError("Client version was not replaced. Something went wrong during build!");
-}
-
-// hack to make sure process is defined in non-node environments
-// NEXT_LINE_PLATFORM js react
-const process = (globalThis as any).process ?? { env: {} };
-
-type RequestLike = {
-  headers: {
-    get: (name: string) => string | null,
-  },
-};
-
-type RedirectMethod = "window"
-// NEXT_LINE_PLATFORM next
-| "nextjs"
-| "none"
-| {
-  useNavigate: () => (to: string) => void,
-  navigate?: (to: string) => void,
-}
-
-export type TokenStoreInit<HasTokenStore extends boolean = boolean> =
-  HasTokenStore extends true ? (
-    | "cookie"
-    | "nextjs-cookie"
-    | "memory"
-    | RequestLike
-    | { accessToken: string, refreshToken: string }
-  )
-  : HasTokenStore extends false ? null
-  : TokenStoreInit<true> | TokenStoreInit<false>;
-
-export type HandlerUrls = {
-  handler: string,
-  signIn: string,
-  signUp: string,
-  afterSignIn: string,
-  afterSignUp: string,
-  signOut: string,
-  afterSignOut: string,
-  emailVerification: string,
-  passwordReset: string,
-  forgotPassword: string,
-  home: string,
-  oauthCallback: string,
-  magicLinkCallback: string,
-  accountSettings: string,
-  teamInvitation: string,
-  error: string,
-}
-
-export type OAuthScopesOnSignIn = {
-  [key in ProviderType]: string[];
-};
-
-
-type ProjectCurrentUser<ProjectId> = ProjectId extends "internal" ? CurrentInternalUser : CurrentUser;
-type ProjectCurrentServerUser<ProjectId> = ProjectId extends "internal" ? CurrentInternalServerUser : CurrentServerUser;
-
-function getUrls(partial: Partial<HandlerUrls>): HandlerUrls {
-  const handler = partial.handler ?? "/handler";
-  const home = partial.home ?? "/";
-  const afterSignIn = partial.afterSignIn ?? home;
-  return {
-    handler,
-    signIn: `${handler}/sign-in`,
-    afterSignIn: home,
-    signUp: `${handler}/sign-up`,
-    afterSignUp: afterSignIn,
-    signOut: `${handler}/sign-out`,
-    afterSignOut: home,
-    emailVerification: `${handler}/email-verification`,
-    passwordReset: `${handler}/password-reset`,
-    forgotPassword: `${handler}/forgot-password`,
-    oauthCallback: `${handler}/oauth-callback`,
-    magicLinkCallback: `${handler}/magic-link-callback`,
-    home: home,
-    accountSettings: `${handler}/account-settings`,
-    error: `${handler}/error`,
-    teamInvitation: `${handler}/team-invitation`,
-    ...filterUndefined(partial),
-  };
-}
-
-function getDefaultProjectId() {
-  return process.env.NEXT_PUBLIC_STACK_PROJECT_ID || throwErr(new Error("Welcome to Stack Auth! It seems that you haven't provided a project ID. Please create a project on the Stack dashboard at https://app.stack-auth.com and put it in the NEXT_PUBLIC_STACK_PROJECT_ID environment variable."));
-}
-
-function getDefaultPublishableClientKey() {
-  return process.env.NEXT_PUBLIC_STACK_PUBLISHABLE_CLIENT_KEY || throwErr(new Error("Welcome to Stack Auth! It seems that you haven't provided a publishable client key. Please create an API key for your project on the Stack dashboard at https://app.stack-auth.com and copy your publishable client key into the NEXT_PUBLIC_STACK_PUBLISHABLE_CLIENT_KEY environment variable."));
-}
-
-function getDefaultSecretServerKey() {
-  return process.env.STACK_SECRET_SERVER_KEY || throwErr(new Error("No secret server key provided. Please copy your key from the Stack dashboard and put your it in the STACK_SECRET_SERVER_KEY environment variable."));
-}
-
-function getDefaultSuperSecretAdminKey() {
-  return process.env.STACK_SUPER_SECRET_ADMIN_KEY || throwErr(new Error("No super secret admin key provided. Please copy your key from the Stack dashboard and put it in the STACK_SUPER_SECRET_ADMIN_KEY environment variable."));
-}
-
-/**
- * Returns the base URL for the Stack API.
- *
- * The URL can be specified in several ways, in order of precedence:
- * 1. Directly through userSpecifiedBaseUrl parameter as string or browser/server object
- * 2. Through environment variables:
- *    - Browser: NEXT_PUBLIC_BROWSER_STACK_API_URL
- *    - Server: NEXT_PUBLIC_SERVER_STACK_API_URL
- *    - Fallback: NEXT_PUBLIC_STACK_API_URL or NEXT_PUBLIC_STACK_URL
- * 3. Default base URL if none of the above are specified
- *
- * The function also ensures the URL doesn't end with a trailing slash
- * by removing it if present.
- *
- * @param userSpecifiedBaseUrl - Optional URL override as string or {browser, server} object
- * @returns The configured base URL without trailing slash
-
- */
-function getBaseUrl(userSpecifiedBaseUrl: string | { browser: string, server: string } | undefined) {
-  let url;
-  if (userSpecifiedBaseUrl) {
-    if (typeof userSpecifiedBaseUrl === "string") {
-      url = userSpecifiedBaseUrl;
-    } else {
-      if (isBrowserLike()) {
-        url = userSpecifiedBaseUrl.browser;
-      } else {
-        url = userSpecifiedBaseUrl.server;
-      }
-    }
-  } else {
-    if (isBrowserLike()) {
-      url = process.env.NEXT_PUBLIC_BROWSER_STACK_API_URL;
-    } else {
-      url = process.env.NEXT_PUBLIC_SERVER_STACK_API_URL;
-    }
-    url = url || process.env.NEXT_PUBLIC_STACK_API_URL || process.env.NEXT_PUBLIC_STACK_URL || defaultBaseUrl;
-  }
-
-  return url.endsWith('/') ? url.slice(0, -1) : url;
-}
-
-export type StackClientAppConstructorOptions<HasTokenStore extends boolean, ProjectId extends string> = {
-  baseUrl?: string | { browser: string, server: string },
-  projectId?: ProjectId,
-  publishableClientKey?: string,
-  urls?: Partial<HandlerUrls>,
-  oauthScopesOnSignIn?: Partial<OAuthScopesOnSignIn>,
-  tokenStore: TokenStoreInit<HasTokenStore>,
-  redirectMethod?: RedirectMethod,
-
-  /**
-   * By default, the Stack app will automatically prefetch some data from Stack's server when this app is first
-   * constructed. This improves the performance of your app, but will create network requests that are unnecessary if
-   * the app is never used or disposed of immediately. To disable this behavior, set this option to true.
-   */
-  noAutomaticPrefetch?: boolean,
-};
-
-export type StackServerAppConstructorOptions<HasTokenStore extends boolean, ProjectId extends string> = StackClientAppConstructorOptions<HasTokenStore, ProjectId> & {
-  secretServerKey?: string,
-};
-
-export type StackAdminAppConstructorOptions<HasTokenStore extends boolean, ProjectId extends string> = (
-  | (
-    & StackServerAppConstructorOptions<HasTokenStore, ProjectId>
-    & {
-      superSecretAdminKey?: string,
-    }
-  )
-  | (
-    & Omit<StackServerAppConstructorOptions<HasTokenStore, ProjectId>, "publishableClientKey" | "secretServerKey">
-    & {
-      projectOwnerSession: InternalSession,
-    }
-  )
-);
-
-export type StackClientAppJson<HasTokenStore extends boolean, ProjectId extends string> = StackClientAppConstructorOptions<HasTokenStore, ProjectId> & {
-  uniqueIdentifier: string,
-  // note: if you add more fields here, make sure to ensure the checkString in the constructor has/doesn't have them
-};
-
-const defaultBaseUrl = "https://api.stack-auth.com";
-
-type TokenObject = {
-  accessToken: string | null,
-  refreshToken: string | null,
-};
-
-function createEmptyTokenStore() {
-  return new Store<TokenObject>({
-    refreshToken: null,
-    accessToken: null,
-  });
-}
-
-const cachePromiseByComponentId = new Map<string, ReactPromise<Result<unknown>>>();
-
-// IF_PLATFORM react-like
-function useAsyncCache<D extends any[], T>(cache: AsyncCache<D, Result<T>>, dependencies: D, caller: string): T {
-  // we explicitly don't want to run this hook in SSR
-  suspendIfSsr(caller);
-
-  const id = React.useId();
-
-  const subscribe = useCallback((cb: () => void) => {
-    const { unsubscribe } = cache.onStateChange(dependencies, () => {
-      cachePromiseByComponentId.delete(id);
-      cb();
-    });
-    return unsubscribe;
-  }, [cache, ...dependencies]);
-  const getSnapshot = useCallback(() => {
-    // React checks whether a promise passed to `use` is still the same as the previous one by comparing the reference.
-    // If we didn't cache here, this wouldn't work because the promise would be recreated every time the value changes.
-    if (!cachePromiseByComponentId.has(id)) {
-      cachePromiseByComponentId.set(id, cache.getOrWait(dependencies, "read-write"));
-    }
-    return cachePromiseByComponentId.get(id) as ReactPromise<Result<T>>;
-  }, [cache, ...dependencies]);
-
-  // note: we must use React.useSyncExternalStore instead of importing the function directly, as it will otherwise
-  // throw an error ("can't import useSyncExternalStore from the server")
-  const promise = React.useSyncExternalStore(
-    subscribe,
-    getSnapshot,
-    () => throwErr(new Error("getServerSnapshot should never be called in useAsyncCache because we restrict to CSR earlier"))
-  );
-
-  const result = React.use(promise);
-  if (result.status === "error") {
-    const error = result.error;
-    if (error instanceof Error && !(error as any).__stackHasConcatenatedStacktraces) {
-      concatStacktraces(error, new Error());
-      (error as any).__stackHasConcatenatedStacktraces = true;
-    }
-    throw error;
-  }
-  return result.data;
-}
-// END_PLATFORM
-
-/** @internal */
-export const stackAppInternalsSymbol = Symbol.for("StackAuth--DO-NOT-USE-OR-YOU-WILL-BE-FIRED--StackAppInternals");
-
-const allClientApps = new Map<string, [checkString: string, app: StackClientApp<any, any>]>();
-
-const createCache = <D extends any[], T>(fetcher: (dependencies: D) => Promise<T>) => {
-  return new AsyncCache<D, Result<T>>(
-    async (dependencies) => await Result.fromThrowingAsync(async () => await fetcher(dependencies)),
-    {},
-  );
-};
-
-const createCacheBySession = <D extends any[], T>(fetcher: (session: InternalSession, extraDependencies: D) => Promise<T> ) => {
-  return new AsyncCache<[InternalSession, ...D], Result<T>>(
-    async ([session, ...extraDependencies]) => await Result.fromThrowingAsync(async () => await fetcher(session, extraDependencies)),
-    {
-      onSubscribe: ([session], refresh) => {
-        const handler = session.onInvalidate(() => refresh());
-        return () => handler.unsubscribe();
-      },
-    },
-  );
-};
-
-let numberOfAppsCreated = 0;
-
-class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends string = string> {
-  protected _uniqueIdentifier: string | undefined = undefined;
-  protected _interface: StackClientInterface;
-  protected readonly _tokenStoreInit: TokenStoreInit<HasTokenStore>;
-  protected readonly _redirectMethod: RedirectMethod | undefined;
-  protected readonly _urlOptions: Partial<HandlerUrls>;
-  protected readonly _oauthScopesOnSignIn: Partial<OAuthScopesOnSignIn>;
-
-  private __DEMO_ENABLE_SLIGHT_FETCH_DELAY = false;
-  private readonly _ownedAdminApps = new DependenciesMap<[InternalSession, string], _StackAdminAppImpl<false, string>>();
-
-  private readonly _currentUserCache = createCacheBySession(async (session) => {
-    if (this.__DEMO_ENABLE_SLIGHT_FETCH_DELAY) {
-      await wait(2000);
-    }
-    if (session.isKnownToBeInvalid()) {
-      // let's save ourselves a network request
-      //
-      // this also makes a certain race condition less likely to happen. particularly, it's quite common for code to
-      // look like this:
-      //
-      //     const user = await useUser({ or: "required" });
-      //     const something = user.useSomething();
-      //
-      // now, let's say the session is invalidated. this will trigger a refresh to refresh both the user and the
-      // something. however, it's not guaranteed that the user will return first, so useUser might still return a
-      // user object while the something request has already completed (and failed, because the session is invalid).
-      // by returning null quickly here without a request, it is very very likely for the user request to complete
-      // first.
-      //
-      // TODO HACK: the above is a bit of a hack, and we should probably think of more consistent ways to handle this.
-      // it also only works for the user endpoint, and only if the session is known to be invalid.
-      return null;
-    }
-    return await this._interface.getClientUserByToken(session);
-  });
-  private readonly _currentProjectCache = createCache(async () => {
-    return Result.orThrow(await this._interface.getClientProject());
-  });
-  private readonly _ownedProjectsCache = createCacheBySession(async (session) => {
-    return await this._interface.listProjects(session);
-  });
-  private readonly _currentUserPermissionsCache = createCacheBySession<
-    [string, boolean],
-    TeamPermissionsCrud['Client']['Read'][]
-  >(async (session, [teamId, recursive]) => {
-    return await this._interface.listCurrentUserTeamPermissions({ teamId, recursive }, session);
-  });
-  private readonly _currentUserTeamsCache = createCacheBySession(async (session) => {
-    return await this._interface.listCurrentUserTeams(session);
-  });
-  private readonly _currentUserOAuthConnectionAccessTokensCache = createCacheBySession<[string, string], { accessToken: string } | null>(
-    async (session, [providerId, scope]) => {
-      try {
-        const result = await this._interface.createProviderAccessToken(providerId, scope || "", session);
-        return { accessToken: result.access_token };
-      } catch (err) {
-        if (!(err instanceof KnownErrors.OAuthConnectionDoesNotHaveRequiredScope || err instanceof KnownErrors.OAuthConnectionNotConnectedToUser)) {
-          throw err;
-        }
-      }
-      return null;
-    }
-  );
-  private readonly _currentUserOAuthConnectionCache = createCacheBySession<[ProviderType, string, boolean], OAuthConnection | null>(
-    async (session, [providerId, scope, redirect]) => {
-      return await this._getUserOAuthConnectionCacheFn({
-        getUser: async () => Result.orThrow(await this._currentUserCache.getOrWait([session], "write-only")),
-        getOrWaitOAuthToken: async () => Result.orThrow(await this._currentUserOAuthConnectionAccessTokensCache.getOrWait([session, providerId, scope || ""] as const, "write-only")),
-        // IF_PLATFORM react-like
-        useOAuthToken: () => useAsyncCache(this._currentUserOAuthConnectionAccessTokensCache, [session, providerId, scope || ""] as const, "useOAuthToken"),
-        // END_PLATFORM
-        providerId,
-        scope,
-        redirect,
-        session,
-      });
-    }
-  );
-  private readonly _teamMemberProfilesCache = createCacheBySession<[string], TeamMemberProfilesCrud['Client']['Read'][]>(
-    async (session, [teamId]) => {
-      return await this._interface.listTeamMemberProfiles({ teamId }, session);
-    }
-  );
-  private readonly _teamInvitationsCache = createCacheBySession<[string], TeamInvitationCrud['Client']['Read'][]>(
-    async (session, [teamId]) => {
-      return await this._interface.listTeamInvitations({ teamId }, session);
-    }
-  );
-  private readonly _currentUserTeamProfileCache = createCacheBySession<[string], TeamMemberProfilesCrud['Client']['Read']>(
-    async (session, [teamId]) => {
-      return await this._interface.getTeamMemberProfile({ teamId, userId: 'me' }, session);
-    }
-  );
-  private readonly _clientContactChannelsCache = createCacheBySession<[], ContactChannelsCrud['Client']['Read'][]>(
-    async (session) => {
-      return await this._interface.listClientContactChannels(session);
-    }
-  );
-
-  protected async _createCookieHelper(): Promise<CookieHelper> {
-    if (this._tokenStoreInit === 'nextjs-cookie' || this._tokenStoreInit === 'cookie') {
-      return await createCookieHelper();
-    } else {
-      return await createPlaceholderCookieHelper();
-    }
-  }
-
-  protected async _getUserOAuthConnectionCacheFn(options: {
-    getUser: () => Promise<CurrentUserCrud['Client']['Read'] | null>,
-    getOrWaitOAuthToken: () => Promise<{ accessToken: string } | null>,
-    // IF_PLATFORM react-like
-    useOAuthToken: () => { accessToken: string } | null,
-    // END_PLATFORM
-    providerId: ProviderType,
-    scope: string | null,
-  } & ({ redirect: true, session: InternalSession | null } | { redirect: false }),) {
-    const user = await options.getUser();
-    let hasConnection = true;
-    if (!user || !user.oauth_providers.find((p) => p.id === options.providerId)) {
-      hasConnection = false;
-    }
-
-    const token = await options.getOrWaitOAuthToken();
-    if (!token) {
-      hasConnection = false;
-    }
-
-    if (!hasConnection && options.redirect) {
-      if (!options.session) {
-        throw new Error("No session found. You might be calling getConnectedAccount with redirect without having a user session.");
-      }
-      await addNewOAuthProviderOrScope(
-          this._interface,
-          {
-            provider: options.providerId,
-            redirectUrl: this.urls.oauthCallback,
-            errorRedirectUrl: this.urls.error,
-            providerScope: mergeScopeStrings(options.scope || "", (this._oauthScopesOnSignIn[options.providerId] ?? []).join(" ")),
-          },
-          options.session,
-        );
-      return await neverResolve();
-    } else if (!hasConnection) {
-      return null;
-    }
-
-    const app = this;
-    return {
-      id: options.providerId,
-      async getAccessToken() {
-        const result = await options.getOrWaitOAuthToken();
-        if (!result) {
-          throw new StackAssertionError("No access token available");
-        }
-        return result;
-      },
-      // IF_PLATFORM react-like
-      useAccessToken() {
-        const result = options.useOAuthToken();
-        if (!result) {
-          throw new StackAssertionError("No access token available");
-        }
-        return result;
-      }
-      // END_PLATFORM
-    };
-  }
-
-  constructor(protected readonly _options:
-    & {
-      uniqueIdentifier?: string,
-      checkString?: string,
-    }
-    & (
-      | StackClientAppConstructorOptions<HasTokenStore, ProjectId>
-      | Exclude<StackClientAppConstructorOptions<HasTokenStore, ProjectId>, "baseUrl" | "projectId" | "publishableClientKey"> & {
-        interface: StackClientInterface,
-      }
-    )
-  ) {
-    if ("interface" in _options) {
-      this._interface = _options.interface;
-    } else {
-      this._interface = new StackClientInterface({
-        getBaseUrl: () => getBaseUrl(_options.baseUrl),
-        projectId: _options.projectId ?? getDefaultProjectId(),
-        clientVersion,
-        publishableClientKey: _options.publishableClientKey ?? getDefaultPublishableClientKey(),
-        prepareRequest: async () => {
-          // NEXT_LINE_PLATFORM next
-          await cookies?.();
-        }
-      });
-    }
-
-    this._tokenStoreInit = _options.tokenStore;
-    this._redirectMethod = _options.redirectMethod || "none";
-    // NEXT_LINE_PLATFORM next
-    this._redirectMethod = _options.redirectMethod || "nextjs";
-    this._urlOptions = _options.urls ?? {};
-    this._oauthScopesOnSignIn = _options.oauthScopesOnSignIn ?? {};
-
-    if (_options.uniqueIdentifier) {
-      this._uniqueIdentifier = _options.uniqueIdentifier;
-      this._initUniqueIdentifier();
-    }
-
-    if (!_options.noAutomaticPrefetch) {
-      numberOfAppsCreated++;
-      if (numberOfAppsCreated > 10) {
-        (process.env.NODE_ENV === "development" ? console.log : console.warn)(`You have created more than 10 Stack apps with automatic pre-fetch enabled (${numberOfAppsCreated}). This is usually a sign of a memory leak, but can sometimes be caused by hot reload of your tech stack. If you are getting this error and it is not caused by hot reload, make sure to minimize the number of Stack apps per page (usually, one per project). (If it is caused by hot reload and does not occur in production, you can safely ignore it.)`);
-      }
-    }
-  }
-
-  protected _initUniqueIdentifier() {
-    if (!this._uniqueIdentifier) {
-      throw new StackAssertionError("Unique identifier not initialized");
-    }
-    if (allClientApps.has(this._uniqueIdentifier)) {
-      throw new StackAssertionError("A Stack client app with the same unique identifier already exists");
-    }
-    allClientApps.set(this._uniqueIdentifier, [this._options.checkString ?? "default check string", this]);
-  }
-
-  /**
-   * Cloudflare workers does not allow use of randomness on the global scope (on which the Stack app is probably
-   * initialized). For that reason, we generate the unique identifier lazily when it is first needed instead of in the
-   * constructor.
-   */
-  protected _getUniqueIdentifier() {
-    if (!this._uniqueIdentifier) {
-      this._uniqueIdentifier = generateUuid();
-      this._initUniqueIdentifier();
-    }
-    return this._uniqueIdentifier!;
-  }
-
-  protected async _checkFeatureSupport(name: string, options: any) {
-    return await this._interface.checkFeatureSupport({ ...options, name });
-  }
-
-  protected _useCheckFeatureSupport(name: string, options: any): never {
-    runAsynchronously(this._checkFeatureSupport(name, options));
-    throw new StackAssertionError(`${name} is not currently supported. Please reach out to Stack support for more information.`);
-  }
-
-  protected _memoryTokenStore = createEmptyTokenStore();
-  protected _nextServerCookiesTokenStores = new WeakMap<object, Store<TokenObject>>();
-  protected _requestTokenStores = new WeakMap<RequestLike, Store<TokenObject>>();
-  protected _storedBrowserCookieTokenStore: Store<TokenObject> | null = null;
-  protected get _refreshTokenCookieName() {
-    return `stack-refresh-${this.projectId}`;
-  }
-  protected _getTokensFromCookies(cookies: { refreshTokenCookie: string | null, accessTokenCookie: string | null }): TokenObject {
-    const refreshToken = cookies.refreshTokenCookie;
-    const accessTokenObject = cookies.accessTokenCookie?.startsWith('[\"') ? JSON.parse(cookies.accessTokenCookie) : null;  // gotta check for validity first for backwards-compat, and also in case someone messes with the cookie value
-    const accessToken = accessTokenObject && refreshToken === accessTokenObject[0] ? accessTokenObject[1] : null;  // if the refresh token has changed, the access token is invalid
-    return {
-      refreshToken,
-      accessToken,
-    };
-  }
-  protected get _accessTokenCookieName() {
-    // The access token, unlike the refresh token, should not depend on the project ID. We never want to store the
-    // access token in cookies more than once because of how big it is (there's a limit of 4096 bytes for all cookies
-    // together). This means that, if you have multiple projects on the same domain, some of them will need to refetch
-    // the access token on page reload.
-    return `stack-access`;
-  }
-  protected _getBrowserCookieTokenStore(): Store<TokenObject> {
-    if (!isBrowserLike()) {
-      throw new Error("Cannot use cookie token store on the server!");
-    }
-
-    if (this._storedBrowserCookieTokenStore === null) {
-      const getCurrentValue = (old: TokenObject | null) => {
-        const tokens = this._getTokensFromCookies({
-          refreshTokenCookie: getCookieClient(this._refreshTokenCookieName) ?? getCookieClient('stack-refresh'),  // keep old cookie name for backwards-compatibility
-          accessTokenCookie: getCookieClient(this._accessTokenCookieName),
-        });
-        return {
-          refreshToken: tokens.refreshToken,
-          accessToken: tokens.accessToken ?? (old?.refreshToken === tokens.refreshToken ? old.accessToken : null),
-        };
-      };
-      this._storedBrowserCookieTokenStore = new Store<TokenObject>(getCurrentValue(null));
-      let hasSucceededInWriting = true;
-
-      setInterval(() => {
-        if (hasSucceededInWriting) {
-          const oldValue = this._storedBrowserCookieTokenStore!.get();
-          const currentValue = getCurrentValue(oldValue);
-          if (!deepPlainEquals(currentValue, oldValue)) {
-            this._storedBrowserCookieTokenStore!.set(currentValue);
-          }
-        }
-      }, 100);
-      this._storedBrowserCookieTokenStore.onChange((value) => {
-        try {
-          setOrDeleteCookieClient(this._refreshTokenCookieName, value.refreshToken, { maxAge: 60 * 60 * 24 * 365 });
-          setOrDeleteCookieClient(this._accessTokenCookieName, value.accessToken ? JSON.stringify([value.refreshToken, value.accessToken]) : null, { maxAge: 60 * 60 * 24 });
-          deleteCookieClient('stack-refresh');  // delete cookie name from previous versions (for backwards-compatibility)
-          hasSucceededInWriting = true;
-        } catch (e) {
-          if (!isBrowserLike()) {
-            // Setting cookies inside RSCs is not allowed, so we just ignore it
-            hasSucceededInWriting = false;
-          } else {
-            throw e;
-          }
-        }
-      });
-    }
-
-    return this._storedBrowserCookieTokenStore;
-  };
-  protected _getOrCreateTokenStore(cookieHelper: CookieHelper, overrideTokenStoreInit?: TokenStoreInit): Store<TokenObject> {
-    const tokenStoreInit = overrideTokenStoreInit === undefined ? this._tokenStoreInit : overrideTokenStoreInit;
-
-    switch (tokenStoreInit) {
-      case "cookie": {
-        return this._getBrowserCookieTokenStore();
-      }
-      case "nextjs-cookie": {
-        if (isBrowserLike()) {
-          return this._getBrowserCookieTokenStore();
-        } else {
-          const tokens = this._getTokensFromCookies({
-            refreshTokenCookie: cookieHelper.get(this._refreshTokenCookieName) ?? cookieHelper.get('stack-refresh'),  // keep old cookie name for backwards-compatibility
-            accessTokenCookie: cookieHelper.get(this._accessTokenCookieName),
-          });
-          const store = new Store<TokenObject>(tokens);
-          store.onChange((value) => {
-            runAsynchronously(async () => {
-              // TODO HACK this is a bit of a hack; while the order happens to work in practice (because the only actual
-              // async operation is waiting for the `cookies()` to resolve which always happens at the same time during
-              // the same request), it's not guaranteed to be free of race conditions if there are many updates happening
-              // at the same time
-              //
-              // instead, we should create a per-request cookie helper outside of the store onChange and reuse that
-              //
-              // but that's kinda hard to do because Next.js doesn't expose a documented way to find out which request
-              // we're currently processing, and hence we can't find out which per-request cookie helper to use
-              //
-              // so hack it is
-              await Promise.all([
-                setOrDeleteCookie(this._refreshTokenCookieName, value.refreshToken, { maxAge: 60 * 60 * 24 * 365, noOpIfServerComponent: true }),
-                setOrDeleteCookie(this._accessTokenCookieName, value.accessToken ? JSON.stringify([value.refreshToken, value.accessToken]) : null, { maxAge: 60 * 60 * 24, noOpIfServerComponent: true }),
-              ]);
-            });
-          });
-          return store;
-        }
-      }
-      case "memory": {
-        return this._memoryTokenStore;
-      }
-      default: {
-        if (tokenStoreInit === null) {
-          return createEmptyTokenStore();
-        } else if (typeof tokenStoreInit === "object" && "headers" in tokenStoreInit) {
-          if (this._requestTokenStores.has(tokenStoreInit)) return this._requestTokenStores.get(tokenStoreInit)!;
-
-          // x-stack-auth header
-          const stackAuthHeader = tokenStoreInit.headers.get("x-stack-auth");
-          if (stackAuthHeader) {
-            let parsed;
-            try {
-              parsed = JSON.parse(stackAuthHeader);
-              if (typeof parsed !== "object") throw new Error("x-stack-auth header must be a JSON object");
-              if (parsed === null) throw new Error("x-stack-auth header must not be null");
-            } catch (e) {
-              throw new Error(`Invalid x-stack-auth header: ${stackAuthHeader}`, { cause: e });
-            }
-            return this._getOrCreateTokenStore(cookieHelper, {
-              accessToken: parsed.accessToken ?? null,
-              refreshToken: parsed.refreshToken ?? null,
-            });
-          }
-
-          // read from cookies
-          const cookieHeader = tokenStoreInit.headers.get("cookie");
-          const parsed = cookie.parse(cookieHeader || "");
-          const res = new Store<TokenObject>({
-            refreshToken: parsed[this._refreshTokenCookieName] || parsed['stack-refresh'] || null,  // keep old cookie name for backwards-compatibility
-            accessToken: parsed[this._accessTokenCookieName] || null,
-          });
-          this._requestTokenStores.set(tokenStoreInit, res);
-          return res;
-        } else if ("accessToken" in tokenStoreInit || "refreshToken" in tokenStoreInit) {
-          return new Store<TokenObject>({
-            refreshToken: tokenStoreInit.refreshToken,
-            accessToken: tokenStoreInit.accessToken,
-          });
-        }
-
-        throw new Error(`Invalid token store ${tokenStoreInit}`);
-      }
-    }
-  }
-
-  // IF_PLATFORM react-like
-  protected _useTokenStore(overrideTokenStoreInit?: TokenStoreInit): Store<TokenObject> {
-    suspendIfSsr();
-    const cookieHelper = createBrowserCookieHelper();
-    const tokenStore = this._getOrCreateTokenStore(cookieHelper, overrideTokenStoreInit);
-    return tokenStore;
-  }
-  // END_PLATFORM
-
-  /**
-   * A map from token stores and session keys to sessions.
-   *
-   * This isn't just a map from session keys to sessions for two reasons:
-   *
-   * - So we can garbage-collect Session objects when the token store is garbage-collected
-   * - So different token stores are separated and don't leak information between each other, eg. if the same user sends two requests to the same server they should get a different session object
-   */
-  private _sessionsByTokenStoreAndSessionKey = new WeakMap<Store<TokenObject>, Map<string, InternalSession>>();
-  protected _getSessionFromTokenStore(tokenStore: Store<TokenObject>): InternalSession {
-    const tokenObj = tokenStore.get();
-    const sessionKey = InternalSession.calculateSessionKey(tokenObj);
-    const existing = sessionKey ? this._sessionsByTokenStoreAndSessionKey.get(tokenStore)?.get(sessionKey) : null;
-    if (existing) return existing;
-
-    const session = this._interface.createSession({
-      refreshToken: tokenObj.refreshToken,
-      accessToken: tokenObj.accessToken,
-    });
-    session.onAccessTokenChange((newAccessToken) => {
-      tokenStore.update((old) => ({
-        ...old,
-        accessToken: newAccessToken?.token ?? null
-      }));
-    });
-    session.onInvalidate(() => {
-      tokenStore.update((old) => ({
-        ...old,
-        accessToken: null,
-        refreshToken: null,
-      }));
-    });
-
-    let sessionsBySessionKey = this._sessionsByTokenStoreAndSessionKey.get(tokenStore) ?? new Map();
-    this._sessionsByTokenStoreAndSessionKey.set(tokenStore, sessionsBySessionKey);
-    sessionsBySessionKey.set(sessionKey, session);
-    return session;
-  }
-
-  protected async _getSession(overrideTokenStoreInit?: TokenStoreInit): Promise<InternalSession> {
-    const tokenStore = this._getOrCreateTokenStore(await this._createCookieHelper(), overrideTokenStoreInit);
-    return this._getSessionFromTokenStore(tokenStore);
-  }
-
-  // IF_PLATFORM react-like
-  protected _useSession(overrideTokenStoreInit?: TokenStoreInit): InternalSession {
-    const tokenStore = this._useTokenStore(overrideTokenStoreInit);
-    const subscribe = useCallback((cb: () => void) => {
-      const { unsubscribe } = tokenStore.onChange(() => {
-        cb();
-      });
-      return unsubscribe;
-    }, [tokenStore]);
-    const getSnapshot = useCallback(() => this._getSessionFromTokenStore(tokenStore), [tokenStore]);
-    return React.useSyncExternalStore(subscribe, getSnapshot, getSnapshot);
-  }
-  // END_PLATFORM
-
-  protected async _signInToAccountWithTokens(tokens: { accessToken: string | null, refreshToken: string }) {
-    if (!("accessToken" in tokens) || !("refreshToken" in tokens)) {
-      throw new StackAssertionError("Invalid tokens object; can't sign in with this", { tokens });
-    }
-    const tokenStore = this._getOrCreateTokenStore(await this._createCookieHelper());
-    tokenStore.set(tokens);
-  }
-
-  protected _hasPersistentTokenStore(overrideTokenStoreInit?: TokenStoreInit): this is StackClientApp<true, ProjectId> {
-    return (overrideTokenStoreInit !== undefined ? overrideTokenStoreInit : this._tokenStoreInit) !== null;
-  }
-
-  protected _ensurePersistentTokenStore(overrideTokenStoreInit?: TokenStoreInit): asserts this is StackClientApp<true, ProjectId>  {
-    if (!this._hasPersistentTokenStore(overrideTokenStoreInit)) {
-      throw new Error("Cannot call this function on a Stack app without a persistent token store. Make sure the tokenStore option on the constructor is set to a non-null value when initializing Stack.\n\nStack uses token stores to access access tokens of the current user. For example, on web frontends it is commonly the string value 'cookies' for cookie storage.");
-    }
-  }
-
-  protected _isInternalProject(): this is { projectId: "internal" } {
-    return this.projectId === "internal";
-  }
-
-  protected _ensureInternalProject(): asserts this is { projectId: "internal" } {
-    if (!this._isInternalProject()) {
-      throw new Error("Cannot call this function on a Stack app with a project ID other than 'internal'.");
-    }
-  }
-
-  protected _clientProjectFromCrud(crud: ProjectsCrud['Client']['Read']): Project {
-    return {
-      id: crud.id,
-      displayName: crud.display_name,
-      config: {
-        signUpEnabled: crud.config.sign_up_enabled,
-        credentialEnabled: crud.config.credential_enabled,
-        magicLinkEnabled: crud.config.magic_link_enabled,
-        passkeyEnabled: crud.config.passkey_enabled,
-        clientTeamCreationEnabled: crud.config.client_team_creation_enabled,
-        clientUserDeletionEnabled: crud.config.client_user_deletion_enabled,
-        oauthProviders: crud.config.enabled_oauth_providers.map((p) => ({
-          id: p.id,
-        })),
-      }
-    };
-  }
-
-  protected _clientTeamPermissionFromCrud(crud: TeamPermissionsCrud['Client']['Read']): TeamPermission {
-    return {
-      id: crud.id,
-    };
-  }
-
-  protected _clientTeamUserFromCrud(crud: TeamMemberProfilesCrud['Client']['Read']): TeamUser {
-    return {
-      id: crud.user_id,
-      teamProfile: {
-        displayName: crud.display_name,
-        profileImageUrl: crud.profile_image_url,
-      }
-    };
-  }
-
-  protected _clientTeamInvitationFromCrud(session: InternalSession, crud: TeamInvitationCrud['Client']['Read']): TeamInvitation {
-    return {
-      id: crud.id,
-      recipientEmail: crud.recipient_email,
-      expiresAt: new Date(crud.expires_at_millis),
-      revoke: async () => {
-        await this._interface.revokeTeamInvitation(crud.id, crud.team_id, session);
-        await this._teamInvitationsCache.refresh([session, crud.team_id]);
-      },
-    };
-  }
-
-  protected _clientTeamFromCrud(crud: TeamsCrud['Client']['Read'], session: InternalSession): Team {
-    const app = this;
-    return {
-      id: crud.id,
-      displayName: crud.display_name,
-      profileImageUrl: crud.profile_image_url,
-      clientMetadata: crud.client_metadata,
-      clientReadOnlyMetadata: crud.client_read_only_metadata,
-      async inviteUser(options: { email: string, callbackUrl?: string }) {
-        if (!options.callbackUrl && !await app._getCurrentUrl()) {
-          throw new Error("Cannot invite user without a callback URL from the server or without a redirect method. Make sure you pass the `callbackUrl` option: `inviteUser({ email, callbackUrl: ... })`");
-        }
-        await app._interface.sendTeamInvitation({
-          teamId: crud.id,
-          email: options.email,
-          session,
-          callbackUrl: options.callbackUrl ?? constructRedirectUrl(app.urls.teamInvitation),
-        });
-        await app._teamInvitationsCache.refresh([session, crud.id]);
-      },
-      async listUsers() {
-        const result = Result.orThrow(await app._teamMemberProfilesCache.getOrWait([session, crud.id], "write-only"));
-        return result.map((crud) => app._clientTeamUserFromCrud(crud));
-      },
-      // IF_PLATFORM react-like
-      useUsers() {
-        const result = useAsyncCache(app._teamMemberProfilesCache, [session, crud.id] as const, "team.useUsers()");
-        return result.map((crud) => app._clientTeamUserFromCrud(crud));
-      },
-      // END_PLATFORM
-      async listInvitations() {
-        const result = Result.orThrow(await app._teamInvitationsCache.getOrWait([session, crud.id], "write-only"));
-        return result.map((crud) => app._clientTeamInvitationFromCrud(session, crud));
-      },
-      // IF_PLATFORM react-like
-      useInvitations() {
-        const result = useAsyncCache(app._teamInvitationsCache, [session, crud.id] as const, "team.useInvitations()");
-        return result.map((crud) => app._clientTeamInvitationFromCrud(session, crud));
-      },
-      // END_PLATFORM
-      async update(data: TeamUpdateOptions){
-        await app._interface.updateTeam({ data: teamUpdateOptionsToCrud(data), teamId: crud.id }, session);
-        await app._currentUserTeamsCache.refresh([session]);
-      },
-      async delete() {
-        await app._interface.deleteTeam(crud.id, session);
-        await app._currentUserTeamsCache.refresh([session]);
-      },
-    };
-  }
-
-  protected _clientContactChannelFromCrud(crud: ContactChannelsCrud['Client']['Read'], session: InternalSession): ContactChannel {
-    const app = this;
-    return {
-      id: crud.id,
-      value: crud.value,
-      type: crud.type,
-      isVerified: crud.is_verified,
-      isPrimary: crud.is_primary,
-      usedForAuth: crud.used_for_auth,
-
-      async sendVerificationEmail() {
-        await app._interface.sendCurrentUserContactChannelVerificationEmail(crud.id, constructRedirectUrl(app.urls.emailVerification), session);
-      },
-      async update(data: ContactChannelUpdateOptions) {
-        await app._interface.updateClientContactChannel(crud.id, contactChannelUpdateOptionsToCrud(data), session);
-        await app._clientContactChannelsCache.refresh([session]);
-      },
-      async delete() {
-        await app._interface.deleteClientContactChannel(crud.id, session);
-        await app._clientContactChannelsCache.refresh([session]);
-      },
-    };
-  }
-  protected _createAuth(session: InternalSession): Auth {
-    const app = this;
-    return {
-      _internalSession: session,
-      currentSession: {
-        async getTokens() {
-          const tokens = await session.getOrFetchLikelyValidTokens(20_000);
-          return {
-            accessToken: tokens?.accessToken.token ?? null,
-            refreshToken: tokens?.refreshToken?.token ?? null,
-          };
-        },
-      },
-      async getAuthHeaders(): Promise<{ "x-stack-auth": string }> {
-        return {
-          "x-stack-auth": JSON.stringify(await this.getAuthJson()),
-        };
-      },
-      async getAuthJson(): Promise<{ accessToken: string | null, refreshToken: string | null }> {
-        const tokens = await this.currentSession.getTokens();
-        return tokens;
-      },
-      async registerPasskey(options?: { hostname?: string }): Promise<Result<undefined, KnownErrors["PasskeyRegistrationFailed"] | KnownErrors["PasskeyWebAuthnError"]>> {
-        const hostname = (await app._getCurrentUrl())?.hostname;
-        if (!hostname) {
-          throw new StackAssertionError("hostname must be provided if the Stack App does not have a redirect method");
-        }
-
-        const initiationResult = await app._interface.initiatePasskeyRegistration({}, session);
-
-        if (initiationResult.status !== "ok") {
-          return Result.error(new KnownErrors.PasskeyRegistrationFailed("Failed to get initiation options for passkey registration"));
-        }
-
-        const { options_json, code } = initiationResult.data;
-
-        // HACK: Override the rpID to be the actual domain
-        if (options_json.rp.id !== "THIS_VALUE_WILL_BE_REPLACED.example.com") {
-          throw new StackAssertionError(`Expected returned RP ID from server to equal sentinel, but found ${options_json.rp.id}`);
-        }
-
-        options_json.rp.id = hostname;
-
-        let attResp;
-        try {
-          attResp = await startRegistration({ optionsJSON: options_json });
-          debugger;
-        } catch (error: any) {
-          if (error instanceof WebAuthnError) {
-            return Result.error(new KnownErrors.PasskeyWebAuthnError(error.message, error.name));
-          } else {
-            // This should never happen
-            return Result.error(new KnownErrors.PasskeyRegistrationFailed("Failed to start passkey registration"));
-          }
-        }
-
-
-        const registrationResult = await app._interface.registerPasskey({ credential: attResp, code }, session);
-
-        await app._refreshUser(session);
-        return registrationResult;
-      },
-      signOut(options?: { redirectUrl?: URL | string }) {
-        return app._signOut(session, options);
-      },
-    };
-  }
-
-  protected _editableTeamProfileFromCrud(crud: TeamMemberProfilesCrud['Client']['Read'], session: InternalSession): EditableTeamMemberProfile {
-    const app = this;
-    return {
-      displayName: crud.display_name,
-      profileImageUrl: crud.profile_image_url,
-      async update(update: { displayName?: string, profileImageUrl?: string }) {
-        await app._interface.updateTeamMemberProfile({
-          teamId: crud.team_id,
-          userId: crud.user_id,
-          profile: {
-            display_name: update.displayName,
-            profile_image_url: update.profileImageUrl,
-          },
-        }, session);
-        await app._currentUserTeamProfileCache.refresh([session, crud.team_id]);
-      }
-    };
-  }
-
-  protected _createBaseUser(crud: NonNullable<CurrentUserCrud['Client']['Read']> | UsersCrud['Server']['Read']): BaseUser {
-    return {
-      id: crud.id,
-      displayName: crud.display_name,
-      primaryEmail: crud.primary_email,
-      primaryEmailVerified: crud.primary_email_verified,
-      profileImageUrl: crud.profile_image_url,
-      signedUpAt: new Date(crud.signed_up_at_millis),
-      clientMetadata: crud.client_metadata,
-      clientReadOnlyMetadata: crud.client_read_only_metadata,
-      hasPassword: crud.has_password,
-      emailAuthEnabled: crud.auth_with_email,
-      otpAuthEnabled: crud.otp_auth_enabled,
-      oauthProviders: crud.oauth_providers,
-      passkeyAuthEnabled: crud.passkey_auth_enabled,
-      isMultiFactorRequired: crud.requires_totp_mfa,
-      toClientJson(): CurrentUserCrud['Client']['Read'] {
-        return crud;
-      }
-    };
-  }
-
-  protected _createUserExtraFromCurrent(crud: NonNullable<CurrentUserCrud['Client']['Read']>, session: InternalSession): UserExtra {
-    const app = this;
-    async function getConnectedAccount(id: ProviderType, options?: { scopes?: string[] }): Promise<OAuthConnection | null>;
-    async function getConnectedAccount(id: ProviderType, options: { or: 'redirect', scopes?: string[] }): Promise<OAuthConnection>;
-    async function getConnectedAccount(id: ProviderType, options?: { or?: 'redirect', scopes?: string[] }): Promise<OAuthConnection | null> {
-      const scopeString = options?.scopes?.join(" ");
-      return Result.orThrow(await app._currentUserOAuthConnectionCache.getOrWait([session, id, scopeString || "", options?.or === 'redirect'], "write-only"));
-    }
-
-    // IF_PLATFORM react-like
-    function useConnectedAccount(id: ProviderType, options?: { scopes?: string[] }): OAuthConnection | null;
-    function useConnectedAccount(id: ProviderType, options: { or: 'redirect', scopes?: string[] }): OAuthConnection;
-    function useConnectedAccount(id: ProviderType, options?: { or?: 'redirect', scopes?: string[] }): OAuthConnection | null {
-      const scopeString = options?.scopes?.join(" ");
-      return useAsyncCache(app._currentUserOAuthConnectionCache, [session, id, scopeString || "", options?.or === 'redirect'] as const, "user.useConnectedAccount()");
-    }
-    // END_PLATFORM
-    return {
-      setDisplayName(displayName: string) {
-        return this.update({ displayName });
-      },
-      setClientMetadata(metadata: Record<string, any>) {
-        return this.update({ clientMetadata: metadata });
-      },
-      async setSelectedTeam(team: Team | null) {
-        await this.update({ selectedTeamId: team?.id ?? null });
-      },
-      getConnectedAccount,
-      // NEXT_LINE_PLATFORM react-like
-      useConnectedAccount,
-      async getTeam(teamId: string) {
-        const teams = await this.listTeams();
-        return teams.find((t) => t.id === teamId) ?? null;
-      },
-      // IF_PLATFORM react-like
-      useTeam(teamId: string) {
-        const teams = this.useTeams();
-        return useMemo(() => {
-          return teams.find((t) => t.id === teamId) ?? null;
-        }, [teams, teamId]);
-      },
-      // END_PLATFORM
-      async listTeams() {
-        const teams = Result.orThrow(await app._currentUserTeamsCache.getOrWait([session], "write-only"));
-        return teams.map((crud) => app._clientTeamFromCrud(crud, session));
-      },
-      // IF_PLATFORM react-like
-      useTeams() {
-        const teams = useAsyncCache(app._currentUserTeamsCache, [session], "user.useTeams()");
-        return useMemo(() => teams.map((crud) => app._clientTeamFromCrud(crud, session)), [teams]);
-      },
-      // END_PLATFORM
-      async createTeam(data: TeamCreateOptions) {
-        const crud = await app._interface.createClientTeam(teamCreateOptionsToCrud(data, 'me'), session);
-        await app._currentUserTeamsCache.refresh([session]);
-        return app._clientTeamFromCrud(crud, session);
-      },
-      async leaveTeam(team: Team) {
-        await app._interface.leaveTeam(team.id, session);
-        // TODO: refresh cache
-      },
-      async listPermissions(scope: Team, options?: { recursive?: boolean }): Promise<TeamPermission[]> {
-        const recursive = options?.recursive ?? true;
-        const permissions = Result.orThrow(await app._currentUserPermissionsCache.getOrWait([session, scope.id, recursive], "write-only"));
-        return permissions.map((crud) => app._clientTeamPermissionFromCrud(crud));
-      },
-      // IF_PLATFORM react-like
-      usePermissions(scope: Team, options?: { recursive?: boolean }): TeamPermission[] {
-        const recursive = options?.recursive ?? true;
-        const permissions = useAsyncCache(app._currentUserPermissionsCache, [session, scope.id, recursive] as const, "user.usePermissions()");
-        return useMemo(() => permissions.map((crud) => app._clientTeamPermissionFromCrud(crud)), [permissions]);
-      },
-      // END_PLATFORM
-      // IF_PLATFORM react-like
-      usePermission(scope: Team, permissionId: string): TeamPermission | null {
-        const permissions = this.usePermissions(scope);
-        return useMemo(() => permissions.find((p) => p.id === permissionId) ?? null, [permissions, permissionId]);
-      },
-      // END_PLATFORM
-      async getPermission(scope: Team, permissionId: string): Promise<TeamPermission | null> {
-        const permissions = await this.listPermissions(scope);
-        return permissions.find((p) => p.id === permissionId) ?? null;
-      },
-      async hasPermission(scope: Team, permissionId: string): Promise<boolean> {
-        return (await this.getPermission(scope, permissionId)) !== null;
-      },
-      async update(update) {
-        return await app._updateClientUser(update, session);
-      },
-      async sendVerificationEmail(options?: { callbackUrl?: string }) {
-        if (!crud.primary_email) {
-          throw new StackAssertionError("User does not have a primary email");
-        }
-        if (!options?.callbackUrl && !await app._getCurrentUrl()) {
-          throw new Error("Cannot send verification email without a callback URL from the server or without a redirect method. Make sure you pass the `callbackUrl` option: `sendVerificationEmail({ callbackUrl: ... })`");
-        }
-        return await app._interface.sendVerificationEmail(crud.primary_email, options?.callbackUrl ?? constructRedirectUrl(app.urls.emailVerification), session);
-      },
-      async updatePassword(options: { oldPassword: string, newPassword: string}) {
-        const result = await app._interface.updatePassword(options, session);
-        await app._currentUserCache.refresh([session]);
-        return result;
-      },
-      async setPassword(options: { password: string }) {
-        const result = await app._interface.setPassword(options, session);
-        await app._currentUserCache.refresh([session]);
-        return result;
-      },
-      selectedTeam: crud.selected_team && this._clientTeamFromCrud(crud.selected_team, session),
-      async getTeamProfile(team: Team) {
-        const result = Result.orThrow(await app._currentUserTeamProfileCache.getOrWait([session, team.id], "write-only"));
-        return app._editableTeamProfileFromCrud(result, session);
-      },
-      // IF_PLATFORM react-like
-      useTeamProfile(team: Team) {
-        const result = useAsyncCache(app._currentUserTeamProfileCache, [session, team.id] as const, "user.useTeamProfile()");
-        return app._editableTeamProfileFromCrud(result, session);
-      },
-      // END_PLATFORM
-      async delete() {
-        await app._interface.deleteCurrentUser(session);
-        session.markInvalid();
-      },
-      async listContactChannels() {
-        const result = Result.orThrow(await app._clientContactChannelsCache.getOrWait([session], "write-only"));
-        return result.map((crud) => app._clientContactChannelFromCrud(crud, session));
-      },
-      // IF_PLATFORM react-like
-      useContactChannels() {
-        const result = useAsyncCache(app._clientContactChannelsCache, [session] as const, "user.useContactChannels()");
-        return result.map((crud) => app._clientContactChannelFromCrud(crud, session));
-      },
-      // END_PLATFORM
-      async createContactChannel(data: ContactChannelCreateOptions) {
-        const crud = await app._interface.createClientContactChannel(contactChannelCreateOptionsToCrud('me', data), session);
-        await app._clientContactChannelsCache.refresh([session]);
-        return app._clientContactChannelFromCrud(crud, session);
-      },
-    };
-  }
-
-  protected _createInternalUserExtra(session: InternalSession): InternalUserExtra {
-    const app = this;
-    this._ensureInternalProject();
-    return {
-      createProject(newProject: AdminProjectUpdateOptions & { displayName: string }) {
-        return app._createProject(session, newProject);
-      },
-      listOwnedProjects() {
-        return app._listOwnedProjects(session);
-      },
-      // IF_PLATFORM react-like
-      useOwnedProjects() {
-        return app._useOwnedProjects(session);
-      },
-      // END_PLATFORM
-    };
-  }
-
-  protected _currentUserFromCrud(crud: NonNullable<CurrentUserCrud['Client']['Read']>, session: InternalSession): ProjectCurrentUser<ProjectId> {
-    const currentUser = {
-      ...this._createBaseUser(crud),
-      ...this._createAuth(session),
-      ...this._createUserExtraFromCurrent(crud, session),
-      ...this._isInternalProject() ? this._createInternalUserExtra(session) : {},
-    } satisfies CurrentUser;
-
-    Object.freeze(currentUser);
-    return currentUser as ProjectCurrentUser<ProjectId>;
-  }
-
-  protected _getOwnedAdminApp(forProjectId: string, session: InternalSession): _StackAdminAppImpl<false, string> {
-    if (!this._ownedAdminApps.has([session, forProjectId])) {
-      this._ownedAdminApps.set([session, forProjectId], new _StackAdminAppImpl({
-        baseUrl: this._interface.options.getBaseUrl(),
-        projectId: forProjectId,
-        tokenStore: null,
-        projectOwnerSession: session,
-        noAutomaticPrefetch: true,
-      }));
-    }
-    return this._ownedAdminApps.get([session, forProjectId])!;
-  }
-
-  get projectId(): ProjectId {
-    return this._interface.projectId as ProjectId;
-  }
-
-  protected async _isTrusted(url: string): Promise<boolean> {
-    return isRelative(url);
-  }
-
-  get urls(): Readonly<HandlerUrls> {
-    return getUrls(this._urlOptions);
-  }
-
-  protected async _getCurrentUrl() {
-    if (this._redirectMethod === "none") {
-      return null;
-    }
-    return new URL(window.location.href);
-  }
-
-  protected async _redirectTo(options: { url: URL | string, replace?: boolean }) {
-    if (this._redirectMethod === "none") {
-      return;
-    // IF_PLATFORM next
-    } else if (isReactServer && this._redirectMethod === "nextjs") {
-      NextNavigation.redirect(options.url.toString(), options.replace ? NextNavigation.RedirectType.replace : NextNavigation.RedirectType.push);
-    // END_PLATFORM
-    } else if (typeof this._redirectMethod === "object" && this._redirectMethod.navigate) {
-      this._redirectMethod.navigate(options.url.toString());
-    } else {
-      if (options.replace) {
-        window.location.replace(options.url);
-      } else {
-        window.location.assign(options.url);
-      }
-    }
-
-    await wait(2000);
-  }
-
-  // IF_PLATFORM react-like
-  useNavigate(): (to: string) => void {
-    if (typeof this._redirectMethod === "object") {
-      return this._redirectMethod.useNavigate();
-    } else if (this._redirectMethod === "window") {
-      return () => window.location.assign;
-    // IF_PLATFORM next
-    } else if (this._redirectMethod === "nextjs") {
-      return NextNavigation.useRouter().push;
-    // END_PLATFORM
-    } else {
-      return (to: string) => {};
-    }
-  }
-  // END_PLATFORM
-  protected async _redirectIfTrusted(url: string, options?: RedirectToOptions) {
-    if (!await this._isTrusted(url)) {
-      throw new Error(`Redirect URL ${url} is not trusted; should be relative.`);
-    }
-    return await this._redirectTo({ url, ...options });
-  }
-
-  protected async _redirectToHandler(handlerName: keyof HandlerUrls, options?: RedirectToOptions) {
-    let url = this.urls[handlerName];
-    if (!url) {
-      throw new Error(`No URL for handler name ${handlerName}`);
-    }
-
-    if (!options?.noRedirectBack) {
-      if (handlerName === "afterSignIn" || handlerName === "afterSignUp") {
-        if (isReactServer || typeof window === "undefined") {
-          try {
-            await this._checkFeatureSupport("rsc-handler-" + handlerName, {});
-          } catch (e) {}
-        } else {
-          const queryParams = new URLSearchParams(window.location.search);
-          url = queryParams.get("after_auth_return_to") || url;
-        }
-      } else if (handlerName === "signIn" || handlerName === "signUp") {
-        if (isReactServer || typeof window === "undefined") {
-          try {
-            await this._checkFeatureSupport("rsc-handler-" + handlerName, {});
-          } catch (e) {}
-        } else {
-          const currentUrl = new URL(window.location.href);
-          const nextUrl = new URL(url, currentUrl);
-          if (currentUrl.searchParams.has("after_auth_return_to")) {
-            nextUrl.searchParams.set("after_auth_return_to", currentUrl.searchParams.get("after_auth_return_to")!);
-          } else if (currentUrl.protocol === nextUrl.protocol && currentUrl.host === nextUrl.host) {
-            nextUrl.searchParams.set("after_auth_return_to", getRelativePart(currentUrl));
-          }
-          url = getRelativePart(nextUrl);
-        }
-      }
-    }
-
-    await this._redirectIfTrusted(url, options);
-  }
-
-  async redirectToSignIn(options?: RedirectToOptions) { return await this._redirectToHandler("signIn", options); }
-  async redirectToSignUp(options?: RedirectToOptions) { return await this._redirectToHandler("signUp", options); }
-  async redirectToSignOut(options?: RedirectToOptions) { return await this._redirectToHandler("signOut", options); }
-  async redirectToEmailVerification(options?: RedirectToOptions) { return await this._redirectToHandler("emailVerification", options); }
-  async redirectToPasswordReset(options?: RedirectToOptions) { return await this._redirectToHandler("passwordReset", options); }
-  async redirectToForgotPassword(options?: RedirectToOptions) { return await this._redirectToHandler("forgotPassword", options); }
-  async redirectToHome(options?: RedirectToOptions) { return await this._redirectToHandler("home", options); }
-  async redirectToOAuthCallback(options?: RedirectToOptions) { return await this._redirectToHandler("oauthCallback", options); }
-  async redirectToMagicLinkCallback(options?: RedirectToOptions) { return await this._redirectToHandler("magicLinkCallback", options); }
-  async redirectToAfterSignIn(options?: RedirectToOptions) { return await this._redirectToHandler("afterSignIn", options); }
-  async redirectToAfterSignUp(options?: RedirectToOptions) { return await this._redirectToHandler("afterSignUp", options); }
-  async redirectToAfterSignOut(options?: RedirectToOptions) { return await this._redirectToHandler("afterSignOut", options); }
-  async redirectToAccountSettings(options?: RedirectToOptions) { return await this._redirectToHandler("accountSettings", options); }
-  async redirectToError(options?: RedirectToOptions) { return await this._redirectToHandler("error", options); }
-  async redirectToTeamInvitation(options?: RedirectToOptions) { return await this._redirectToHandler("teamInvitation", options); }
-
-  async sendForgotPasswordEmail(email: string, options?: { callbackUrl?: string }): Promise<Result<undefined, KnownErrors["UserNotFound"]>> {
-    if (!options?.callbackUrl && !await this._getCurrentUrl()) {
-      throw new Error("Cannot send forgot password email without a callback URL from the server or without a redirect method. Make sure you pass the `callbackUrl` option: `sendForgotPasswordEmail({ email, callbackUrl: ... })`");
-    }
-    return await this._interface.sendForgotPasswordEmail(email, options?.callbackUrl ?? constructRedirectUrl(this.urls.passwordReset));
-  }
-
-  async sendMagicLinkEmail(email: string, options?: { callbackUrl?: string }): Promise<Result<{ nonce: string }, KnownErrors["RedirectUrlNotWhitelisted"]>> {
-    if (!options?.callbackUrl && !await this._getCurrentUrl()) {
-      throw new Error("Cannot send magic link email without a callback URL from the server or without a redirect method. Make sure you pass the `callbackUrl` option: `sendMagicLinkEmail({ email, callbackUrl: ... })`");
-    }
-    return await this._interface.sendMagicLinkEmail(email, options?.callbackUrl ?? constructRedirectUrl(this.urls.magicLinkCallback));
-  }
-
-  async resetPassword(options: { password: string, code: string }): Promise<Result<undefined, KnownErrors["VerificationCodeError"]>> {
-    return await this._interface.resetPassword(options);
-  }
-
-  async verifyPasswordResetCode(code: string): Promise<Result<undefined, KnownErrors["VerificationCodeError"]>> {
-    return await this._interface.verifyPasswordResetCode(code);
-  }
-
-  async verifyTeamInvitationCode(code: string): Promise<Result<undefined, KnownErrors["VerificationCodeError"]>> {
-    return await this._interface.acceptTeamInvitation({
-      type: 'check',
-      code,
-      session: await this._getSession(),
-    });
-  }
-
-  async acceptTeamInvitation(code: string): Promise<Result<undefined, KnownErrors["VerificationCodeError"]>> {
-    const result = await this._interface.acceptTeamInvitation({
-      type: 'use',
-      code,
-      session: await this._getSession(),
-    });
-
-    if (result.status === 'ok') {
-      return Result.ok(undefined);
-    } else {
-      return Result.error(result.error);
-    }
-  }
-
-  async getTeamInvitationDetails(code: string): Promise<Result<{ teamDisplayName: string }, KnownErrors["VerificationCodeError"]>> {
-    const result = await this._interface.acceptTeamInvitation({
-      type: 'details',
-      code,
-      session: await this._getSession(),
-    });
-
-    if (result.status === 'ok') {
-      return Result.ok({ teamDisplayName: result.data.team_display_name });
-    } else {
-      return Result.error(result.error);
-    }
-  }
-
-  async verifyEmail(code: string): Promise<Result<undefined, KnownErrors["VerificationCodeError"]>> {
-    const result = await this._interface.verifyEmail(code);
-    await this._currentUserCache.refresh([await this._getSession()]);
-    await this._clientContactChannelsCache.refresh([await this._getSession()]);
-    return result;
-  }
-
-  async getUser(options: GetUserOptions<HasTokenStore> & { or: 'redirect' }): Promise<ProjectCurrentUser<ProjectId>>;
-  async getUser(options: GetUserOptions<HasTokenStore> & { or: 'throw' }): Promise<ProjectCurrentUser<ProjectId>>;
-  async getUser(options?: GetUserOptions<HasTokenStore>): Promise<ProjectCurrentUser<ProjectId> | null>;
-  async getUser(options?: GetUserOptions<HasTokenStore>): Promise<ProjectCurrentUser<ProjectId> | null> {
-    this._ensurePersistentTokenStore(options?.tokenStore);
-    const session = await this._getSession(options?.tokenStore);
-    const crud = Result.orThrow(await this._currentUserCache.getOrWait([session], "write-only"));
-
-    if (crud === null) {
-      switch (options?.or) {
-        case 'redirect': {
-          await this.redirectToSignIn({ replace: true });
-          break;
-        }
-        case 'throw': {
-          throw new Error("User is not signed in but getUser was called with { or: 'throw' }");
-        }
-        default: {
-          return null;
-        }
-      }
-    }
-
-    return crud && this._currentUserFromCrud(crud, session);
-  }
-
-  // IF_PLATFORM react-like
-  useUser(options: GetUserOptions<HasTokenStore> & { or: 'redirect' }): ProjectCurrentUser<ProjectId>;
-  useUser(options: GetUserOptions<HasTokenStore> & { or: 'throw' }): ProjectCurrentUser<ProjectId>;
-  useUser(options?: GetUserOptions<HasTokenStore>): ProjectCurrentUser<ProjectId> | null;
-  useUser(options?: GetUserOptions<HasTokenStore>): ProjectCurrentUser<ProjectId> | null {
-    this._ensurePersistentTokenStore(options?.tokenStore);
-
-    const session = this._useSession(options?.tokenStore);
-    const crud = useAsyncCache(this._currentUserCache, [session], "useUser()");
-
-    if (crud === null) {
-      switch (options?.or) {
-        case 'redirect': {
-          runAsynchronously(this.redirectToSignIn({ replace: true }));
-          suspend();
-          throw new StackAssertionError("suspend should never return");
-        }
-        case 'throw': {
-          throw new Error("User is not signed in but useUser was called with { or: 'throw' }");
-        }
-        case undefined:
-        case "return-null": {
-          // do nothing
-        }
-      }
-    }
-
-    return useMemo(() => {
-      return crud && this._currentUserFromCrud(crud, session);
-    }, [crud, session, options?.or]);
-  }
-  // END_PLATFORM
-
-  protected async _updateClientUser(update: UserUpdateOptions, session: InternalSession) {
-    const res = await this._interface.updateClientUser(userUpdateOptionsToCrud(update), session);
-    await this._refreshUser(session);
-    return res;
-  }
-
-  async signInWithOAuth(provider: ProviderType) {
-    if (typeof window === "undefined") {
-      throw new Error("signInWithOAuth can currently only be called in a browser environment");
-    }
-
-    this._ensurePersistentTokenStore();
-    await signInWithOAuth(
-      this._interface, {
-        provider,
-        redirectUrl: this.urls.oauthCallback,
-        errorRedirectUrl: this.urls.error,
-        providerScope: this._oauthScopesOnSignIn[provider]?.join(" "),
-      }
-    );
-  }
-
-  /**
-   * @deprecated
-   * TODO remove
-   */
-  protected async _experimentalMfa(error: KnownErrors['MultiFactorAuthenticationRequired'], session: InternalSession) {
-    const otp = prompt('Please enter the six-digit TOTP code from your authenticator app.');
-    if (!otp) {
-      throw new KnownErrors.InvalidTotpCode();
-    }
-
-    return await this._interface.totpMfa(
-      (error.details as any)?.attempt_code ?? throwErr("attempt code missing"),
-      otp,
-      session
-    );
-  }
-
-  /**
-   * @deprecated
-   * TODO remove
-   */
-  protected async _catchMfaRequiredError<T, E>(callback: () => Promise<Result<T, E>>): Promise<Result<T | { accessToken: string, refreshToken: string, newUser: boolean }, E>> {
-    try {
-      return await callback();
-    } catch (e) {
-      if (e instanceof KnownErrors.MultiFactorAuthenticationRequired) {
-        return Result.ok(await this._experimentalMfa(e, await this._getSession()));
-      }
-      throw e;
-    }
-  }
-
-  async signInWithCredential(options: {
-    email: string,
-    password: string,
-    noRedirect?: boolean,
-  }): Promise<Result<undefined, KnownErrors["EmailPasswordMismatch"] | KnownErrors["InvalidTotpCode"]>> {
-    this._ensurePersistentTokenStore();
-    const session = await this._getSession();
-    let result;
-    try {
-      result = await this._catchMfaRequiredError(async () => {
-        return await this._interface.signInWithCredential(options.email, options.password, session);
-      });
-    } catch (e) {
-      if (e instanceof KnownErrors.InvalidTotpCode) {
-        return Result.error(e);
-      }
-      throw e;
-    }
-
-    if (result.status === 'ok') {
-      await this._signInToAccountWithTokens(result.data);
-      if (!options.noRedirect) {
-        await this.redirectToAfterSignIn({ replace: true });
-      }
-      return Result.ok(undefined);
-    } else {
-      return Result.error(result.error);
-    }
-  }
-
-  async signUpWithCredential(options: {
-    email: string,
-    password: string,
-    noRedirect?: boolean,
-    verificationCallbackUrl?: string,
-  }): Promise<Result<undefined, KnownErrors["UserEmailAlreadyExists"] | KnownErrors['PasswordRequirementsNotMet']>> {
-    this._ensurePersistentTokenStore();
-    const session = await this._getSession();
-    const emailVerificationRedirectUrl = options.verificationCallbackUrl ?? constructRedirectUrl(this.urls.emailVerification);
-    const result = await this._interface.signUpWithCredential(
-      options.email,
-      options.password,
-      emailVerificationRedirectUrl,
-      session
-    );
-    if (result.status === 'ok') {
-      await this._signInToAccountWithTokens(result.data);
-      if (!options.noRedirect) {
-        await this.redirectToAfterSignUp({ replace: true });
-      }
-      return Result.ok(undefined);
-    } else {
-      return Result.error(result.error);
-    }
-  }
-
-  async signInWithMagicLink(code: string): Promise<Result<undefined, KnownErrors["VerificationCodeError"] | KnownErrors["InvalidTotpCode"]>> {
-    this._ensurePersistentTokenStore();
-    let result;
-    try {
-      result = await this._catchMfaRequiredError(async () => {
-        return await this._interface.signInWithMagicLink(code);
-      });
-    } catch (e) {
-      if (e instanceof KnownErrors.InvalidTotpCode) {
-        return Result.error(e);
-      }
-      throw e;
-    }
-
-    if (result.status === 'ok') {
-      await this._signInToAccountWithTokens(result.data);
-      if (result.data.newUser) {
-        await this.redirectToAfterSignUp({ replace: true });
-      } else {
-        await this.redirectToAfterSignIn({ replace: true });
-      }
-      return Result.ok(undefined);
-    } else {
-      return Result.error(result.error);
-    }
-  }
-
-  async signInWithPasskey(): Promise<Result<undefined, KnownErrors["PasskeyAuthenticationFailed"] | KnownErrors["InvalidTotpCode"] | KnownErrors["PasskeyWebAuthnError"]>> {
-    this._ensurePersistentTokenStore();
-    const session = await this._getSession();
-    let result;
-    try {
-      result = await this._catchMfaRequiredError(async () => {
-        const initiationResult = await this._interface.initiatePasskeyAuthentication({}, session);
-        if (initiationResult.status !== "ok") {
-          return Result.error(new KnownErrors.PasskeyAuthenticationFailed("Failed to get initiation options for passkey authentication"));
-        }
-
-        const { options_json, code } = initiationResult.data;
-
-        // HACK: Override the rpID to be the actual domain
-        if (options_json.rpId !== "THIS_VALUE_WILL_BE_REPLACED.example.com") {
-          throw new StackAssertionError(`Expected returned RP ID from server to equal sentinel, but found ${options_json.rpId}`);
-        }
-        options_json.rpId = window.location.hostname;
-
-        const authentication_response = await startAuthentication({ optionsJSON: options_json });
-        return await this._interface.signInWithPasskey({ authentication_response, code });
-      });
-    } catch (error) {
-      if (error instanceof WebAuthnError) {
-        return Result.error(new KnownErrors.PasskeyWebAuthnError(error.message, error.name));
-      } else {
-        // This should never happen
-        return Result.error(new KnownErrors.PasskeyAuthenticationFailed("Failed to sign in with passkey"));
-      }
-    }
-
-    if (result.status === 'ok') {
-      await this._signInToAccountWithTokens(result.data);
-      await this.redirectToAfterSignIn({ replace: true });
-      return Result.ok(undefined);
-    } else {
-      return Result.error(result.error);
-    }
-  }
-
-
-  async callOAuthCallback() {
-    if (typeof window === "undefined") {
-      throw new Error("callOAuthCallback can currently only be called in a browser environment");
-    }
-    this._ensurePersistentTokenStore();
-    let result;
-    try {
-      result = await this._catchMfaRequiredError(async () => {
-        return await callOAuthCallback(this._interface, this.urls.oauthCallback);
-      });
-    } catch (e) {
-      if (e instanceof KnownErrors.InvalidTotpCode) {
-        alert("Invalid TOTP code. Please try signing in again.");
-        return false;
-      } else {
-        throw e;
-      }
-    }
-    if (result.status === 'ok' && result.data) {
-      await this._signInToAccountWithTokens(result.data);
-      // TODO fix afterCallbackRedirectUrl for MFA (currently not passed because /mfa/sign-in doesn't return it)
-      // or just get rid of afterCallbackRedirectUrl entirely tbh
-      if ("afterCallbackRedirectUrl" in result.data && result.data.afterCallbackRedirectUrl) {
-        await this._redirectTo({ url: result.data.afterCallbackRedirectUrl, replace: true });
-        return true;
-      } else if (result.data.newUser) {
-        await this.redirectToAfterSignUp({ replace: true });
-        return true;
-      } else {
-        await this.redirectToAfterSignIn({ replace: true });
-        return true;
-      }
-    }
-    return false;
-  }
-
-  protected async _signOut(session: InternalSession, options?: { redirectUrl?: URL | string }): Promise<void> {
-    await storeLock.withWriteLock(async () => {
-      await this._interface.signOut(session);
-      if (options?.redirectUrl) {
-        await this._redirectTo({ url: options.redirectUrl, replace: true });
-      } else {
-        await this.redirectToAfterSignOut();
-      }
-    });
-  }
-
-  async signOut(options?: { redirectUrl?: URL | string }): Promise<void> {
-    const user = await this.getUser();
-    if (user) {
-      await user.signOut(options);
-    }
-  }
-
-  async getProject(): Promise<Project> {
-    const crud = Result.orThrow(await this._currentProjectCache.getOrWait([], "write-only"));
-    return this._clientProjectFromCrud(crud);
-  }
-
-  // IF_PLATFORM react-like
-  useProject(): Project {
-    const crud = useAsyncCache(this._currentProjectCache, [], "useProject()");
-    return useMemo(() => this._clientProjectFromCrud(crud), [crud]);
-  }
-  // END_PLATFORM
-
-  protected async _listOwnedProjects(session: InternalSession): Promise<AdminOwnedProject[]> {
-    this._ensureInternalProject();
-    const crud = Result.orThrow(await this._ownedProjectsCache.getOrWait([session], "write-only"));
-    return crud.map((j) => this._getOwnedAdminApp(j.id, session)._adminOwnedProjectFromCrud(
-      j,
-      () => this._refreshOwnedProjects(session),
-    ));
-  }
-
-  // IF_PLATFORM react-like
-  protected _useOwnedProjects(session: InternalSession): AdminOwnedProject[] {
-    this._ensureInternalProject();
-    const projects = useAsyncCache(this._ownedProjectsCache, [session], "useOwnedProjects()");
-    return useMemo(() => projects.map((j) => this._getOwnedAdminApp(j.id, session)._adminOwnedProjectFromCrud(
-      j,
-      () => this._refreshOwnedProjects(session),
-    )), [projects]);
-  }
-  // END_PLATFORM
-  protected async _createProject(session: InternalSession, newProject: AdminProjectUpdateOptions & { displayName: string }): Promise<AdminOwnedProject> {
-    this._ensureInternalProject();
-    const crud = await this._interface.createProject(adminProjectCreateOptionsToCrud(newProject), session);
-    const res = this._getOwnedAdminApp(crud.id, session)._adminOwnedProjectFromCrud(
-      crud,
-      () => this._refreshOwnedProjects(session),
-    );
-    await this._refreshOwnedProjects(session);
-    return res;
-  }
-
-  protected async _refreshUser(session: InternalSession) {
-    // TODO this should take a user ID instead of a session, and automatically refresh all sessions with that user ID
-    await this._refreshSession(session);
-  }
-
-  protected async _refreshSession(session: InternalSession) {
-    await this._currentUserCache.refresh([session]);
-  }
-
-  protected async _refreshUsers() {
-    // nothing yet
-  }
-
-  protected async _refreshProject() {
-    await this._currentProjectCache.refresh([]);
-  }
-
-  protected async _refreshOwnedProjects(session: InternalSession) {
-    await this._ownedProjectsCache.refresh([session]);
-  }
-
-  static get [stackAppInternalsSymbol]() {
-    return {
-      fromClientJson: <HasTokenStore extends boolean, ProjectId extends string>(
-        json: StackClientAppJson<HasTokenStore, ProjectId>
-      ): StackClientApp<HasTokenStore, ProjectId> => {
-        const providedCheckString = JSON.stringify(omit(json, [/* none currently */]));
-        const existing = allClientApps.get(json.uniqueIdentifier);
-        if (existing) {
-          const [existingCheckString, clientApp] = existing;
-          if (existingCheckString !== providedCheckString) {
-            throw new StackAssertionError("The provided app JSON does not match the configuration of the existing client app with the same unique identifier", { providedObj: json, existingString: existingCheckString });
-          }
-          return clientApp as any;
-        }
-
-        return new _StackClientAppImpl<HasTokenStore, ProjectId>({
-          ...json,
-          checkString: providedCheckString,
-        });
-      }
-    };
-  }
-
-  get [stackAppInternalsSymbol]() {
-    return {
-      toClientJson: (): StackClientAppJson<HasTokenStore, ProjectId> => {
-        if (!("publishableClientKey" in this._interface.options)) {
-          // TODO find a way to do this
-          throw new StackAssertionError("Cannot serialize to JSON from an application without a publishable client key");
-        }
-
-        if (typeof this._redirectMethod !== "string") {
-          throw new StackAssertionError("Cannot serialize to JSON from an application with a non-string redirect method");
-        }
-
-        return {
-          baseUrl: this._options.baseUrl,
-          projectId: this.projectId,
-          publishableClientKey: this._interface.options.publishableClientKey,
-          tokenStore: this._tokenStoreInit,
-          urls: this._urlOptions,
-          oauthScopesOnSignIn: this._oauthScopesOnSignIn,
-          uniqueIdentifier: this._getUniqueIdentifier(),
-          redirectMethod: this._redirectMethod,
-        };
-      },
-      setCurrentUser: (userJsonPromise: Promise<CurrentUserCrud['Client']['Read'] | null>) => {
-        runAsynchronously(async () => {
-          await this._currentUserCache.forceSetCachedValueAsync([await this._getSession()], Result.fromPromise(userJsonPromise));
-        });
-      },
-      sendRequest: async (
-        path: string,
-        requestOptions: RequestInit,
-        requestType: "client" | "server" | "admin" = "client",
-      ) => {
-        return await this._interface.sendClientRequest(path, requestOptions, await this._getSession(), requestType);
-      },
-    };
-  };
-}
-
-class _StackServerAppImpl<HasTokenStore extends boolean, ProjectId extends string> extends _StackClientAppImpl<HasTokenStore, ProjectId>
-{
-  declare protected _interface: StackServerInterface;
-
-  // TODO override the client user cache to use the server user cache, so we save some requests
-  private readonly _currentServerUserCache = createCacheBySession(async (session) => {
-    if (session.isKnownToBeInvalid()) {
-      // see comment in _currentUserCache for more details on why we do this
-      return null;
-    }
-    return await this._interface.getServerUserByToken(session);
-  });
-  private readonly _serverUsersCache = createCache<[
-    cursor?: string,
-    limit?: number,
-    orderBy?: 'signedUpAt',
-    desc?: boolean,
-    query?: string,
-  ], UsersCrud['Server']['List']>(async ([cursor, limit, orderBy, desc, query]) => {
-    return await this._interface.listServerUsers({ cursor, limit, orderBy, desc, query });
-  });
-  private readonly _serverUserCache = createCache<string[], UsersCrud['Server']['Read'] | null>(async ([userId]) => {
-    const user = await this._interface.getServerUserById(userId);
-    return Result.or(user, null);
-  });
-  private readonly _serverTeamsCache = createCache<[string | undefined], TeamsCrud['Server']['Read'][]>(async ([userId]) => {
-    return await this._interface.listServerTeams({ userId });
-  });
-  private readonly _serverTeamUserPermissionsCache = createCache<
-    [string, string, boolean],
-    TeamPermissionsCrud['Server']['Read'][]
-  >(async ([teamId, userId, recursive]) => {
-    return await this._interface.listServerTeamPermissions({ teamId, userId, recursive }, null);
-  });
-  private readonly _serverUserOAuthConnectionAccessTokensCache = createCache<[string, string, string], { accessToken: string } | null>(
-    async ([userId, providerId, scope]) => {
-      try {
-        const result = await this._interface.createServerProviderAccessToken(userId, providerId, scope || "");
-        return { accessToken: result.access_token };
-      } catch (err) {
-        if (!(err instanceof KnownErrors.OAuthConnectionDoesNotHaveRequiredScope || err instanceof KnownErrors.OAuthConnectionNotConnectedToUser)) {
-          throw err;
-        }
-      }
-      return null;
-    }
-  );
-  private readonly _serverUserOAuthConnectionCache = createCache<[string, ProviderType, string, boolean], OAuthConnection | null>(
-    async ([userId, providerId, scope, redirect]) => {
-      return await this._getUserOAuthConnectionCacheFn({
-        getUser: async () => Result.orThrow(await this._serverUserCache.getOrWait([userId], "write-only")),
-        getOrWaitOAuthToken: async () => Result.orThrow(await this._serverUserOAuthConnectionAccessTokensCache.getOrWait([userId, providerId, scope || ""] as const, "write-only")),
-        // IF_PLATFORM react-like
-        useOAuthToken: () => useAsyncCache(this._serverUserOAuthConnectionAccessTokensCache, [userId, providerId, scope || ""] as const, "user.useConnectedAccount()"),
-        // END_PLATFORM
-        providerId,
-        scope,
-        redirect,
-        session: null,
-      });
-    }
-  );
-  private readonly _serverTeamMemberProfilesCache = createCache<[string], TeamMemberProfilesCrud['Server']['Read'][]>(
-    async ([teamId]) => {
-      return await this._interface.listServerTeamMemberProfiles({ teamId });
-    }
-  );
-  private readonly _serverTeamInvitationsCache = createCache<[string], TeamInvitationCrud['Server']['Read'][]>(
-    async ([teamId]) => {
-      return await this._interface.listServerTeamInvitations({ teamId });
-    }
-  );
-  private readonly _serverUserTeamProfileCache = createCache<[string, string], TeamMemberProfilesCrud['Client']['Read']>(
-    async ([teamId, userId]) => {
-      return await this._interface.getServerTeamMemberProfile({ teamId, userId });
-    }
-  );
-  private readonly _serverContactChannelsCache = createCache<[string], ContactChannelsCrud['Server']['Read'][]>(
-    async ([userId]) => {
-      return await this._interface.listServerContactChannels(userId);
-    }
-  );
-
-  private async _updateServerUser(userId: string, update: ServerUserUpdateOptions): Promise<UsersCrud['Server']['Read']> {
-    const result = await this._interface.updateServerUser(userId, serverUserUpdateOptionsToCrud(update));
-    await this._refreshUsers();
-    return result;
-  }
-
-  protected _serverEditableTeamProfileFromCrud(crud: TeamMemberProfilesCrud['Client']['Read']): EditableTeamMemberProfile {
-    const app = this;
-    return {
-      displayName: crud.display_name,
-      profileImageUrl: crud.profile_image_url,
-      async update(update: { displayName?: string, profileImageUrl?: string }) {
-        await app._interface.updateServerTeamMemberProfile({
-          teamId: crud.team_id,
-          userId: crud.user_id,
-          profile: {
-            display_name: update.displayName,
-            profile_image_url: update.profileImageUrl,
-          },
-        });
-        await app._serverUserTeamProfileCache.refresh([crud.team_id, crud.user_id]);
-      }
-    };
-  }
-
-  protected _serverContactChannelFromCrud(userId: string, crud: ContactChannelsCrud['Server']['Read']): ServerContactChannel {
-    const app = this;
-    return {
-      id: crud.id,
-      value: crud.value,
-      type: crud.type,
-      isVerified: crud.is_verified,
-      isPrimary: crud.is_primary,
-      usedForAuth: crud.used_for_auth,
-      async sendVerificationEmail(options?: { callbackUrl?: string }) {
-        if (!options?.callbackUrl && !await app._getCurrentUrl()) {
-          throw new Error("Cannot send verification email without a callback URL from the server or without a redirect method. Make sure you pass the `callbackUrl` option: `sendVerificationEmail({ callbackUrl: ... })`");
-        }
-
-        await app._interface.sendServerContactChannelVerificationEmail(userId, crud.id, options?.callbackUrl ?? constructRedirectUrl(app.urls.emailVerification));
-      },
-      async update(data: ServerContactChannelUpdateOptions) {
-        await app._interface.updateServerContactChannel(userId, crud.id, serverContactChannelUpdateOptionsToCrud(data));
-      },
-      async delete() {
-        await app._interface.deleteServerContactChannel(userId, crud.id);
-      },
-    };
-  }
-
-  constructor(options:
-    | StackServerAppConstructorOptions<HasTokenStore, ProjectId>
-    | {
-      interface: StackServerInterface,
-      tokenStore: TokenStoreInit<HasTokenStore>,
-      urls: Partial<HandlerUrls> | undefined,
-      oauthScopesOnSignIn?: Partial<OAuthScopesOnSignIn> | undefined,
-    }
-  ) {
-    super("interface" in options ? {
-      interface: options.interface,
-      tokenStore: options.tokenStore,
-      urls: options.urls,
-      oauthScopesOnSignIn: options.oauthScopesOnSignIn,
-    } : {
-      interface: new StackServerInterface({
-        getBaseUrl: () => getBaseUrl(options.baseUrl),
-        projectId: options.projectId ?? getDefaultProjectId(),
-        clientVersion,
-        publishableClientKey: options.publishableClientKey ?? getDefaultPublishableClientKey(),
-        secretServerKey: options.secretServerKey ?? getDefaultSecretServerKey(),
-      }),
-      baseUrl: options.baseUrl,
-      projectId: options.projectId,
-      publishableClientKey: options.publishableClientKey,
-      tokenStore: options.tokenStore,
-      urls: options.urls ?? {},
-      oauthScopesOnSignIn: options.oauthScopesOnSignIn ?? {},
-      redirectMethod: options.redirectMethod,
-    });
-  }
-
-  protected _serverUserFromCrud(crud: UsersCrud['Server']['Read']): ServerUser {
-    const app = this;
-
-    async function getConnectedAccount(id: ProviderType, options?: { scopes?: string[] }): Promise<OAuthConnection | null>;
-    async function getConnectedAccount(id: ProviderType, options: { or: 'redirect', scopes?: string[] }): Promise<OAuthConnection>;
-    async function getConnectedAccount(id: ProviderType, options?: { or?: 'redirect', scopes?: string[] }): Promise<OAuthConnection | null> {
-      const scopeString = options?.scopes?.join(" ");
-      return Result.orThrow(await app._serverUserOAuthConnectionCache.getOrWait([crud.id, id, scopeString || "", options?.or === 'redirect'], "write-only"));
-    }
-
-    // IF_PLATFORM react-like
-    function useConnectedAccount(id: ProviderType, options?: { scopes?: string[] }): OAuthConnection | null;
-    function useConnectedAccount(id: ProviderType, options: { or: 'redirect', scopes?: string[] }): OAuthConnection;
-    function useConnectedAccount(id: ProviderType, options?: { or?: 'redirect', scopes?: string[] }): OAuthConnection | null {
-      const scopeString = options?.scopes?.join(" ");
-      return useAsyncCache(app._serverUserOAuthConnectionCache, [crud.id, id, scopeString || "", options?.or === 'redirect'] as const, "user.useConnectedAccount()");
-    }
-    // END_PLATFORM
-
-    return {
-      ...super._createBaseUser(crud),
-      lastActiveAt: new Date(crud.last_active_at_millis),
-      serverMetadata: crud.server_metadata,
-      async setPrimaryEmail(email: string | null, options?: { verified?: boolean }) {
-        await app._updateServerUser(crud.id, { primaryEmail: email, primaryEmailVerified: options?.verified });
-      },
-      async grantPermission(scope: Team, permissionId: string): Promise<void> {
-        await app._interface.grantServerTeamUserPermission(scope.id, crud.id, permissionId);
-        for (const recursive of [true, false]) {
-          await app._serverTeamUserPermissionsCache.refresh([scope.id, crud.id, recursive]);
-        }
-      },
-      async revokePermission(scope: Team, permissionId: string): Promise<void> {
-        await app._interface.revokeServerTeamUserPermission(scope.id, crud.id, permissionId);
-        for (const recursive of [true, false]) {
-          await app._serverTeamUserPermissionsCache.refresh([scope.id, crud.id, recursive]);
-        }
-      },
-      async delete() {
-        const res = await app._interface.deleteServerServerUser(crud.id);
-        await app._refreshUsers();
-        return res;
-      },
-      async createSession(options: { expiresInMillis?: number }) {
-        // TODO this should also refresh the access token when it expires (like InternalSession)
-        const tokens = await app._interface.createServerUserSession(crud.id, options.expiresInMillis ?? 1000 * 60 * 60 * 24 * 365);
-        return {
-          async getTokens() {
-            return tokens;
-          },
-        };
-      },
-      async setDisplayName(displayName: string) {
-        return await this.update({ displayName });
-      },
-      async setClientMetadata(metadata: Record<string, any>) {
-        return await this.update({ clientMetadata: metadata });
-      },
-      async setClientReadOnlyMetadata(metadata: Record<string, any>) {
-        return await this.update({ clientReadOnlyMetadata: metadata });
-      },
-      async setServerMetadata(metadata: Record<string, any>) {
-        return await this.update({ serverMetadata: metadata });
-      },
-      async setSelectedTeam(team: Team | null) {
-        return await this.update({ selectedTeamId: team?.id ?? null });
-      },
-      getConnectedAccount,
-      // NEXT_LINE_PLATFORM react-like
-      useConnectedAccount,
-      selectedTeam: crud.selected_team ? app._serverTeamFromCrud(crud.selected_team) : null,
-      async getTeam(teamId: string) {
-        const teams = await this.listTeams();
-        return teams.find((t) => t.id === teamId) ?? null;
-      },
-      // IF_PLATFORM react-like
-      useTeam(teamId: string) {
-        const teams = this.useTeams();
-        return useMemo(() => {
-          return teams.find((t) => t.id === teamId) ?? null;
-        }, [teams, teamId]);
-      },
-      // END_PLATFORM
-      async listTeams() {
-        const teams = Result.orThrow(await app._serverTeamsCache.getOrWait([crud.id], "write-only"));
-        return teams.map((t) => app._serverTeamFromCrud(t));
-      },
-      // IF_PLATFORM react-like
-      useTeams() {
-        const teams = useAsyncCache(app._serverTeamsCache, [crud.id], "user.useTeams()");
-        return useMemo(() => teams.map((t) => app._serverTeamFromCrud(t)), [teams]);
-      },
-      // END_PLATFORM
-      createTeam: async (data: Omit<ServerTeamCreateOptions, "creatorUserId">) => {
-        const team = await app._interface.createServerTeam(serverTeamCreateOptionsToCrud({
-          creatorUserId: crud.id,
-          ...data,
-        }));
-        await app._serverTeamsCache.refresh([undefined]);
-        return app._serverTeamFromCrud(team);
-      },
-      leaveTeam: async (team: Team) => {
-        await app._interface.leaveServerTeam({ teamId: team.id, userId: crud.id });
-        // TODO: refresh cache
-      },
-      async listPermissions(scope: Team, options?: { recursive?: boolean }): Promise<AdminTeamPermission[]> {
-        const recursive = options?.recursive ?? true;
-        const permissions = Result.orThrow(await app._serverTeamUserPermissionsCache.getOrWait([scope.id, crud.id, recursive], "write-only"));
-        return permissions.map((crud) => app._serverPermissionFromCrud(crud));
-      },
-      // IF_PLATFORM react-like
-      usePermissions(scope: Team, options?: { recursive?: boolean }): AdminTeamPermission[] {
-        const recursive = options?.recursive ?? true;
-        const permissions = useAsyncCache(app._serverTeamUserPermissionsCache, [scope.id, crud.id, recursive] as const, "user.usePermissions()");
-        return useMemo(() => permissions.map((crud) => app._serverPermissionFromCrud(crud)), [permissions]);
-      },
-      // END_PLATFORM
-      async getPermission(scope: Team, permissionId: string): Promise<AdminTeamPermission | null> {
-        const permissions = await this.listPermissions(scope);
-        return permissions.find((p) => p.id === permissionId) ?? null;
-      },
-      // IF_PLATFORM react-like
-      usePermission(scope: Team, permissionId: string): AdminTeamPermission | null {
-        const permissions = this.usePermissions(scope);
-        return useMemo(() => permissions.find((p) => p.id === permissionId) ?? null, [permissions, permissionId]);
-      },
-      // END_PLATFORM
-      async hasPermission(scope: Team, permissionId: string): Promise<boolean> {
-        return await this.getPermission(scope, permissionId) !== null;
-      },
-      async update(update: ServerUserUpdateOptions) {
-        await app._updateServerUser(crud.id, update);
-      },
-      async sendVerificationEmail() {
-        return await app._checkFeatureSupport("sendVerificationEmail() on ServerUser", {});
-      },
-      async updatePassword(options: { oldPassword?: string, newPassword: string}) {
-        const result = await this.update({ password: options.newPassword });
-        await app._serverUserCache.refresh([crud.id]);
-        return result;
-      },
-      async setPassword(options: { password: string }) {
-        const result = await this.update(options);
-        await app._serverUserCache.refresh([crud.id]);
-        return result;
-      },
-      async getTeamProfile(team: Team) {
-        const result = Result.orThrow(await app._serverUserTeamProfileCache.getOrWait([team.id, crud.id], "write-only"));
-        return app._serverEditableTeamProfileFromCrud(result);
-      },
-      // IF_PLATFORM react-like
-      useTeamProfile(team: Team) {
-        const result = useAsyncCache(app._serverUserTeamProfileCache, [team.id, crud.id] as const, "user.useTeamProfile()");
-        return useMemo(() => app._serverEditableTeamProfileFromCrud(result), [result]);
-      },
-      // END_PLATFORM
-      async listContactChannels() {
-        const result = Result.orThrow(await app._serverContactChannelsCache.getOrWait([crud.id], "write-only"));
-        return result.map((data) => app._serverContactChannelFromCrud(crud.id, data));
-      },
-      // IF_PLATFORM react-like
-      useContactChannels() {
-        const result = useAsyncCache(app._serverContactChannelsCache, [crud.id] as const, "user.useContactChannels()");
-        return useMemo(() => result.map((data) => app._serverContactChannelFromCrud(crud.id, data)), [result]);
-      },
-      // END_PLATFORM
-      createContactChannel: async (data: ServerContactChannelCreateOptions) => {
-        const contactChannel = await app._interface.createServerContactChannel(serverContactChannelCreateOptionsToCrud(crud.id, data));
-        await app._serverContactChannelsCache.refresh([crud.id]);
-        return app._serverContactChannelFromCrud(crud.id, contactChannel);
-      },
-    };
-  }
-
-  protected _serverTeamUserFromCrud(crud: TeamMemberProfilesCrud["Server"]["Read"]): ServerTeamUser {
-    return {
-      ...this._serverUserFromCrud(crud.user),
-      teamProfile: {
-        displayName: crud.display_name,
-        profileImageUrl: crud.profile_image_url,
-      },
-    };
-  }
-
-  protected _serverTeamInvitationFromCrud(crud: TeamInvitationCrud['Server']['Read']): TeamInvitation {
-    return {
-      id: crud.id,
-      recipientEmail: crud.recipient_email,
-      expiresAt: new Date(crud.expires_at_millis),
-      revoke: async () => {
-        await this._interface.revokeServerTeamInvitation(crud.id, crud.team_id);
-      },
-    };
-  }
-
-  protected override _currentUserFromCrud(crud: UsersCrud['Server']['Read'], session: InternalSession): ProjectCurrentServerUser<ProjectId> {
-    const app = this;
-    const currentUser = {
-      ...this._serverUserFromCrud(crud),
-      ...this._createAuth(session),
-      ...this._isInternalProject() ? this._createInternalUserExtra(session) : {},
-    } satisfies ServerUser;
-
-    Object.freeze(currentUser);
-    return currentUser as ProjectCurrentServerUser<ProjectId>;
-  }
-
-  protected _serverTeamFromCrud(crud: TeamsCrud['Server']['Read']): ServerTeam {
-    const app = this;
-    return {
-      id: crud.id,
-      displayName: crud.display_name,
-      profileImageUrl: crud.profile_image_url,
-      createdAt: new Date(crud.created_at_millis),
-      clientMetadata: crud.client_metadata,
-      clientReadOnlyMetadata: crud.client_read_only_metadata,
-      serverMetadata: crud.server_metadata,
-      async update(update: Partial<ServerTeamUpdateOptions>) {
-        await app._interface.updateServerTeam(crud.id, serverTeamUpdateOptionsToCrud(update));
-        await app._serverTeamsCache.refresh([undefined]);
-      },
-      async delete() {
-        await app._interface.deleteServerTeam(crud.id);
-        await app._serverTeamsCache.refresh([undefined]);
-      },
-      async listUsers() {
-        const result = Result.orThrow(await app._serverTeamMemberProfilesCache.getOrWait([crud.id], "write-only"));
-        return result.map(u => app._serverTeamUserFromCrud(u));
-      },
-      // IF_PLATFORM react-like
-      useUsers() {
-        const result = useAsyncCache(app._serverTeamMemberProfilesCache, [crud.id] as const, "team.useUsers()");
-        return useMemo(() => result.map(u => app._serverTeamUserFromCrud(u)), [result]);
-      },
-      // END_PLATFORM
-      async addUser(userId) {
-        await app._interface.addServerUserToTeam({
-          teamId: crud.id,
-          userId,
-        });
-        await app._serverTeamMemberProfilesCache.refresh([crud.id]);
-      },
-      async removeUser(userId) {
-        await app._interface.removeServerUserFromTeam({
-          teamId: crud.id,
-          userId,
-        });
-        await app._serverTeamMemberProfilesCache.refresh([crud.id]);
-      },
-      async inviteUser(options: { email: string, callbackUrl?: string }) {
-        if (!options.callbackUrl && !await app._getCurrentUrl()) {
-          throw new Error("Cannot invite user without a callback URL from the server or without a redirect method. Make sure you pass the `callbackUrl` option: `inviteUser({ email, callbackUrl: ... })`");
-        }
-
-        await app._interface.sendServerTeamInvitation({
-          teamId: crud.id,
-          email: options.email,
-          callbackUrl: options.callbackUrl ?? constructRedirectUrl(app.urls.teamInvitation),
-        });
-        await app._serverTeamInvitationsCache.refresh([crud.id]);
-      },
-      async listInvitations() {
-        const result = Result.orThrow(await app._serverTeamInvitationsCache.getOrWait([crud.id], "write-only"));
-        return result.map((crud) => app._serverTeamInvitationFromCrud(crud));
-      },
-      // IF_PLATFORM react-like
-      useInvitations() {
-        const result = useAsyncCache(app._serverTeamInvitationsCache, [crud.id] as const, "team.useInvitations()");
-        return useMemo(() => result.map((crud) => app._serverTeamInvitationFromCrud(crud)), [result]);
-      },
-      // END_PLATFORM
-    };
-  }
-
-  async createUser(options: ServerUserCreateOptions): Promise<ServerUser> {
-    const crud = await this._interface.createServerUser(serverUserCreateOptionsToCrud(options));
-    await this._refreshUsers();
-    return this._serverUserFromCrud(crud);
-  }
-
-  async getUser(options: GetUserOptions<HasTokenStore> & { or: 'redirect' }): Promise<ProjectCurrentServerUser<ProjectId>>;
-  async getUser(options: GetUserOptions<HasTokenStore> & { or: 'throw' }): Promise<ProjectCurrentServerUser<ProjectId>>;
-  async getUser(options?: GetUserOptions<HasTokenStore>): Promise<ProjectCurrentServerUser<ProjectId> | null>;
-  async getUser(id: string): Promise<ServerUser | null>;
-  async getUser(options?: string | GetUserOptions<HasTokenStore>): Promise<ProjectCurrentServerUser<ProjectId> | ServerUser | null> {
-    if (typeof options === "string") {
-      return await this.getServerUserById(options);
-    } else {
-      // TODO this code is duplicated from the client app; fix that
-      this._ensurePersistentTokenStore(options?.tokenStore);
-      const session = await this._getSession(options?.tokenStore);
-      const crud = Result.orThrow(await this._currentServerUserCache.getOrWait([session], "write-only"));
-
-      if (crud === null) {
-        switch (options?.or) {
-          case 'redirect': {
-            await this.redirectToSignIn({ replace: true });
-            break;
-          }
-          case 'throw': {
-            throw new Error("User is not signed in but getUser was called with { or: 'throw' }");
-          }
-          default: {
-            return null;
-          }
-        }
-      }
-
-      return crud && this._currentUserFromCrud(crud, session);
-    }
-  }
-
-  async getServerUser(): Promise<ProjectCurrentServerUser<ProjectId> | null> {
-    console.warn("stackServerApp.getServerUser is deprecated; use stackServerApp.getUser instead");
-    return await this.getUser();
-  }
-
-  async getServerUserById(userId: string): Promise<ServerUser | null> {
-    const crud = Result.orThrow(await this._serverUserCache.getOrWait([userId], "write-only"));
-    return crud && this._serverUserFromCrud(crud);
-  }
-
-  // IF_PLATFORM react-like
-  useUser(options: GetUserOptions<HasTokenStore> & { or: 'redirect' }): ProjectCurrentServerUser<ProjectId>;
-  useUser(options: GetUserOptions<HasTokenStore> & { or: 'throw' }): ProjectCurrentServerUser<ProjectId>;
-  useUser(options?: GetUserOptions<HasTokenStore>): ProjectCurrentServerUser<ProjectId> | null;
-  useUser(id: string): ServerUser | null;
-  useUser(options?: GetUserOptions<HasTokenStore> | string): ProjectCurrentServerUser<ProjectId> | ServerUser | null {
-    if (typeof options === "string") {
-      return this.useUserById(options);
-    } else {
-      // TODO this code is duplicated from the client app; fix that
-      this._ensurePersistentTokenStore(options?.tokenStore);
-
-      const session = this._useSession(options?.tokenStore);
-      const crud = useAsyncCache(this._currentServerUserCache, [session], "useUser()");
-
-      if (crud === null) {
-        switch (options?.or) {
-          case 'redirect': {
-            runAsynchronously(this.redirectToSignIn({ replace: true }));
-            suspend();
-            throw new StackAssertionError("suspend should never return");
-          }
-          case 'throw': {
-            throw new Error("User is not signed in but useUser was called with { or: 'throw' }");
-          }
-          case undefined:
-          case "return-null": {
-            // do nothing
-          }
-        }
-      }
-
-      return useMemo(() => {
-        return crud && this._currentUserFromCrud(crud, session);
-      }, [crud, session, options?.or]);
-    }
-  }
-  // END_PLATFORM
-  // IF_PLATFORM react-like
-  useUserById(userId: string): ServerUser | null {
-    const crud = useAsyncCache(this._serverUserCache, [userId], "useUserById()");
-    return useMemo(() => {
-      return crud && this._serverUserFromCrud(crud);
-    }, [crud]);
-  }
-  // END_PLATFORM
-
-  async listUsers(options?: ServerListUsersOptions): Promise<ServerUser[] & { nextCursor: string | null }> {
-    const crud = Result.orThrow(await this._serverUsersCache.getOrWait([options?.cursor, options?.limit, options?.orderBy, options?.desc, options?.query], "write-only"));
-    const result: any = crud.items.map((j) => this._serverUserFromCrud(j));
-    result.nextCursor = crud.pagination?.next_cursor ?? null;
-    return result as any;
-  }
-
-  // IF_PLATFORM react-like
-  useUsers(options?: ServerListUsersOptions): ServerUser[] & { nextCursor: string | null } {
-    const crud = useAsyncCache(this._serverUsersCache, [options?.cursor, options?.limit, options?.orderBy, options?.desc, options?.query] as const, "useServerUsers()");
-    const result: any = crud.items.map((j) => this._serverUserFromCrud(j));
-    result.nextCursor = crud.pagination?.next_cursor ?? null;
-    return result as any;
-  }
-  // END_PLATFORM
-
-  _serverPermissionFromCrud(crud: TeamPermissionsCrud['Server']['Read']): AdminTeamPermission {
-    return {
-      id: crud.id,
-    };
-  }
-
-  _serverTeamPermissionDefinitionFromCrud(crud: TeamPermissionDefinitionsCrud['Admin']['Read']): AdminTeamPermissionDefinition {
-    return {
-      id: crud.id,
-      description: crud.description,
-      containedPermissionIds: crud.contained_permission_ids,
-    };
-  }
-
-  async listTeams(): Promise<ServerTeam[]> {
-    const teams = Result.orThrow(await this._serverTeamsCache.getOrWait([undefined], "write-only"));
-    return teams.map((t) => this._serverTeamFromCrud(t));
-  }
-
-  async createTeam(data: ServerTeamCreateOptions): Promise<ServerTeam> {
-    const team = await this._interface.createServerTeam(serverTeamCreateOptionsToCrud(data));
-    await this._serverTeamsCache.refresh([undefined]);
-    return this._serverTeamFromCrud(team);
-  }
-
-  // IF_PLATFORM react-like
-  useTeams(): ServerTeam[] {
-    const teams = useAsyncCache(this._serverTeamsCache, [undefined], "useServerTeams()");
-    return useMemo(() => {
-      return teams.map((t) => this._serverTeamFromCrud(t));
-    }, [teams]);
-  }
-  // END_PLATFORM
-
-  async getTeam(teamId: string): Promise<ServerTeam | null> {
-    const teams = await this.listTeams();
-    return teams.find((t) => t.id === teamId) ?? null;
-  }
-
-  // IF_PLATFORM react-like
-  useTeam(teamId: string): ServerTeam | null {
-    const teams = this.useTeams();
-    return useMemo(() => {
-      return teams.find((t) => t.id === teamId) ?? null;
-    }, [teams, teamId]);
-  }
-  // END_PLATFORM
-
-  protected override async _refreshSession(session: InternalSession) {
-    await Promise.all([
-      super._refreshUser(session),
-      this._currentServerUserCache.refresh([session]),
-    ]);
-  }
-
-  protected override async _refreshUsers() {
-    await Promise.all([
-      super._refreshUsers(),
-      this._serverUserCache.refreshWhere(() => true),
-      this._serverUsersCache.refreshWhere(() => true),
-      this._serverContactChannelsCache.refreshWhere(() => true),
-    ]);
-  }
-}
-
-class _StackAdminAppImpl<HasTokenStore extends boolean, ProjectId extends string> extends _StackServerAppImpl<HasTokenStore, ProjectId>
-{
-  declare protected _interface: StackAdminInterface;
-
-  private readonly _adminProjectCache = createCache(async () => {
-    return await this._interface.getProject();
-  });
-  private readonly _apiKeysCache = createCache(async () => {
-    return await this._interface.listApiKeys();
-  });
-  private readonly _adminEmailTemplatesCache = createCache(async () => {
-    return await this._interface.listEmailTemplates();
-  });
-  private readonly _adminTeamPermissionDefinitionsCache = createCache(async () => {
-    return await this._interface.listPermissionDefinitions();
-  });
-  private readonly _svixTokenCache = createCache(async () => {
-    return await this._interface.getSvixToken();
-  });
-  private readonly _metricsCache = createCache(async () => {
-    return await this._interface.getMetrics();
-  });
-
-  constructor(options: StackAdminAppConstructorOptions<HasTokenStore, ProjectId>) {
-    super({
-      interface: new StackAdminInterface({
-        getBaseUrl: () => getBaseUrl(options.baseUrl),
-        projectId: options.projectId ?? getDefaultProjectId(),
-        clientVersion,
-        ..."projectOwnerSession" in options ? {
-          projectOwnerSession: options.projectOwnerSession,
-        } : {
-          publishableClientKey: options.publishableClientKey ?? getDefaultPublishableClientKey(),
-          secretServerKey: options.secretServerKey ?? getDefaultSecretServerKey(),
-          superSecretAdminKey: options.superSecretAdminKey ?? getDefaultSuperSecretAdminKey(),
-        },
-      }),
-      baseUrl: options.baseUrl,
-      projectId: options.projectId,
-      tokenStore: options.tokenStore,
-      urls: options.urls,
-      oauthScopesOnSignIn: options.oauthScopesOnSignIn,
-      redirectMethod: options.redirectMethod,
-    });
-  }
-
-  _adminOwnedProjectFromCrud(data: InternalProjectsCrud['Admin']['Read'], onRefresh: () => Promise<void>): AdminOwnedProject {
-    if (this._tokenStoreInit !== null) {
-      throw new StackAssertionError("Owned apps must always have tokenStore === null — did you not create this project with app._createOwnedApp()?");;
-    }
-    return {
-      ...this._adminProjectFromCrud(data, onRefresh),
-      app: this as StackAdminApp<false>,
-    };
-  }
-
-  _adminProjectFromCrud(data: InternalProjectsCrud['Admin']['Read'], onRefresh: () => Promise<void>): AdminProject {
-    if (data.id !== this.projectId) {
-      throw new StackAssertionError(`The project ID of the provided project JSON (${data.id}) does not match the project ID of the app (${this.projectId})!`);
-    }
-
-    const app = this;
-    return {
-      id: data.id,
-      displayName: data.display_name,
-      description: data.description,
-      createdAt: new Date(data.created_at_millis),
-      userCount: data.user_count,
-      isProductionMode: data.is_production_mode,
-      config: {
-        id: data.config.id,
-        signUpEnabled: data.config.sign_up_enabled,
-        credentialEnabled: data.config.credential_enabled,
-        magicLinkEnabled: data.config.magic_link_enabled,
-        passkeyEnabled: data.config.passkey_enabled,
-        clientTeamCreationEnabled: data.config.client_team_creation_enabled,
-        clientUserDeletionEnabled: data.config.client_user_deletion_enabled,
-        allowLocalhost: data.config.allow_localhost,
-        oauthProviders: data.config.oauth_providers.map((p) => ((p.type === 'shared' ? {
-          id: p.id,
-          enabled: p.enabled,
-          type: 'shared',
-        } as const : {
-          id: p.id,
-          enabled: p.enabled,
-          type: 'standard',
-          clientId: p.client_id ?? throwErr("Client ID is missing"),
-          clientSecret: p.client_secret ?? throwErr("Client secret is missing"),
-          facebookConfigId: p.facebook_config_id,
-          microsoftTenantId: p.microsoft_tenant_id,
-        } as const))),
-        emailConfig: data.config.email_config.type === 'shared' ? {
-          type: 'shared'
-        } : {
-          type: 'standard',
-          host: data.config.email_config.host ?? throwErr("Email host is missing"),
-          port: data.config.email_config.port ?? throwErr("Email port is missing"),
-          username: data.config.email_config.username ?? throwErr("Email username is missing"),
-          password: data.config.email_config.password ?? throwErr("Email password is missing"),
-          senderName: data.config.email_config.sender_name ?? throwErr("Email sender name is missing"),
-          senderEmail: data.config.email_config.sender_email ?? throwErr("Email sender email is missing"),
-        },
-        domains: data.config.domains.map((d) => ({
-          domain: d.domain,
-          handlerPath: d.handler_path,
-        })),
-        createTeamOnSignUp: data.config.create_team_on_sign_up,
-        teamCreatorDefaultPermissions: data.config.team_creator_default_permissions,
-        teamMemberDefaultPermissions: data.config.team_member_default_permissions,
-      },
-
-      async update(update: AdminProjectUpdateOptions) {
-        await app._interface.updateProject(adminProjectUpdateOptionsToCrud(update));
-        await onRefresh();
-      },
-      async delete() {
-        await app._interface.deleteProject();
-      },
-      async getProductionModeErrors() {
-        return getProductionModeErrors(data);
-      },
-      useProductionModeErrors() {
-        return getProductionModeErrors(data);
-      },
-    };
-  }
-
-  _adminEmailTemplateFromCrud(data: EmailTemplateCrud['Admin']['Read']): AdminEmailTemplate {
-    return {
-      type: data.type,
-      subject: data.subject,
-      content: data.content,
-      isDefault: data.is_default,
-    };
-  }
-
-  override async getProject(): Promise<AdminProject> {
-    return this._adminProjectFromCrud(
-      Result.orThrow(await this._adminProjectCache.getOrWait([], "write-only")),
-      () => this._refreshProject()
-    );
-  }
-
-  // IF_PLATFORM react-like
-  override useProject(): AdminProject {
-    const crud = useAsyncCache(this._adminProjectCache, [], "useProjectAdmin()");
-    return useMemo(() => this._adminProjectFromCrud(
-      crud,
-      () => this._refreshProject()
-    ), [crud]);
-  }
-  // END_PLATFORM
-
-  protected _createApiKeyBaseFromCrud(data: ApiKeyBaseCrudRead): ApiKeyBase {
-    const app = this;
-    return {
-      id: data.id,
-      description: data.description,
-      expiresAt: new Date(data.expires_at_millis),
-      manuallyRevokedAt: data.manually_revoked_at_millis ? new Date(data.manually_revoked_at_millis) : null,
-      createdAt: new Date(data.created_at_millis),
-      isValid() {
-        return this.whyInvalid() === null;
-      },
-      whyInvalid() {
-        if (this.expiresAt.getTime() < Date.now()) return "expired";
-        if (this.manuallyRevokedAt) return "manually-revoked";
-        return null;
-      },
-      async revoke() {
-        const res = await app._interface.revokeApiKeyById(data.id);
-        await app._refreshApiKeys();
-        return res;
-      }
-    };
-  }
-
-  protected _createApiKeyFromCrud(data: ApiKeysCrud["Admin"]["Read"]): ApiKey {
-    return {
-      ...this._createApiKeyBaseFromCrud(data),
-      publishableClientKey: data.publishable_client_key ? { lastFour: data.publishable_client_key.last_four } : null,
-      secretServerKey: data.secret_server_key ? { lastFour: data.secret_server_key.last_four } : null,
-      superSecretAdminKey: data.super_secret_admin_key ? { lastFour: data.super_secret_admin_key.last_four } : null,
-    };
-  }
-
-  protected _createApiKeyFirstViewFromCrud(data: ApiKeyCreateCrudResponse): ApiKeyFirstView {
-    return {
-      ...this._createApiKeyBaseFromCrud(data),
-      publishableClientKey: data.publishable_client_key,
-      secretServerKey: data.secret_server_key,
-      superSecretAdminKey: data.super_secret_admin_key,
-    };
-  }
-
-  async listApiKeys(): Promise<ApiKey[]> {
-    const crud = Result.orThrow(await this._apiKeysCache.getOrWait([], "write-only"));
-    return crud.map((j) => this._createApiKeyFromCrud(j));
-  }
-
-  // IF_PLATFORM react-like
-  useApiKeys(): ApiKey[] {
-    const crud = useAsyncCache(this._apiKeysCache, [], "useApiKeys()");
-    return useMemo(() => {
-      return crud.map((j) => this._createApiKeyFromCrud(j));
-    }, [crud]);
-  }
-  // END_PLATFORM
-
-  async createApiKey(options: ApiKeyCreateOptions): Promise<ApiKeyFirstView> {
-    const crud = await this._interface.createApiKey(apiKeyCreateOptionsToCrud(options));
-    await this._refreshApiKeys();
-    return this._createApiKeyFirstViewFromCrud(crud);
-  }
-
-  // IF_PLATFORM react-like
-  useEmailTemplates(): AdminEmailTemplate[] {
-    const crud = useAsyncCache(this._adminEmailTemplatesCache, [], "useEmailTemplates()");
-    return useMemo(() => {
-      return crud.map((j) => this._adminEmailTemplateFromCrud(j));
-    }, [crud]);
-  }
-  // END_PLATFORM
-  async listEmailTemplates(): Promise<AdminEmailTemplate[]> {
-    const crud = Result.orThrow(await this._adminEmailTemplatesCache.getOrWait([], "write-only"));
-    return crud.map((j) => this._adminEmailTemplateFromCrud(j));
-  }
-
-  async updateEmailTemplate(type: EmailTemplateType, data: AdminEmailTemplateUpdateOptions): Promise<void> {
-    await this._interface.updateEmailTemplate(type, adminEmailTemplateUpdateOptionsToCrud(data));
-    await this._adminEmailTemplatesCache.refresh([]);
-  }
-
-  async resetEmailTemplate(type: EmailTemplateType) {
-    await this._interface.resetEmailTemplate(type);
-    await this._adminEmailTemplatesCache.refresh([]);
-  }
-
-  async createTeamPermissionDefinition(data: AdminTeamPermissionDefinitionCreateOptions): Promise<AdminTeamPermission>{
-    const crud = await this._interface.createPermissionDefinition(serverTeamPermissionDefinitionCreateOptionsToCrud(data));
-    await this._adminTeamPermissionDefinitionsCache.refresh([]);
-    return this._serverTeamPermissionDefinitionFromCrud(crud);
-  }
-
-  async updateTeamPermissionDefinition(permissionId: string, data: AdminTeamPermissionDefinitionUpdateOptions) {
-    await this._interface.updatePermissionDefinition(permissionId, serverTeamPermissionDefinitionUpdateOptionsToCrud(data));
-    await this._adminTeamPermissionDefinitionsCache.refresh([]);
-  }
-
-  async deleteTeamPermissionDefinition(permissionId: string): Promise<void> {
-    await this._interface.deletePermissionDefinition(permissionId);
-    await this._adminTeamPermissionDefinitionsCache.refresh([]);
-  }
-
-  async listTeamPermissionDefinitions(): Promise<AdminTeamPermissionDefinition[]> {
-    const crud = Result.orThrow(await this._adminTeamPermissionDefinitionsCache.getOrWait([], "write-only"));
-    return crud.map((p) => this._serverTeamPermissionDefinitionFromCrud(p));
-  }
-
-  // IF_PLATFORM react-like
-  useTeamPermissionDefinitions(): AdminTeamPermissionDefinition[] {
-    const crud = useAsyncCache(this._adminTeamPermissionDefinitionsCache, [], "usePermissions()");
-    return useMemo(() => {
-      return crud.map((p) => this._serverTeamPermissionDefinitionFromCrud(p));
-    }, [crud]);
-  }
-  // END_PLATFORM
-  // IF_PLATFORM react-like
-  useSvixToken(): string {
-    const crud = useAsyncCache(this._svixTokenCache, [], "useSvixToken()");
-    return crud.token;
-  }
-  // END_PLATFORM
-
-  protected override async _refreshProject() {
-    await Promise.all([
-      super._refreshProject(),
-      this._adminProjectCache.refresh([]),
-    ]);
-  }
-
-  protected async _refreshApiKeys() {
-    await this._apiKeysCache.refresh([]);
-  }
-
-  get [stackAppInternalsSymbol]() {
-    return {
-      ...super[stackAppInternalsSymbol],
-      // IF_PLATFORM react-like
-      useMetrics: (): any => {
-        return useAsyncCache(this._metricsCache, [], "useMetrics()");
-      }
-      // END_PLATFORM
-    };
-  }
-
-  async sendTestEmail(options: {
-    recipientEmail: string,
-    emailConfig: EmailConfig,
-  }): Promise<Result<undefined, { errorMessage: string }>> {
-    const response = await this._interface.sendTestEmail({
-      recipient_email: options.recipientEmail,
-      email_config: {
-        ...(pick(options.emailConfig, ['host', 'port', 'username', 'password'])),
-        sender_email: options.emailConfig.senderEmail,
-        sender_name: options.emailConfig.senderName,
-      },
-    });
-
-    if (response.success) {
-      return Result.ok(undefined);
-    } else {
-      return Result.error({ errorMessage: response.error_message ?? throwErr("Email test error not specified") });
-    }
-  }
-}
-
-type _______________CONTACT_CHANNEL_______________ = never;  // this is a marker for VSCode's outline view
-
-type ContactChannel = {
-  id: string,
-  value: string,
-  type: 'email',
-  isPrimary: boolean,
-  isVerified: boolean,
-  usedForAuth: boolean,
-
-  sendVerificationEmail(): Promise<void>,
-  update(data: ContactChannelUpdateOptions): Promise<void>,
-  delete(): Promise<void>,
-}
-
-type ContactChannelCreateOptions = {
-  value: string,
-  type: 'email',
-  usedForAuth: boolean,
-}
-
-function contactChannelCreateOptionsToCrud(userId: string, options: ContactChannelCreateOptions): ContactChannelsCrud["Client"]["Create"] {
-  return {
-    value: options.value,
-    type: options.type,
-    used_for_auth: options.usedForAuth,
-    user_id: userId,
-  };
-}
-
-type ContactChannelUpdateOptions = {
-  usedForAuth?: boolean,
-  value?: string,
-  isPrimary?: boolean,
-}
-
-function contactChannelUpdateOptionsToCrud(options: ContactChannelUpdateOptions): ContactChannelsCrud["Client"]["Update"] {
-  return {
-    value: options.value,
-    used_for_auth: options.usedForAuth,
-    is_primary: options.isPrimary,
-  };
-}
-
-type ServerContactChannel = ContactChannel & {
-  update(data: ServerContactChannelUpdateOptions): Promise<void>,
-}
-type ServerContactChannelUpdateOptions = ContactChannelUpdateOptions & {
-  isVerified?: boolean,
-}
-
-function serverContactChannelUpdateOptionsToCrud(options: ServerContactChannelUpdateOptions): ContactChannelsCrud["Server"]["Update"] {
-  return {
-    value: options.value,
-    is_verified: options.isVerified,
-    used_for_auth: options.usedForAuth,
-  };
-}
-
-type ServerContactChannelCreateOptions = ContactChannelCreateOptions & {
-  isVerified?: boolean,
-}
-function serverContactChannelCreateOptionsToCrud(userId: string, options: ServerContactChannelCreateOptions): ContactChannelsCrud["Server"]["Create"] {
-  return {
-    type: options.type,
-    value: options.value,
-    is_verified: options.isVerified,
-    user_id: userId,
-    used_for_auth: options.usedForAuth,
-  };
-}
-
-type _______________USER_______________ = never;  // this is a marker for VSCode's outline view
-type ___________client_user = never;  // this is a marker for VSCode's outline view
-
-type Session = {
-  getTokens(): Promise<{ accessToken: string | null, refreshToken: string | null }>,
-};
-
-/**
- * Contains everything related to the current user session.
- */
-type Auth = {
-  readonly _internalSession: InternalSession,
-  readonly currentSession: Session,
-  signOut(options?: { redirectUrl?: URL | string }): Promise<void>,
-
-  /**
-   * Returns headers for sending authenticated HTTP requests to external servers. Most commonly used in cross-origin
-   * requests. Similar to `getAuthJson`, but specifically for HTTP requests.
-   *
-   * If you are using `tokenStore: "cookie"`, you don't need this for same-origin requests. However, most
-   * browsers now disable third-party cookies by default, so we must pass authentication tokens by header instead
-   * if the client and server are on different hostnames.
-   *
-   * This function returns a header object that can be used with `fetch` or other HTTP request libraries to send
-   * authenticated requests.
-   *
-   * On the server, you can then pass in the `Request` object to the `tokenStore` option
-   * of your Stack app. Please note that CORS does not allow most headers by default, so you
-   * must include `x-stack-auth` in the [`Access-Control-Allow-Headers` header](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Headers)
-   * of the CORS preflight response.
-   *
-   * If you are not using HTTP (and hence cannot set headers), you will need to use the `getAuthJson()` function
-   * instead.
-   *
-   * Example:
-   *
-   * ```ts
-   * // client
-   * const res = await fetch("https://api.example.com", {
-   *   headers: {
-   *     ...await stackApp.getAuthHeaders()
-   *     // you can also add your own headers here
-   *   },
-   * });
-   *
-   * // server
-   * function handleRequest(req: Request) {
-   *   const user = await stackServerApp.getUser({ tokenStore: req });
-   *   return new Response("Welcome, " + user.displayName);
-   * }
-   * ```
-   */
-  getAuthHeaders(): Promise<{ "x-stack-auth": string }>,
-
-  /**
-   * Creates a JSON-serializable object containing the information to authenticate a user on an external server.
-   * Similar to `getAuthHeaders`, but returns an object that can be sent over any protocol instead of just
-   * HTTP headers.
-   *
-   * While `getAuthHeaders` is the recommended way to send authentication tokens over HTTP, your app may use
-   * a different protocol, for example WebSockets or gRPC. This function returns a token object that can be JSON-serialized and sent to the server in any way you like.
-   *
-   * On the server, you can pass in this token object into the `tokenStore` option to fetch user details.
-   *
-   * Example:
-   *
-   * ```ts
-   * // client
-   * const res = await rpcCall(rpcEndpoint, {
-   *   data: {
-   *     auth: await stackApp.getAuthJson(),
-   *   },
-   * });
-   *
-   * // server
-   * function handleRequest(data) {
-   *   const user = await stackServerApp.getUser({ tokenStore: data.auth });
-   *   return new Response("Welcome, " + user.displayName);
-   * }
-   * ```
-   */
-  getAuthJson(): Promise<{ accessToken: string | null, refreshToken: string | null }>,
-  registerPasskey(options?: { hostname?: string }): Promise<Result<undefined, KnownErrors["PasskeyRegistrationFailed"] | KnownErrors["PasskeyWebAuthnError"]>>,
-};
-
-/**
- * ```
- * +----------+-------------+-------------------+
- * |    \     |   !Server   |      Server       |
- * +----------+-------------+-------------------+
- * | !Session | User        | ServerUser        |
- * | Session  | CurrentUser | CurrentServerUser |
- * +----------+-------------+-------------------+
- * ```
- *
- * The fields on each of these types are available iff:
- * BaseUser: true
- * Auth: Session
- * ServerBaseUser: Server
- * UserExtra: Session OR Server
- *
- * The types are defined as follows (in the typescript manner):
- * User = BaseUser
- * CurrentUser = BaseUser & Auth & UserExtra
- * ServerUser = BaseUser & ServerBaseUser & UserExtra
- * CurrentServerUser = BaseUser & ServerBaseUser & Auth & UserExtra
- **/
-
-type BaseUser = {
-  readonly id: string,
-
-  readonly displayName: string | null,
-
-  /**
-   * The user's email address.
-   *
-   * Note: This might NOT be unique across multiple users, so always use `id` for unique identification.
-   */
-  readonly primaryEmail: string | null,
-  readonly primaryEmailVerified: boolean,
-  readonly profileImageUrl: string | null,
-
-  readonly signedUpAt: Date,
-
-  readonly clientMetadata: any,
-  readonly clientReadOnlyMetadata: any,
-
-  /**
-   * Whether the user has a password set.
-   */
-  readonly hasPassword: boolean,
-  readonly otpAuthEnabled: boolean,
-  readonly passkeyAuthEnabled: boolean,
-
-  readonly isMultiFactorRequired: boolean,
-  toClientJson(): CurrentUserCrud["Client"]["Read"],
-
-  /**
-   * @deprecated, use contact channel's usedForAuth instead
-   */
-  readonly emailAuthEnabled: boolean,
-  /**
-   * @deprecated
-   */
-  readonly oauthProviders: readonly { id: string }[],
-}
-
-type UserExtra = {
-  setDisplayName(displayName: string): Promise<void>,
-  /** @deprecated Use contact channel's sendVerificationEmail instead */
-  sendVerificationEmail(): Promise<KnownErrors["EmailAlreadyVerified"] | void>,
-  setClientMetadata(metadata: any): Promise<void>,
-  updatePassword(options: { oldPassword: string, newPassword: string}): Promise<KnownErrors["PasswordConfirmationMismatch"] | KnownErrors["PasswordRequirementsNotMet"] | void>,
-  setPassword(options: { password: string }): Promise<KnownErrors["PasswordRequirementsNotMet"] | void>,
-
-  /**
-   * A shorthand method to update multiple fields of the user at once.
-   */
-  update(update: UserUpdateOptions): Promise<void>,
-
-  // NEXT_LINE_PLATFORM react-like
-  useContactChannels(): ContactChannel[],
-  listContactChannels(): Promise<ContactChannel[]>,
-  createContactChannel(data: ContactChannelCreateOptions): Promise<ContactChannel>,
-
-  delete(): Promise<void>,
-
-  getConnectedAccount(id: ProviderType, options: { or: 'redirect', scopes?: string[] }): Promise<OAuthConnection>,
-  getConnectedAccount(id: ProviderType, options?: { or?: 'redirect' | 'throw' | 'return-null', scopes?: string[] }): Promise<OAuthConnection | null>,
-
-  // IF_PLATFORM react-like
-  useConnectedAccount(id: ProviderType, options: { or: 'redirect', scopes?: string[] }): OAuthConnection,
-  useConnectedAccount(id: ProviderType, options?: { or?: 'redirect' | 'throw' | 'return-null', scopes?: string[] }): OAuthConnection | null,
-  // END_PLATFORM
-
-  hasPermission(scope: Team, permissionId: string): Promise<boolean>,
-
-  readonly selectedTeam: Team | null,
-  setSelectedTeam(team: Team | null): Promise<void>,
-  createTeam(data: TeamCreateOptions): Promise<Team>,
-  leaveTeam(team: Team): Promise<void>,
-
-  getTeamProfile(team: Team): Promise<EditableTeamMemberProfile>,
-  // NEXT_LINE_PLATFORM react-like
-  useTeamProfile(team: Team): EditableTeamMemberProfile,
-}
-& AsyncStoreProperty<"team", [id: string], Team | null, false>
-& AsyncStoreProperty<"teams", [], Team[], true>
-& AsyncStoreProperty<"permission", [scope: Team, permissionId: string, options?: { recursive?: boolean }], TeamPermission | null, false>
-& AsyncStoreProperty<"permissions", [scope: Team, options?: { recursive?: boolean }], TeamPermission[], true>;
-
-type InternalUserExtra =
-  & {
-    createProject(newProject: AdminProjectUpdateOptions & { displayName: string }): Promise<AdminProject>,
-  }
-  & AsyncStoreProperty<"ownedProjects", [], AdminOwnedProject[], true>
-
-export type User = BaseUser;
-
-export type CurrentUser = BaseUser & Auth & UserExtra;
-
-export type CurrentInternalUser = CurrentUser & InternalUserExtra;
-
-type UserUpdateOptions = {
-  displayName?: string,
-  clientMetadata?: ReadonlyJson,
-  selectedTeamId?: string | null,
-  totpMultiFactorSecret?: Uint8Array | null,
-  profileImageUrl?: string | null,
-  otpAuthEnabled?: boolean,
-  passkeyAuthEnabled?:boolean,
-}
-function userUpdateOptionsToCrud(options: UserUpdateOptions): CurrentUserCrud["Client"]["Update"] {
-  return {
-    display_name: options.displayName,
-    client_metadata: options.clientMetadata,
-    selected_team_id: options.selectedTeamId,
-    totp_secret_base64: options.totpMultiFactorSecret != null ? encodeBase64(options.totpMultiFactorSecret) : options.totpMultiFactorSecret,
-    profile_image_url: options.profileImageUrl,
-    otp_auth_enabled: options.otpAuthEnabled,
-    passkey_auth_enabled: options.passkeyAuthEnabled,
-  };
-}
-
-
-type ___________server_user = never;  // this is a marker for VSCode's outline view
-
-type ServerBaseUser = {
-  setPrimaryEmail(email: string | null, options?: { verified?: boolean | undefined }): Promise<void>,
-
-  readonly lastActiveAt: Date,
-
-  readonly serverMetadata: any,
-  setServerMetadata(metadata: any): Promise<void>,
-  setClientReadOnlyMetadata(metadata: any): Promise<void>,
-
-  createTeam(data: Omit<ServerTeamCreateOptions, "creatorUserId">): Promise<ServerTeam>,
-
-  // NEXT_LINE_PLATFORM react-like
-  useContactChannels(): ServerContactChannel[],
-  listContactChannels(): Promise<ServerContactChannel[]>,
-  createContactChannel(data: ServerContactChannelCreateOptions): Promise<ServerContactChannel>,
-
-  update(user: ServerUserUpdateOptions): Promise<void>,
-
-  grantPermission(scope: Team, permissionId: string): Promise<void>,
-  revokePermission(scope: Team, permissionId: string): Promise<void>,
-
-  /**
-   * Creates a new session object with a refresh token for this user. Can be used to impersonate them.
-   */
-  createSession(options?: { expiresInMillis?: number }): Promise<Session>,
-}
-& AsyncStoreProperty<"team", [id: string], ServerTeam | null, false>
-& AsyncStoreProperty<"teams", [], ServerTeam[], true>
-& AsyncStoreProperty<"permission", [scope: Team, permissionId: string, options?: { direct?: boolean }], AdminTeamPermission | null, false>
-& AsyncStoreProperty<"permissions", [scope: Team, options?: { direct?: boolean }], AdminTeamPermission[], true>;
-
-/**
- * A user including sensitive fields that should only be used on the server, never sent to the client
- * (such as sensitive information and serverMetadata).
- */
-export type ServerUser = ServerBaseUser & BaseUser & UserExtra;
-
-export type CurrentServerUser = Auth & ServerUser;
-
-export type CurrentInternalServerUser = CurrentServerUser & InternalUserExtra;
-
-type ServerUserUpdateOptions = {
-  primaryEmail?: string | null,
-  primaryEmailVerified?: boolean,
-  primaryEmailAuthEnabled?: boolean,
-  clientReadOnlyMetadata?: ReadonlyJson,
-  serverMetadata?: ReadonlyJson,
-  password?: string,
-} & UserUpdateOptions;
-function serverUserUpdateOptionsToCrud(options: ServerUserUpdateOptions): CurrentUserCrud["Server"]["Update"] {
-  return {
-    display_name: options.displayName,
-    primary_email: options.primaryEmail,
-    client_metadata: options.clientMetadata,
-    client_read_only_metadata: options.clientReadOnlyMetadata,
-    server_metadata: options.serverMetadata,
-    selected_team_id: options.selectedTeamId,
-    primary_email_auth_enabled: options.primaryEmailAuthEnabled,
-    primary_email_verified: options.primaryEmailVerified,
-    password: options.password,
-    profile_image_url: options.profileImageUrl,
-    totp_secret_base64: options.totpMultiFactorSecret != null ? encodeBase64(options.totpMultiFactorSecret) : options.totpMultiFactorSecret,
-  };
-}
-
-
-type ServerUserCreateOptions = {
-  primaryEmail?: string | null,
-  primaryEmailAuthEnabled?: boolean,
-  password?: string,
-  otpAuthEnabled?: boolean,
-  displayName?: string,
-  primaryEmailVerified?: boolean,
-  clientMetadata?: any,
-  clientReadOnlyMetadata?: any,
-  serverMetadata?: any,
-}
-function serverUserCreateOptionsToCrud(options: ServerUserCreateOptions): UsersCrud["Server"]["Create"] {
-  return {
-    primary_email: options.primaryEmail,
-    password: options.password,
-    otp_auth_enabled: options.otpAuthEnabled,
-    primary_email_auth_enabled: options.primaryEmailAuthEnabled,
-    display_name: options.displayName,
-    primary_email_verified: options.primaryEmailVerified,
-    client_metadata: options.clientMetadata,
-    client_read_only_metadata: options.clientReadOnlyMetadata,
-    server_metadata: options.serverMetadata,
-  };
-}
-
-
-type _______________PROJECT_______________ = never;  // this is a marker for VSCode's outline view
-
-export type Project = {
-  readonly id: string,
-  readonly displayName: string,
-  readonly config: ProjectConfig,
-};
-
-export type AdminProject = {
-  readonly id: string,
-  readonly displayName: string,
-  readonly description: string | null,
-  readonly createdAt: Date,
-  readonly userCount: number,
-  readonly isProductionMode: boolean,
-  readonly config: AdminProjectConfig,
-
-  update(this: AdminProject, update: AdminProjectUpdateOptions): Promise<void>,
-  delete(this: AdminProject): Promise<void>,
-
-  getProductionModeErrors(this: AdminProject): Promise<ProductionModeError[]>,
-  useProductionModeErrors(this: AdminProject): ProductionModeError[],
-} & Project;
-
-export type AdminOwnedProject = {
-  readonly app: StackAdminApp<false>,
-} & AdminProject;
-
-export type AdminProjectUpdateOptions = {
-  displayName?: string,
-  description?: string,
-  isProductionMode?: boolean,
-  config?: AdminProjectConfigUpdateOptions,
-};
-function adminProjectUpdateOptionsToCrud(options: AdminProjectUpdateOptions): ProjectsCrud["Admin"]["Update"] {
-  return {
-    display_name: options.displayName,
-    description: options.description,
-    is_production_mode: options.isProductionMode,
-    config: {
-      domains: options.config?.domains?.map((d) => ({
-        domain: d.domain,
-        handler_path: d.handlerPath
-      })),
-      oauth_providers: options.config?.oauthProviders?.map((p) => ({
-        id: p.id as any,
-        enabled: p.enabled,
-        type: p.type,
-        ...(p.type === 'standard' && {
-          client_id: p.clientId,
-          client_secret: p.clientSecret,
-          facebook_config_id: p.facebookConfigId,
-          microsoft_tenant_id: p.microsoftTenantId,
-        }),
-      })),
-      email_config: options.config?.emailConfig && (
-        options.config.emailConfig.type === 'shared' ? {
-          type: 'shared',
-        } : {
-          type: 'standard',
-          host: options.config.emailConfig.host,
-          port: options.config.emailConfig.port,
-          username: options.config.emailConfig.username,
-          password: options.config.emailConfig.password,
-          sender_name: options.config.emailConfig.senderName,
-          sender_email: options.config.emailConfig.senderEmail,
-        }
-      ),
-      sign_up_enabled: options.config?.signUpEnabled,
-      credential_enabled: options.config?.credentialEnabled,
-      magic_link_enabled: options.config?.magicLinkEnabled,
-      passkey_enabled: options.config?.passkeyEnabled,
-      allow_localhost: options.config?.allowLocalhost,
-      create_team_on_sign_up: options.config?.createTeamOnSignUp,
-      client_team_creation_enabled: options.config?.clientTeamCreationEnabled,
-      client_user_deletion_enabled: options.config?.clientUserDeletionEnabled,
-      team_creator_default_permissions: options.config?.teamCreatorDefaultPermissions,
-      team_member_default_permissions: options.config?.teamMemberDefaultPermissions,
-    },
-  };
-}
-
-export type AdminProjectCreateOptions = Omit<AdminProjectUpdateOptions, 'displayName'> & {
-  displayName: string,
-};
-function adminProjectCreateOptionsToCrud(options: AdminProjectCreateOptions): InternalProjectsCrud["Server"]["Create"] {
-  return {
-    ...adminProjectUpdateOptionsToCrud(options),
-    display_name: options.displayName,
-  };
-}
-
-type _______________PROJECT_CONFIG_______________ = never;  // this is a marker for VSCode's outline view
-
-export type ProjectConfig = {
-  readonly signUpEnabled: boolean,
-  readonly credentialEnabled: boolean,
-  readonly magicLinkEnabled: boolean,
-  readonly passkeyEnabled: boolean,
-  readonly clientTeamCreationEnabled: boolean,
-  readonly clientUserDeletionEnabled: boolean,
-  readonly oauthProviders: OAuthProviderConfig[],
-};
-
-export type OAuthProviderConfig = {
-  readonly id: string,
-};
-
-export type AdminProjectConfig = {
-  readonly id: string,
-  readonly signUpEnabled: boolean,
-  readonly credentialEnabled: boolean,
-  readonly magicLinkEnabled: boolean,
-  readonly passkeyEnabled: boolean,
-  readonly clientTeamCreationEnabled: boolean,
-  readonly clientUserDeletionEnabled: boolean,
-  readonly allowLocalhost: boolean,
-  readonly oauthProviders: AdminOAuthProviderConfig[],
-  readonly emailConfig?: AdminEmailConfig,
-  readonly domains: AdminDomainConfig[],
-  readonly createTeamOnSignUp: boolean,
-  readonly teamCreatorDefaultPermissions: AdminTeamPermission[],
-  readonly teamMemberDefaultPermissions: AdminTeamPermission[],
-};
-
-export type AdminEmailConfig = (
-  {
-    type: "standard",
-    senderName: string,
-    senderEmail: string,
-    host: string,
-    port: number,
-    username: string,
-    password: string,
-  }
-  | {
-    type: "shared",
-  }
-);
-
-export type AdminDomainConfig = {
-  domain: string,
-  handlerPath: string,
-};
-
-export type AdminOAuthProviderConfig = {
-  id: string,
-  enabled: boolean,
-} & (
-  | { type: 'shared' }
-  | {
-    type: 'standard',
-    clientId: string,
-    clientSecret: string,
-    facebookConfigId?: string,
-    microsoftTenantId?: string,
-  }
-) & OAuthProviderConfig;
-
-export type AdminProjectConfigUpdateOptions = {
-  domains?: {
-    domain: string,
-    handlerPath: string,
-  }[],
-  oauthProviders?: AdminOAuthProviderConfig[],
-  signUpEnabled?: boolean,
-  credentialEnabled?: boolean,
-  magicLinkEnabled?: boolean,
-  passkeyEnabled?: boolean,
-  clientTeamCreationEnabled?: boolean,
-  clientUserDeletionEnabled?: boolean,
-  allowLocalhost?: boolean,
-  createTeamOnSignUp?: boolean,
-  emailConfig?: AdminEmailConfig,
-  teamCreatorDefaultPermissions?: { id: string }[],
-  teamMemberDefaultPermissions?: { id: string }[],
-};
-
-type _______________API_KEYS_______________ = never;  // this is a marker for VSCode's outline view
-
-export type ApiKeyBase = {
-  id: string,
-  description: string,
-  expiresAt: Date,
-  manuallyRevokedAt: Date | null,
-  createdAt: Date,
-  isValid(): boolean,
-  whyInvalid(): "expired" | "manually-revoked" | null,
-  revoke(): Promise<void>,
-};
-
-export type ApiKeyBaseCrudRead = Pick<ApiKeysCrud["Admin"]["Read"], "id" | "created_at_millis" | "description" | "expires_at_millis" | "manually_revoked_at_millis">;
-
-export type ApiKeyFirstView = {
-  publishableClientKey?: string,
-  secretServerKey?: string,
-  superSecretAdminKey?: string,
-} & ApiKeyBase;
-
-export type ApiKey = {
-  publishableClientKey: null | {
-    lastFour: string,
-  },
-  secretServerKey: null | {
-    lastFour: string,
-  },
-  superSecretAdminKey: null | {
-    lastFour: string,
-  },
-} & ApiKeyBase;
-
-export type ApiKeyCreateOptions = {
-  description: string,
-  expiresAt: Date,
-  hasPublishableClientKey: boolean,
-  hasSecretServerKey: boolean,
-  hasSuperSecretAdminKey: boolean,
-};
-function apiKeyCreateOptionsToCrud(options: ApiKeyCreateOptions): ApiKeyCreateCrudRequest {
-  return {
-    description: options.description,
-    expires_at_millis: options.expiresAt.getTime(),
-    has_publishable_client_key: options.hasPublishableClientKey,
-    has_secret_server_key: options.hasSecretServerKey,
-    has_super_secret_admin_key: options.hasSuperSecretAdminKey,
-  };
-}
-
-type _______________TEAM_______________ = never;  // this is a marker for VSCode's outline view
-type ___________client_team = never;  // this is a marker for VSCode's outline view
-
-export type TeamMemberProfile = {
-  displayName: string | null,
-  profileImageUrl: string | null,
-}
-
-type TeamMemberProfileUpdateOptions = {
-  displayName?: string,
-  profileImageUrl?: string | null,
-};
-
-export type EditableTeamMemberProfile = TeamMemberProfile & {
-  update(update: TeamMemberProfileUpdateOptions): Promise<void>,
-}
-
-export type TeamUser = {
-  id: string,
-  teamProfile: TeamMemberProfile,
-}
-
-export type TeamInvitation = {
-  id: string,
-  recipientEmail: string | null,
-  expiresAt: Date,
-  revoke(): Promise<void>,
-}
-
-export type Team = {
-  id: string,
-  displayName: string,
-  profileImageUrl: string | null,
-  clientMetadata: any,
-  clientReadOnlyMetadata: any,
-  inviteUser(options: { email: string, callbackUrl?: string }): Promise<void>,
-  listUsers(): Promise<TeamUser[]>,
-  // NEXT_LINE_PLATFORM react-like
-  useUsers(): TeamUser[],
-  listInvitations(): Promise<TeamInvitation[]>,
-  // NEXT_LINE_PLATFORM react-like
-  useInvitations(): TeamInvitation[],
-  update(update: TeamUpdateOptions): Promise<void>,
-  delete(): Promise<void>,
-};
-
-export type TeamUpdateOptions = {
-  displayName?: string,
-  profileImageUrl?: string | null,
-  clientMetadata?: ReadonlyJson,
-};
-function teamUpdateOptionsToCrud(options: TeamUpdateOptions): TeamsCrud["Client"]["Update"] {
-  return {
-    display_name: options.displayName,
-    profile_image_url: options.profileImageUrl,
-    client_metadata: options.clientMetadata,
-  };
-}
-
-export type TeamCreateOptions = {
-  displayName: string,
-  profileImageUrl?: string,
-}
-function teamCreateOptionsToCrud(options: TeamCreateOptions, creatorUserId: string): TeamsCrud["Client"]["Create"] {
-  return {
-    display_name: options.displayName,
-    profile_image_url: options.profileImageUrl,
-    creator_user_id: creatorUserId,
-  };
-}
-
-type ___________server_team = never;  // this is a marker for VSCode's outline view
-
-export type ServerTeamMemberProfile = TeamMemberProfile;
-
-export type ServerTeamUser = ServerUser & {
-  teamProfile: ServerTeamMemberProfile,
-}
-
-export type ServerTeam = {
-  createdAt: Date,
-  serverMetadata: any,
-  listUsers(): Promise<ServerTeamUser[]>,
-  // NEXT_LINE_PLATFORM react-like
-  useUsers(): ServerUser[],
-  update(update: ServerTeamUpdateOptions): Promise<void>,
-  delete(): Promise<void>,
-  addUser(userId: string): Promise<void>,
-  inviteUser(options: { email: string, callbackUrl?: string }): Promise<void>,
-  removeUser(userId: string): Promise<void>,
-} & Team;
-
-export type ServerListUsersOptions = {
-  cursor?: string,
-  limit?: number,
-  orderBy?: 'signedUpAt',
-  desc?: boolean,
-  query?: string,
-};
-
-export type ServerTeamCreateOptions = TeamCreateOptions & {
-  creatorUserId?: string,
-};
-function serverTeamCreateOptionsToCrud(options: ServerTeamCreateOptions): TeamsCrud["Server"]["Create"] {
-  return {
-    display_name: options.displayName,
-    profile_image_url: options.profileImageUrl,
-    creator_user_id: options.creatorUserId,
-  };
-}
-
-export type ServerTeamUpdateOptions = TeamUpdateOptions & {
-  clientReadOnlyMetadata?: ReadonlyJson,
-  serverMetadata?: ReadonlyJson,
-};
-function serverTeamUpdateOptionsToCrud(options: ServerTeamUpdateOptions): TeamsCrud["Server"]["Update"] {
-  return {
-    display_name: options.displayName,
-    profile_image_url: options.profileImageUrl,
-    client_metadata: options.clientMetadata,
-    client_read_only_metadata: options.clientReadOnlyMetadata,
-    server_metadata: options.serverMetadata,
-  };
-}
-
-type _______________PERMISSION_______________ = never;  // this is a marker for VSCode's outline view
-
-
-export type TeamPermission = {
-  id: string,
-};
-
-export type AdminTeamPermission = TeamPermission;
-
-export type AdminTeamPermissionDefinition = {
-  id: string,
-  description?: string,
-  containedPermissionIds: string[],
-};
-
-export type AdminTeamPermissionDefinitionCreateOptions = {
-  id: string,
-  description?: string,
-  containedPermissionIds: string[],
-};
-export function serverTeamPermissionDefinitionCreateOptionsToCrud(options: AdminTeamPermissionDefinitionCreateOptions): TeamPermissionDefinitionsCrud["Admin"]["Create"] {
-  return {
-    id: options.id,
-    description: options.description,
-    contained_permission_ids: options.containedPermissionIds,
-  };
-}
-
-export type AdminTeamPermissionDefinitionUpdateOptions = Partial<AdminTeamPermissionDefinitionCreateOptions>;
-export function serverTeamPermissionDefinitionUpdateOptionsToCrud(options: AdminTeamPermissionDefinitionUpdateOptions): TeamPermissionDefinitionsCrud["Admin"]["Update"] {
-  return {
-    id: options.id,
-    description: options.description,
-    contained_permission_ids: options.containedPermissionIds,
-  };
-}
-
-
-type _______________CONNECTED_ACCOUNT_______________ = never;  // this is a marker for VSCode's outline view
-
-
-export type Connection = {
-  id: string,
-};
-
-export type OAuthConnection = {
-  getAccessToken(): Promise<{ accessToken: string }>,
-  // NEXT_LINE_PLATFORM react-like
-  useAccessToken(): { accessToken: string },
-} & Connection;
-
-type _______________STACK_APP_______________ = never;  // this is a marker for VSCode's outline view
-
-export type GetUserOptions<HasTokenStore> =
-  & {
-    or?: 'redirect' | 'throw' | 'return-null',
-    tokenStore?: TokenStoreInit,
-  }
-  & (HasTokenStore extends false ? {
-    tokenStore: TokenStoreInit,
-  } : {});
-
-export type StackClientApp<HasTokenStore extends boolean = boolean, ProjectId extends string = string> = (
-  & {
-    readonly projectId: ProjectId,
-
-    readonly urls: Readonly<HandlerUrls>,
-
-    signInWithOAuth(provider: string): Promise<void>,
-    signInWithCredential(options: { email: string, password: string, noRedirect?: boolean }): Promise<Result<undefined, KnownErrors["EmailPasswordMismatch"] | KnownErrors["InvalidTotpCode"]>>,
-    signUpWithCredential(options: { email: string, password: string, noRedirect?: boolean, verificationCallbackUrl?: string }): Promise<Result<undefined, KnownErrors["UserEmailAlreadyExists"] | KnownErrors["PasswordRequirementsNotMet"]>>,
-    signInWithPasskey(): Promise<Result<undefined, KnownErrors["PasskeyAuthenticationFailed"]| KnownErrors["InvalidTotpCode"] | KnownErrors["PasskeyWebAuthnError"]>>,
-    callOAuthCallback(): Promise<boolean>,
-    sendForgotPasswordEmail(email: string, options?: { callbackUrl?: string }): Promise<Result<undefined, KnownErrors["UserNotFound"]>>,
-    sendMagicLinkEmail(email: string, options?: { callbackUrl?: string }): Promise<Result<{ nonce: string }, KnownErrors["RedirectUrlNotWhitelisted"]>>,
-    resetPassword(options: { code: string, password: string }): Promise<Result<undefined, KnownErrors["VerificationCodeError"]>>,
-    verifyPasswordResetCode(code: string): Promise<Result<undefined, KnownErrors["VerificationCodeError"]>>,
-    verifyTeamInvitationCode(code: string): Promise<Result<undefined, KnownErrors["VerificationCodeError"]>>,
-    acceptTeamInvitation(code: string): Promise<Result<undefined, KnownErrors["VerificationCodeError"]>>,
-    getTeamInvitationDetails(code: string): Promise<Result<{ teamDisplayName: string }, KnownErrors["VerificationCodeError"]>>,
-    verifyEmail(code: string): Promise<Result<undefined, KnownErrors["VerificationCodeError"]>>,
-    signInWithMagicLink(code: string): Promise<Result<undefined, KnownErrors["VerificationCodeError"] | KnownErrors["InvalidTotpCode"]>>,
-
-    redirectToOAuthCallback(): Promise<void>,
-
-    // IF_PLATFORM react-like
-    useUser(options: GetUserOptions<HasTokenStore> & { or: 'redirect' }): ProjectCurrentUser<ProjectId>,
-    useUser(options: GetUserOptions<HasTokenStore> & { or: 'throw' }): ProjectCurrentUser<ProjectId>,
-    useUser(options?: GetUserOptions<HasTokenStore>): ProjectCurrentUser<ProjectId> | null,
-    // END_PLATFORM
-
-    getUser(options: GetUserOptions<HasTokenStore> & { or: 'redirect' }): Promise<ProjectCurrentUser<ProjectId>>,
-    getUser(options: GetUserOptions<HasTokenStore> & { or: 'throw' }): Promise<ProjectCurrentUser<ProjectId>>,
-    getUser(options?: GetUserOptions<HasTokenStore>): Promise<ProjectCurrentUser<ProjectId> | null>,
-
-    // NEXT_LINE_PLATFORM react-like
-    useNavigate(): (to: string) => void,
-
-    [stackAppInternalsSymbol]: {
-      toClientJson(): StackClientAppJson<HasTokenStore, ProjectId>,
-      setCurrentUser(userJsonPromise: Promise<CurrentUserCrud['Client']['Read'] | null>): void,
-    },
-  }
-  & AsyncStoreProperty<"project", [], Project, false>
-  & { [K in `redirectTo${Capitalize<keyof Omit<HandlerUrls, 'handler' | 'oauthCallback'>>}`]: (options?: RedirectToOptions) => Promise<void> }
-);
-export type StackClientAppConstructor = {
-  new <
-    TokenStoreType extends string,
-    HasTokenStore extends (TokenStoreType extends {} ? true : boolean),
-    ProjectId extends string
-  >(options: StackClientAppConstructorOptions<HasTokenStore, ProjectId>): StackClientApp<HasTokenStore, ProjectId>,
-  new (options: StackClientAppConstructorOptions<boolean, string>): StackClientApp<boolean, string>,
-
-  [stackAppInternalsSymbol]: {
-    fromClientJson<HasTokenStore extends boolean, ProjectId extends string>(
-      json: StackClientAppJson<HasTokenStore, ProjectId>
-    ): StackClientApp<HasTokenStore, ProjectId>,
-  },
-};
-export const StackClientApp: StackClientAppConstructor = _StackClientAppImpl;
-
-export type StackServerApp<HasTokenStore extends boolean = boolean, ProjectId extends string = string> = (
-  & {
-    createTeam(data: ServerTeamCreateOptions): Promise<ServerTeam>,
-    /**
-     * @deprecated use `getUser()` instead
-     */
-    getServerUser(): Promise<ProjectCurrentServerUser<ProjectId> | null>,
-
-    createUser(options: ServerUserCreateOptions): Promise<ServerUser>,
-
-    // IF_PLATFORM react-like
-    useUser(options: GetUserOptions<HasTokenStore> & { or: 'redirect' }): ProjectCurrentServerUser<ProjectId>,
-    useUser(options: GetUserOptions<HasTokenStore> & { or: 'throw' }): ProjectCurrentServerUser<ProjectId>,
-    useUser(options?: GetUserOptions<HasTokenStore>): ProjectCurrentServerUser<ProjectId> | null,
-    // END_PLATFORM
-
-    getUser(options: GetUserOptions<HasTokenStore> & { or: 'redirect' }): Promise<ProjectCurrentServerUser<ProjectId>>,
-    getUser(options: GetUserOptions<HasTokenStore> & { or: 'throw' }): Promise<ProjectCurrentServerUser<ProjectId>>,
-    getUser(options?: GetUserOptions<HasTokenStore>): Promise<ProjectCurrentServerUser<ProjectId> | null>,
-
-    // NEXT_LINE_PLATFORM react-like
-    useUsers(options?: ServerListUsersOptions): ServerUser[] & { nextCursor: string | null },
-    listUsers(options?: ServerListUsersOptions): Promise<ServerUser[] & { nextCursor: string | null }>,
-  }
-  & AsyncStoreProperty<"user", [id: string], ServerUser | null, false>
-  & Omit<AsyncStoreProperty<"users", [], ServerUser[], true>, "listUsers" | "useUsers">
-  & AsyncStoreProperty<"team", [id: string], ServerTeam | null, false>
-  & AsyncStoreProperty<"teams", [], ServerTeam[], true>
-  & StackClientApp<HasTokenStore, ProjectId>
-);
-export type StackServerAppConstructor = {
-  new <
-    TokenStoreType extends string,
-    HasTokenStore extends (TokenStoreType extends {} ? true : boolean),
-    ProjectId extends string
-  >(options: StackServerAppConstructorOptions<HasTokenStore, ProjectId>): StackServerApp<HasTokenStore, ProjectId>,
-  new (options: StackServerAppConstructorOptions<boolean, string>): StackServerApp<boolean, string>,
-};
-export const StackServerApp: StackServerAppConstructor = _StackServerAppImpl;
-
-export type StackAdminApp<HasTokenStore extends boolean = boolean, ProjectId extends string = string> = (
-  & AsyncStoreProperty<"project", [], AdminProject, false>
-  & AsyncStoreProperty<"apiKeys", [], ApiKey[], true>
-  & AsyncStoreProperty<"teamPermissionDefinitions", [], AdminTeamPermissionDefinition[], true>
-  & {
-    // NEXT_LINE_PLATFORM react-like
-    useEmailTemplates(): AdminEmailTemplate[],
-    listEmailTemplates(): Promise<AdminEmailTemplate[]>,
-    updateEmailTemplate(type: EmailTemplateType, data: AdminEmailTemplateUpdateOptions): Promise<void>,
-    resetEmailTemplate(type: EmailTemplateType): Promise<void>,
-
-    createApiKey(options: ApiKeyCreateOptions): Promise<ApiKeyFirstView>,
-
-    createTeamPermissionDefinition(data: AdminTeamPermissionDefinitionCreateOptions): Promise<AdminTeamPermission>,
-    updateTeamPermissionDefinition(permissionId: string, data: AdminTeamPermissionDefinitionUpdateOptions): Promise<void>,
-    deleteTeamPermissionDefinition(permissionId: string): Promise<void>,
-
-    // NEXT_LINE_PLATFORM react-like
-    useSvixToken(): string,
-
-    sendTestEmail(options: {
-      recipientEmail: string,
-      emailConfig: EmailConfig,
-    }): Promise<Result<undefined, { errorMessage: string }>>,
-  }
-  & StackServerApp<HasTokenStore, ProjectId>
-);
-export type StackAdminAppConstructor = {
-  new <
-    HasTokenStore extends boolean,
-    ProjectId extends string
-  >(options: StackAdminAppConstructorOptions<HasTokenStore, ProjectId>): StackAdminApp<HasTokenStore, ProjectId>,
-  new (options: StackAdminAppConstructorOptions<boolean, string>): StackAdminApp<boolean, string>,
-};
-export const StackAdminApp: StackAdminAppConstructor = _StackAdminAppImpl;
-
-type _______________EMAIL_TEMPLATE_______________ = never;  // this is a marker for VSCode's outline view
-
-type AdminEmailTemplate = {
-  type: EmailTemplateType,
-  subject: string,
-  content: any,
-  isDefault: boolean,
-}
-
-type AdminEmailTemplateUpdateOptions = {
-  subject?: string,
-  content?: any,
-};
-function adminEmailTemplateUpdateOptionsToCrud(options: AdminEmailTemplateUpdateOptions): EmailTemplateCrud['Admin']['Update'] {
-  return {
-    subject: options.subject,
-    content: options.content,
-  };
-}
-
-type _______________VARIOUS_______________ = never;  // this is a marker for VSCode's outline view
-
-type RedirectToOptions = {
-  replace?: boolean,
-  noRedirectBack?: boolean,
-};
-
-type AsyncStoreProperty<Name extends string, Args extends any[], Value, IsMultiple extends boolean> =
-  & { [key in `${IsMultiple extends true ? "list" : "get"}${Capitalize<Name>}`]: (...args: Args) => Promise<Value> }
-  // NEXT_LINE_PLATFORM react-like
-  & { [key in `use${Capitalize<Name>}`]: (...args: Args) => Value }
-
-type EmailConfig = {
-  host: string,
-  port: number,
-  username: string,
-  password: string,
-  senderEmail: string,
-  senderName: string,
-}
