+++ packages/template/src/lib/stack-app/apps/implementations/server-app-impl.ts
@@ -1,0 +1,664 @@
+import { WebAuthnError, startAuthentication, startRegistration } from "@simplewebauthn/browser";
+import { KnownErrors, StackAdminInterface, StackClientInterface, StackServerInterface } from "@stackframe/stack-shared";
+import { ProductionModeError, getProductionModeErrors } from "@stackframe/stack-shared/dist/helpers/production-mode";
+import { ApiKeyCreateCrudRequest, ApiKeyCreateCrudResponse } from "@stackframe/stack-shared/dist/interface/adminInterface";
+import { ApiKeysCrud } from "@stackframe/stack-shared/dist/interface/crud/api-keys";
+import { ContactChannelsCrud } from "@stackframe/stack-shared/dist/interface/crud/contact-channels";
+import { CurrentUserCrud } from "@stackframe/stack-shared/dist/interface/crud/current-user";
+import { EmailTemplateCrud, EmailTemplateType } from "@stackframe/stack-shared/dist/interface/crud/email-templates";
+import { InternalProjectsCrud, ProjectsCrud } from "@stackframe/stack-shared/dist/interface/crud/projects";
+import { TeamInvitationCrud } from "@stackframe/stack-shared/dist/interface/crud/team-invitation";
+import { TeamMemberProfilesCrud } from "@stackframe/stack-shared/dist/interface/crud/team-member-profiles";
+import { TeamPermissionDefinitionsCrud, TeamPermissionsCrud } from "@stackframe/stack-shared/dist/interface/crud/team-permissions";
+import { TeamsCrud } from "@stackframe/stack-shared/dist/interface/crud/teams";
+import { UsersCrud } from "@stackframe/stack-shared/dist/interface/crud/users";
+import { InternalSession } from "@stackframe/stack-shared/dist/sessions";
+import { encodeBase64 } from "@stackframe/stack-shared/dist/utils/bytes";
+import { AsyncCache } from "@stackframe/stack-shared/dist/utils/caches";
+import { scrambleDuringCompileTime } from "@stackframe/stack-shared/dist/utils/compile-time";
+import { isBrowserLike } from "@stackframe/stack-shared/dist/utils/env";
+import { StackAssertionError, concatStacktraces, throwErr } from "@stackframe/stack-shared/dist/utils/errors";
+import { ReadonlyJson } from "@stackframe/stack-shared/dist/utils/json";
+import { DependenciesMap } from "@stackframe/stack-shared/dist/utils/maps";
+import { ProviderType } from "@stackframe/stack-shared/dist/utils/oauth";
+import { deepPlainEquals, filterUndefined, omit, pick } from "@stackframe/stack-shared/dist/utils/objects";
+import { ReactPromise, neverResolve, runAsynchronously, wait } from "@stackframe/stack-shared/dist/utils/promises";
+import { suspend, suspendIfSsr } from "@stackframe/stack-shared/dist/utils/react";
+import { Result } from "@stackframe/stack-shared/dist/utils/results";
+import { Store, storeLock } from "@stackframe/stack-shared/dist/utils/stores";
+import { mergeScopeStrings } from "@stackframe/stack-shared/dist/utils/strings";
+import { getRelativePart, isRelative } from "@stackframe/stack-shared/dist/utils/urls";
+import { generateUuid } from "@stackframe/stack-shared/dist/utils/uuids";
+import * as cookie from "cookie";
+// NEXT_LINE_PLATFORM react-like
+import React, { useCallback, useMemo } from "react";
+import { constructRedirectUrl } from "../../../../utils/url";
+import { addNewOAuthProviderOrScope, callOAuthCallback, signInWithOAuth } from "../../../auth";
+import { CookieHelper, createBrowserCookieHelper, createCookieHelper, createPlaceholderCookieHelper, deleteCookieClient, getCookieClient, setOrDeleteCookie, setOrDeleteCookieClient } from "../../../cookie";
+import { TokenStoreInit, HandlerUrls, OAuthScopesOnSignIn, GetUserOptions } from "../../common";
+import { OAuthConnection } from "../../connected-accounts";
+import { AdminTeamPermission, AdminTeamPermissionDefinition } from "../../permissions";
+import { EditableTeamMemberProfile, Team, ServerTeamCreateOptions, ServerTeamUser, TeamInvitation, ServerTeam, ServerTeamUpdateOptions, ServerListUsersOptions, serverTeamCreateOptionsToCrud, serverTeamUpdateOptionsToCrud } from "../../teams";
+import { ServerUser, ProjectCurrentServerUser, ServerUserCreateOptions, ServerUserUpdateOptions, serverUserCreateOptionsToCrud, serverUserUpdateOptionsToCrud } from "../../users";
+import { StackServerAppConstructorOptions } from "../interfaces/server-app";
+import { _StackClientAppImpl } from "./client-app-impl";
+import { createCacheBySession, createCache, getDefaultProjectId, getDefaultPublishableClientKey, getDefaultSecretServerKey, getBaseUrl, clientVersion } from "./common";
+// NEXT_LINE_PLATFORM react-like
+import { useAsyncCache } from "./common";
+import { ServerContactChannel, ServerContactChannelUpdateOptions, ServerContactChannelCreateOptions, serverContactChannelCreateOptionsToCrud, serverContactChannelUpdateOptionsToCrud } from "../../contact-channels";
+
+
+export class _StackServerAppImpl<HasTokenStore extends boolean, ProjectId extends string> extends _StackClientAppImpl<HasTokenStore, ProjectId>
+{
+  declare protected _interface: StackServerInterface;
+
+  // TODO override the client user cache to use the server user cache, so we save some requests
+  private readonly _currentServerUserCache = createCacheBySession(async (session) => {
+    if (session.isKnownToBeInvalid()) {
+      // see comment in _currentUserCache for more details on why we do this
+      return null;
+    }
+    return await this._interface.getServerUserByToken(session);
+  });
+  private readonly _serverUsersCache = createCache<[
+    cursor?: string,
+    limit?: number,
+    orderBy?: 'signedUpAt',
+    desc?: boolean,
+    query?: string,
+  ], UsersCrud['Server']['List']>(async ([cursor, limit, orderBy, desc, query]) => {
+    return await this._interface.listServerUsers({ cursor, limit, orderBy, desc, query });
+  });
+  private readonly _serverUserCache = createCache<string[], UsersCrud['Server']['Read'] | null>(async ([userId]) => {
+    const user = await this._interface.getServerUserById(userId);
+    return Result.or(user, null);
+  });
+  private readonly _serverTeamsCache = createCache<[string | undefined], TeamsCrud['Server']['Read'][]>(async ([userId]) => {
+    return await this._interface.listServerTeams({ userId });
+  });
+  private readonly _serverTeamUserPermissionsCache = createCache<
+    [string, string, boolean],
+    TeamPermissionsCrud['Server']['Read'][]
+  >(async ([teamId, userId, recursive]) => {
+    return await this._interface.listServerTeamPermissions({ teamId, userId, recursive }, null);
+  });
+  private readonly _serverUserOAuthConnectionAccessTokensCache = createCache<[string, string, string], { accessToken: string } | null>(
+    async ([userId, providerId, scope]) => {
+      try {
+        const result = await this._interface.createServerProviderAccessToken(userId, providerId, scope || "");
+        return { accessToken: result.access_token };
+      } catch (err) {
+        if (!(err instanceof KnownErrors.OAuthConnectionDoesNotHaveRequiredScope || err instanceof KnownErrors.OAuthConnectionNotConnectedToUser)) {
+          throw err;
+        }
+      }
+      return null;
+    }
+  );
+  private readonly _serverUserOAuthConnectionCache = createCache<[string, ProviderType, string, boolean], OAuthConnection | null>(
+    async ([userId, providerId, scope, redirect]) => {
+      return await this._getUserOAuthConnectionCacheFn({
+        getUser: async () => Result.orThrow(await this._serverUserCache.getOrWait([userId], "write-only")),
+        getOrWaitOAuthToken: async () => Result.orThrow(await this._serverUserOAuthConnectionAccessTokensCache.getOrWait([userId, providerId, scope || ""] as const, "write-only")),
+        // IF_PLATFORM react-like
+        useOAuthToken: () => useAsyncCache(this._serverUserOAuthConnectionAccessTokensCache, [userId, providerId, scope || ""] as const, "user.useConnectedAccount()"),
+        // END_PLATFORM
+        providerId,
+        scope,
+        redirect,
+        session: null,
+      });
+    }
+  );
+  private readonly _serverTeamMemberProfilesCache = createCache<[string], TeamMemberProfilesCrud['Server']['Read'][]>(
+    async ([teamId]) => {
+      return await this._interface.listServerTeamMemberProfiles({ teamId });
+    }
+  );
+  private readonly _serverTeamInvitationsCache = createCache<[string], TeamInvitationCrud['Server']['Read'][]>(
+    async ([teamId]) => {
+      return await this._interface.listServerTeamInvitations({ teamId });
+    }
+  );
+  private readonly _serverUserTeamProfileCache = createCache<[string, string], TeamMemberProfilesCrud['Client']['Read']>(
+    async ([teamId, userId]) => {
+      return await this._interface.getServerTeamMemberProfile({ teamId, userId });
+    }
+  );
+  private readonly _serverContactChannelsCache = createCache<[string], ContactChannelsCrud['Server']['Read'][]>(
+    async ([userId]) => {
+      return await this._interface.listServerContactChannels(userId);
+    }
+  );
+
+  private async _updateServerUser(userId: string, update: ServerUserUpdateOptions): Promise<UsersCrud['Server']['Read']> {
+    const result = await this._interface.updateServerUser(userId, serverUserUpdateOptionsToCrud(update));
+    await this._refreshUsers();
+    return result;
+  }
+
+  protected _serverEditableTeamProfileFromCrud(crud: TeamMemberProfilesCrud['Client']['Read']): EditableTeamMemberProfile {
+    const app = this;
+    return {
+      displayName: crud.display_name,
+      profileImageUrl: crud.profile_image_url,
+      async update(update: { displayName?: string, profileImageUrl?: string }) {
+        await app._interface.updateServerTeamMemberProfile({
+          teamId: crud.team_id,
+          userId: crud.user_id,
+          profile: {
+            display_name: update.displayName,
+            profile_image_url: update.profileImageUrl,
+          },
+        });
+        await app._serverUserTeamProfileCache.refresh([crud.team_id, crud.user_id]);
+      }
+    };
+  }
+
+  protected _serverContactChannelFromCrud(userId: string, crud: ContactChannelsCrud['Server']['Read']): ServerContactChannel {
+    const app = this;
+    return {
+      id: crud.id,
+      value: crud.value,
+      type: crud.type,
+      isVerified: crud.is_verified,
+      isPrimary: crud.is_primary,
+      usedForAuth: crud.used_for_auth,
+      async sendVerificationEmail(options?: { callbackUrl?: string }) {
+        if (!options?.callbackUrl && !await app._getCurrentUrl()) {
+          throw new Error("Cannot send verification email without a callback URL from the server or without a redirect method. Make sure you pass the `callbackUrl` option: `sendVerificationEmail({ callbackUrl: ... })`");
+        }
+
+        await app._interface.sendServerContactChannelVerificationEmail(userId, crud.id, options?.callbackUrl ?? constructRedirectUrl(app.urls.emailVerification));
+      },
+      async update(data: ServerContactChannelUpdateOptions) {
+        await app._interface.updateServerContactChannel(userId, crud.id, serverContactChannelUpdateOptionsToCrud(data));
+      },
+      async delete() {
+        await app._interface.deleteServerContactChannel(userId, crud.id);
+      },
+    };
+  }
+
+  constructor(options:
+    | StackServerAppConstructorOptions<HasTokenStore, ProjectId>
+    | {
+      interface: StackServerInterface,
+      tokenStore: TokenStoreInit<HasTokenStore>,
+      urls: Partial<HandlerUrls> | undefined,
+      oauthScopesOnSignIn?: Partial<OAuthScopesOnSignIn> | undefined,
+    }
+  ) {
+    super("interface" in options ? {
+      interface: options.interface,
+      tokenStore: options.tokenStore,
+      urls: options.urls,
+      oauthScopesOnSignIn: options.oauthScopesOnSignIn,
+    } : {
+      interface: new StackServerInterface({
+        getBaseUrl: () => getBaseUrl(options.baseUrl),
+        projectId: options.projectId ?? getDefaultProjectId(),
+        clientVersion,
+        publishableClientKey: options.publishableClientKey ?? getDefaultPublishableClientKey(),
+        secretServerKey: options.secretServerKey ?? getDefaultSecretServerKey(),
+      }),
+      baseUrl: options.baseUrl,
+      projectId: options.projectId,
+      publishableClientKey: options.publishableClientKey,
+      tokenStore: options.tokenStore,
+      urls: options.urls ?? {},
+      oauthScopesOnSignIn: options.oauthScopesOnSignIn ?? {},
+      redirectMethod: options.redirectMethod,
+    });
+  }
+
+  protected _serverUserFromCrud(crud: UsersCrud['Server']['Read']): ServerUser {
+    const app = this;
+
+    async function getConnectedAccount(id: ProviderType, options?: { scopes?: string[] }): Promise<OAuthConnection | null>;
+    async function getConnectedAccount(id: ProviderType, options: { or: 'redirect', scopes?: string[] }): Promise<OAuthConnection>;
+    async function getConnectedAccount(id: ProviderType, options?: { or?: 'redirect', scopes?: string[] }): Promise<OAuthConnection | null> {
+      const scopeString = options?.scopes?.join(" ");
+      return Result.orThrow(await app._serverUserOAuthConnectionCache.getOrWait([crud.id, id, scopeString || "", options?.or === 'redirect'], "write-only"));
+    }
+
+    // IF_PLATFORM react-like
+    function useConnectedAccount(id: ProviderType, options?: { scopes?: string[] }): OAuthConnection | null;
+    function useConnectedAccount(id: ProviderType, options: { or: 'redirect', scopes?: string[] }): OAuthConnection;
+    function useConnectedAccount(id: ProviderType, options?: { or?: 'redirect', scopes?: string[] }): OAuthConnection | null {
+      const scopeString = options?.scopes?.join(" ");
+      return useAsyncCache(app._serverUserOAuthConnectionCache, [crud.id, id, scopeString || "", options?.or === 'redirect'] as const, "user.useConnectedAccount()");
+    }
+    // END_PLATFORM
+
+    return {
+      ...super._createBaseUser(crud),
+      lastActiveAt: new Date(crud.last_active_at_millis),
+      serverMetadata: crud.server_metadata,
+      async setPrimaryEmail(email: string | null, options?: { verified?: boolean }) {
+        await app._updateServerUser(crud.id, { primaryEmail: email, primaryEmailVerified: options?.verified });
+      },
+      async grantPermission(scope: Team, permissionId: string): Promise<void> {
+        await app._interface.grantServerTeamUserPermission(scope.id, crud.id, permissionId);
+        for (const recursive of [true, false]) {
+          await app._serverTeamUserPermissionsCache.refresh([scope.id, crud.id, recursive]);
+        }
+      },
+      async revokePermission(scope: Team, permissionId: string): Promise<void> {
+        await app._interface.revokeServerTeamUserPermission(scope.id, crud.id, permissionId);
+        for (const recursive of [true, false]) {
+          await app._serverTeamUserPermissionsCache.refresh([scope.id, crud.id, recursive]);
+        }
+      },
+      async delete() {
+        const res = await app._interface.deleteServerServerUser(crud.id);
+        await app._refreshUsers();
+        return res;
+      },
+      async createSession(options: { expiresInMillis?: number }) {
+        // TODO this should also refresh the access token when it expires (like InternalSession)
+        const tokens = await app._interface.createServerUserSession(crud.id, options.expiresInMillis ?? 1000 * 60 * 60 * 24 * 365);
+        return {
+          async getTokens() {
+            return tokens;
+          },
+        };
+      },
+      async setDisplayName(displayName: string) {
+        return await this.update({ displayName });
+      },
+      async setClientMetadata(metadata: Record<string, any>) {
+        return await this.update({ clientMetadata: metadata });
+      },
+      async setClientReadOnlyMetadata(metadata: Record<string, any>) {
+        return await this.update({ clientReadOnlyMetadata: metadata });
+      },
+      async setServerMetadata(metadata: Record<string, any>) {
+        return await this.update({ serverMetadata: metadata });
+      },
+      async setSelectedTeam(team: Team | null) {
+        return await this.update({ selectedTeamId: team?.id ?? null });
+      },
+      getConnectedAccount,
+      // NEXT_LINE_PLATFORM react-like
+      useConnectedAccount,
+      selectedTeam: crud.selected_team ? app._serverTeamFromCrud(crud.selected_team) : null,
+      async getTeam(teamId: string) {
+        const teams = await this.listTeams();
+        return teams.find((t) => t.id === teamId) ?? null;
+      },
+      // IF_PLATFORM react-like
+      useTeam(teamId: string) {
+        const teams = this.useTeams();
+        return useMemo(() => {
+          return teams.find((t) => t.id === teamId) ?? null;
+        }, [teams, teamId]);
+      },
+      // END_PLATFORM
+      async listTeams() {
+        const teams = Result.orThrow(await app._serverTeamsCache.getOrWait([crud.id], "write-only"));
+        return teams.map((t) => app._serverTeamFromCrud(t));
+      },
+      // IF_PLATFORM react-like
+      useTeams() {
+        const teams = useAsyncCache(app._serverTeamsCache, [crud.id], "user.useTeams()");
+        return useMemo(() => teams.map((t) => app._serverTeamFromCrud(t)), [teams]);
+      },
+      // END_PLATFORM
+      createTeam: async (data: Omit<ServerTeamCreateOptions, "creatorUserId">) => {
+        const team = await app._interface.createServerTeam(serverTeamCreateOptionsToCrud({
+          creatorUserId: crud.id,
+          ...data,
+        }));
+        await app._serverTeamsCache.refresh([undefined]);
+        return app._serverTeamFromCrud(team);
+      },
+      leaveTeam: async (team: Team) => {
+        await app._interface.leaveServerTeam({ teamId: team.id, userId: crud.id });
+        // TODO: refresh cache
+      },
+      async listPermissions(scope: Team, options?: { recursive?: boolean }): Promise<AdminTeamPermission[]> {
+        const recursive = options?.recursive ?? true;
+        const permissions = Result.orThrow(await app._serverTeamUserPermissionsCache.getOrWait([scope.id, crud.id, recursive], "write-only"));
+        return permissions.map((crud) => app._serverPermissionFromCrud(crud));
+      },
+      // IF_PLATFORM react-like
+      usePermissions(scope: Team, options?: { recursive?: boolean }): AdminTeamPermission[] {
+        const recursive = options?.recursive ?? true;
+        const permissions = useAsyncCache(app._serverTeamUserPermissionsCache, [scope.id, crud.id, recursive] as const, "user.usePermissions()");
+        return useMemo(() => permissions.map((crud) => app._serverPermissionFromCrud(crud)), [permissions]);
+      },
+      // END_PLATFORM
+      async getPermission(scope: Team, permissionId: string): Promise<AdminTeamPermission | null> {
+        const permissions = await this.listPermissions(scope);
+        return permissions.find((p) => p.id === permissionId) ?? null;
+      },
+      // IF_PLATFORM react-like
+      usePermission(scope: Team, permissionId: string): AdminTeamPermission | null {
+        const permissions = this.usePermissions(scope);
+        return useMemo(() => permissions.find((p) => p.id === permissionId) ?? null, [permissions, permissionId]);
+      },
+      // END_PLATFORM
+      async hasPermission(scope: Team, permissionId: string): Promise<boolean> {
+        return await this.getPermission(scope, permissionId) !== null;
+      },
+      async update(update: ServerUserUpdateOptions) {
+        await app._updateServerUser(crud.id, update);
+      },
+      async sendVerificationEmail() {
+        return await app._checkFeatureSupport("sendVerificationEmail() on ServerUser", {});
+      },
+      async updatePassword(options: { oldPassword?: string, newPassword: string}) {
+        const result = await this.update({ password: options.newPassword });
+        await app._serverUserCache.refresh([crud.id]);
+        return result;
+      },
+      async setPassword(options: { password: string }) {
+        const result = await this.update(options);
+        await app._serverUserCache.refresh([crud.id]);
+        return result;
+      },
+      async getTeamProfile(team: Team) {
+        const result = Result.orThrow(await app._serverUserTeamProfileCache.getOrWait([team.id, crud.id], "write-only"));
+        return app._serverEditableTeamProfileFromCrud(result);
+      },
+      // IF_PLATFORM react-like
+      useTeamProfile(team: Team) {
+        const result = useAsyncCache(app._serverUserTeamProfileCache, [team.id, crud.id] as const, "user.useTeamProfile()");
+        return useMemo(() => app._serverEditableTeamProfileFromCrud(result), [result]);
+      },
+      // END_PLATFORM
+      async listContactChannels() {
+        const result = Result.orThrow(await app._serverContactChannelsCache.getOrWait([crud.id], "write-only"));
+        return result.map((data) => app._serverContactChannelFromCrud(crud.id, data));
+      },
+      // IF_PLATFORM react-like
+      useContactChannels() {
+        const result = useAsyncCache(app._serverContactChannelsCache, [crud.id] as const, "user.useContactChannels()");
+        return useMemo(() => result.map((data) => app._serverContactChannelFromCrud(crud.id, data)), [result]);
+      },
+      // END_PLATFORM
+      createContactChannel: async (data: ServerContactChannelCreateOptions) => {
+        const contactChannel = await app._interface.createServerContactChannel(serverContactChannelCreateOptionsToCrud(crud.id, data));
+        await app._serverContactChannelsCache.refresh([crud.id]);
+        return app._serverContactChannelFromCrud(crud.id, contactChannel);
+      },
+    };
+  }
+
+  protected _serverTeamUserFromCrud(crud: TeamMemberProfilesCrud["Server"]["Read"]): ServerTeamUser {
+    return {
+      ...this._serverUserFromCrud(crud.user),
+      teamProfile: {
+        displayName: crud.display_name,
+        profileImageUrl: crud.profile_image_url,
+      },
+    };
+  }
+
+  protected _serverTeamInvitationFromCrud(crud: TeamInvitationCrud['Server']['Read']): TeamInvitation {
+    return {
+      id: crud.id,
+      recipientEmail: crud.recipient_email,
+      expiresAt: new Date(crud.expires_at_millis),
+      revoke: async () => {
+        await this._interface.revokeServerTeamInvitation(crud.id, crud.team_id);
+      },
+    };
+  }
+
+  protected override _currentUserFromCrud(crud: UsersCrud['Server']['Read'], session: InternalSession): ProjectCurrentServerUser<ProjectId> {
+    const app = this;
+    const currentUser = {
+      ...this._serverUserFromCrud(crud),
+      ...this._createAuth(session),
+      ...this._isInternalProject() ? this._createInternalUserExtra(session) : {},
+    } satisfies ServerUser;
+
+    Object.freeze(currentUser);
+    return currentUser as ProjectCurrentServerUser<ProjectId>;
+  }
+
+  protected _serverTeamFromCrud(crud: TeamsCrud['Server']['Read']): ServerTeam {
+    const app = this;
+    return {
+      id: crud.id,
+      displayName: crud.display_name,
+      profileImageUrl: crud.profile_image_url,
+      createdAt: new Date(crud.created_at_millis),
+      clientMetadata: crud.client_metadata,
+      clientReadOnlyMetadata: crud.client_read_only_metadata,
+      serverMetadata: crud.server_metadata,
+      async update(update: Partial<ServerTeamUpdateOptions>) {
+        await app._interface.updateServerTeam(crud.id, serverTeamUpdateOptionsToCrud(update));
+        await app._serverTeamsCache.refresh([undefined]);
+      },
+      async delete() {
+        await app._interface.deleteServerTeam(crud.id);
+        await app._serverTeamsCache.refresh([undefined]);
+      },
+      async listUsers() {
+        const result = Result.orThrow(await app._serverTeamMemberProfilesCache.getOrWait([crud.id], "write-only"));
+        return result.map(u => app._serverTeamUserFromCrud(u));
+      },
+      // IF_PLATFORM react-like
+      useUsers() {
+        const result = useAsyncCache(app._serverTeamMemberProfilesCache, [crud.id] as const, "team.useUsers()");
+        return useMemo(() => result.map(u => app._serverTeamUserFromCrud(u)), [result]);
+      },
+      // END_PLATFORM
+      async addUser(userId) {
+        await app._interface.addServerUserToTeam({
+          teamId: crud.id,
+          userId,
+        });
+        await app._serverTeamMemberProfilesCache.refresh([crud.id]);
+      },
+      async removeUser(userId) {
+        await app._interface.removeServerUserFromTeam({
+          teamId: crud.id,
+          userId,
+        });
+        await app._serverTeamMemberProfilesCache.refresh([crud.id]);
+      },
+      async inviteUser(options: { email: string, callbackUrl?: string }) {
+        if (!options.callbackUrl && !await app._getCurrentUrl()) {
+          throw new Error("Cannot invite user without a callback URL from the server or without a redirect method. Make sure you pass the `callbackUrl` option: `inviteUser({ email, callbackUrl: ... })`");
+        }
+
+        await app._interface.sendServerTeamInvitation({
+          teamId: crud.id,
+          email: options.email,
+          callbackUrl: options.callbackUrl ?? constructRedirectUrl(app.urls.teamInvitation),
+        });
+        await app._serverTeamInvitationsCache.refresh([crud.id]);
+      },
+      async listInvitations() {
+        const result = Result.orThrow(await app._serverTeamInvitationsCache.getOrWait([crud.id], "write-only"));
+        return result.map((crud) => app._serverTeamInvitationFromCrud(crud));
+      },
+      // IF_PLATFORM react-like
+      useInvitations() {
+        const result = useAsyncCache(app._serverTeamInvitationsCache, [crud.id] as const, "team.useInvitations()");
+        return useMemo(() => result.map((crud) => app._serverTeamInvitationFromCrud(crud)), [result]);
+      },
+      // END_PLATFORM
+    };
+  }
+
+  async createUser(options: ServerUserCreateOptions): Promise<ServerUser> {
+    const crud = await this._interface.createServerUser(serverUserCreateOptionsToCrud(options));
+    await this._refreshUsers();
+    return this._serverUserFromCrud(crud);
+  }
+
+  async getUser(options: GetUserOptions<HasTokenStore> & { or: 'redirect' }): Promise<ProjectCurrentServerUser<ProjectId>>;
+  async getUser(options: GetUserOptions<HasTokenStore> & { or: 'throw' }): Promise<ProjectCurrentServerUser<ProjectId>>;
+  async getUser(options?: GetUserOptions<HasTokenStore>): Promise<ProjectCurrentServerUser<ProjectId> | null>;
+  async getUser(id: string): Promise<ServerUser | null>;
+  async getUser(options?: string | GetUserOptions<HasTokenStore>): Promise<ProjectCurrentServerUser<ProjectId> | ServerUser | null> {
+    if (typeof options === "string") {
+      return await this.getServerUserById(options);
+    } else {
+      // TODO this code is duplicated from the client app; fix that
+      this._ensurePersistentTokenStore(options?.tokenStore);
+      const session = await this._getSession(options?.tokenStore);
+      const crud = Result.orThrow(await this._currentServerUserCache.getOrWait([session], "write-only"));
+
+      if (crud === null) {
+        switch (options?.or) {
+          case 'redirect': {
+            await this.redirectToSignIn({ replace: true });
+            break;
+          }
+          case 'throw': {
+            throw new Error("User is not signed in but getUser was called with { or: 'throw' }");
+          }
+          default: {
+            return null;
+          }
+        }
+      }
+
+      return crud && this._currentUserFromCrud(crud, session);
+    }
+  }
+
+  async getServerUser(): Promise<ProjectCurrentServerUser<ProjectId> | null> {
+    console.warn("stackServerApp.getServerUser is deprecated; use stackServerApp.getUser instead");
+    return await this.getUser();
+  }
+
+  async getServerUserById(userId: string): Promise<ServerUser | null> {
+    const crud = Result.orThrow(await this._serverUserCache.getOrWait([userId], "write-only"));
+    return crud && this._serverUserFromCrud(crud);
+  }
+
+  // IF_PLATFORM react-like
+  useUser(options: GetUserOptions<HasTokenStore> & { or: 'redirect' }): ProjectCurrentServerUser<ProjectId>;
+  useUser(options: GetUserOptions<HasTokenStore> & { or: 'throw' }): ProjectCurrentServerUser<ProjectId>;
+  useUser(options?: GetUserOptions<HasTokenStore>): ProjectCurrentServerUser<ProjectId> | null;
+  useUser(id: string): ServerUser | null;
+  useUser(options?: GetUserOptions<HasTokenStore> | string): ProjectCurrentServerUser<ProjectId> | ServerUser | null {
+    if (typeof options === "string") {
+      return this.useUserById(options);
+    } else {
+      // TODO this code is duplicated from the client app; fix that
+      this._ensurePersistentTokenStore(options?.tokenStore);
+
+      const session = this._useSession(options?.tokenStore);
+      const crud = useAsyncCache(this._currentServerUserCache, [session], "useUser()");
+
+      if (crud === null) {
+        switch (options?.or) {
+          case 'redirect': {
+            runAsynchronously(this.redirectToSignIn({ replace: true }));
+            suspend();
+            throw new StackAssertionError("suspend should never return");
+          }
+          case 'throw': {
+            throw new Error("User is not signed in but useUser was called with { or: 'throw' }");
+          }
+          case undefined:
+          case "return-null": {
+            // do nothing
+          }
+        }
+      }
+
+      return useMemo(() => {
+        return crud && this._currentUserFromCrud(crud, session);
+      }, [crud, session, options?.or]);
+    }
+  }
+  // END_PLATFORM
+  // IF_PLATFORM react-like
+  useUserById(userId: string): ServerUser | null {
+    const crud = useAsyncCache(this._serverUserCache, [userId], "useUserById()");
+    return useMemo(() => {
+      return crud && this._serverUserFromCrud(crud);
+    }, [crud]);
+  }
+  // END_PLATFORM
+
+  async listUsers(options?: ServerListUsersOptions): Promise<ServerUser[] & { nextCursor: string | null }> {
+    const crud = Result.orThrow(await this._serverUsersCache.getOrWait([options?.cursor, options?.limit, options?.orderBy, options?.desc, options?.query], "write-only"));
+    const result: any = crud.items.map((j) => this._serverUserFromCrud(j));
+    result.nextCursor = crud.pagination?.next_cursor ?? null;
+    return result as any;
+  }
+
+  // IF_PLATFORM react-like
+  useUsers(options?: ServerListUsersOptions): ServerUser[] & { nextCursor: string | null } {
+    const crud = useAsyncCache(this._serverUsersCache, [options?.cursor, options?.limit, options?.orderBy, options?.desc, options?.query] as const, "useServerUsers()");
+    const result: any = crud.items.map((j) => this._serverUserFromCrud(j));
+    result.nextCursor = crud.pagination?.next_cursor ?? null;
+    return result as any;
+  }
+  // END_PLATFORM
+
+  _serverPermissionFromCrud(crud: TeamPermissionsCrud['Server']['Read']): AdminTeamPermission {
+    return {
+      id: crud.id,
+    };
+  }
+
+  _serverTeamPermissionDefinitionFromCrud(crud: TeamPermissionDefinitionsCrud['Admin']['Read']): AdminTeamPermissionDefinition {
+    return {
+      id: crud.id,
+      description: crud.description,
+      containedPermissionIds: crud.contained_permission_ids,
+    };
+  }
+
+  async listTeams(): Promise<ServerTeam[]> {
+    const teams = Result.orThrow(await this._serverTeamsCache.getOrWait([undefined], "write-only"));
+    return teams.map((t) => this._serverTeamFromCrud(t));
+  }
+
+  async createTeam(data: ServerTeamCreateOptions): Promise<ServerTeam> {
+    const team = await this._interface.createServerTeam(serverTeamCreateOptionsToCrud(data));
+    await this._serverTeamsCache.refresh([undefined]);
+    return this._serverTeamFromCrud(team);
+  }
+
+  // IF_PLATFORM react-like
+  useTeams(): ServerTeam[] {
+    const teams = useAsyncCache(this._serverTeamsCache, [undefined], "useServerTeams()");
+    return useMemo(() => {
+      return teams.map((t) => this._serverTeamFromCrud(t));
+    }, [teams]);
+  }
+  // END_PLATFORM
+
+  async getTeam(teamId: string): Promise<ServerTeam | null> {
+    const teams = await this.listTeams();
+    return teams.find((t) => t.id === teamId) ?? null;
+  }
+
+  // IF_PLATFORM react-like
+  useTeam(teamId: string): ServerTeam | null {
+    const teams = this.useTeams();
+    return useMemo(() => {
+      return teams.find((t) => t.id === teamId) ?? null;
+    }, [teams, teamId]);
+  }
+  // END_PLATFORM
+
+  protected override async _refreshSession(session: InternalSession) {
+    await Promise.all([
+      super._refreshUser(session),
+      this._currentServerUserCache.refresh([session]),
+    ]);
+  }
+
+  protected override async _refreshUsers() {
+    await Promise.all([
+      super._refreshUsers(),
+      this._serverUserCache.refreshWhere(() => true),
+      this._serverUsersCache.refreshWhere(() => true),
+      this._serverContactChannelsCache.refreshWhere(() => true),
+    ]);
+  }
+}
