+++ packages/template/src/lib/stack-app/apps/index.ts
@@ -1,0 +1,16 @@
+export {
+  StackAdminAppConstructorOptions,
+  StackAdminAppConstructor,
+  StackAdminApp,
+} from "./interfaces/admin-app";
+
+export {
+  StackClientAppConstructorOptions,
+  StackClientAppJson,
+  StackClientApp,
+} from "./interfaces/client-app";
+
+export {
+  StackServerAppConstructorOptions,
+  StackServerApp,
+} from "./interfaces/server-app";
