+++ packages/template/src/lib/stack-app/apps/interfaces/admin-app.ts
@@ -1,0 +1,96 @@
+import { WebAuthnError, startAuthentication, startRegistration } from "@simplewebauthn/browser";
+import { KnownErrors, StackAdminInterface, StackClientInterface, StackServerInterface } from "@stackframe/stack-shared";
+import { ProductionModeError, getProductionModeErrors } from "@stackframe/stack-shared/dist/helpers/production-mode";
+import { ApiKeyCreateCrudRequest, ApiKeyCreateCrudResponse } from "@stackframe/stack-shared/dist/interface/adminInterface";
+import { ApiKeysCrud } from "@stackframe/stack-shared/dist/interface/crud/api-keys";
+import { ContactChannelsCrud } from "@stackframe/stack-shared/dist/interface/crud/contact-channels";
+import { CurrentUserCrud } from "@stackframe/stack-shared/dist/interface/crud/current-user";
+import { EmailTemplateCrud, EmailTemplateType } from "@stackframe/stack-shared/dist/interface/crud/email-templates";
+import { InternalProjectsCrud, ProjectsCrud } from "@stackframe/stack-shared/dist/interface/crud/projects";
+import { TeamInvitationCrud } from "@stackframe/stack-shared/dist/interface/crud/team-invitation";
+import { TeamMemberProfilesCrud } from "@stackframe/stack-shared/dist/interface/crud/team-member-profiles";
+import { TeamPermissionDefinitionsCrud, TeamPermissionsCrud } from "@stackframe/stack-shared/dist/interface/crud/team-permissions";
+import { TeamsCrud } from "@stackframe/stack-shared/dist/interface/crud/teams";
+import { UsersCrud } from "@stackframe/stack-shared/dist/interface/crud/users";
+import { InternalSession } from "@stackframe/stack-shared/dist/sessions";
+import { encodeBase64 } from "@stackframe/stack-shared/dist/utils/bytes";
+import { AsyncCache } from "@stackframe/stack-shared/dist/utils/caches";
+import { scrambleDuringCompileTime } from "@stackframe/stack-shared/dist/utils/compile-time";
+import { isBrowserLike } from "@stackframe/stack-shared/dist/utils/env";
+import { StackAssertionError, concatStacktraces, throwErr } from "@stackframe/stack-shared/dist/utils/errors";
+import { ReadonlyJson } from "@stackframe/stack-shared/dist/utils/json";
+import { DependenciesMap } from "@stackframe/stack-shared/dist/utils/maps";
+import { ProviderType } from "@stackframe/stack-shared/dist/utils/oauth";
+import { deepPlainEquals, filterUndefined, omit, pick } from "@stackframe/stack-shared/dist/utils/objects";
+import { ReactPromise, neverResolve, runAsynchronously, wait } from "@stackframe/stack-shared/dist/utils/promises";
+import { suspend, suspendIfSsr } from "@stackframe/stack-shared/dist/utils/react";
+import { Result } from "@stackframe/stack-shared/dist/utils/results";
+import { Store, storeLock } from "@stackframe/stack-shared/dist/utils/stores";
+import { mergeScopeStrings } from "@stackframe/stack-shared/dist/utils/strings";
+import { getRelativePart, isRelative } from "@stackframe/stack-shared/dist/utils/urls";
+import { generateUuid } from "@stackframe/stack-shared/dist/utils/uuids";
+import * as cookie from "cookie";
+
+import { constructRedirectUrl } from "../../../../utils/url";
+import { addNewOAuthProviderOrScope, callOAuthCallback, signInWithOAuth } from "../../../auth";
+import { CookieHelper, createBrowserCookieHelper, createCookieHelper, createPlaceholderCookieHelper, deleteCookieClient, getCookieClient, setOrDeleteCookie, setOrDeleteCookieClient } from "../../../cookie";
+import { ApiKey, ApiKeyCreateOptions, ApiKeyFirstView } from "../../api-keys";
+import { AsyncStoreProperty, EmailConfig } from "../../common";
+import { AdminEmailTemplate, AdminEmailTemplateUpdateOptions } from "../../email-templates";
+import { AdminTeamPermissionDefinition, AdminTeamPermissionDefinitionCreateOptions, AdminTeamPermission, AdminTeamPermissionDefinitionUpdateOptions } from "../../permissions";
+import { AdminProject } from "../../projects";
+import { _StackAdminAppImpl } from "../implementations/admin-app-impl";
+import { StackServerAppConstructorOptions, StackServerApp } from "./server-app";
+
+
+export type StackAdminAppConstructorOptions<HasTokenStore extends boolean, ProjectId extends string> = (
+  | (
+    & StackServerAppConstructorOptions<HasTokenStore, ProjectId>
+    & {
+      superSecretAdminKey?: string,
+    }
+  )
+  | (
+    & Omit<StackServerAppConstructorOptions<HasTokenStore, ProjectId>, "publishableClientKey" | "secretServerKey">
+    & {
+      projectOwnerSession: InternalSession,
+    }
+  )
+);
+
+
+export type StackAdminApp<HasTokenStore extends boolean = boolean, ProjectId extends string = string> = (
+  & AsyncStoreProperty<"project", [], AdminProject, false>
+  & AsyncStoreProperty<"apiKeys", [], ApiKey[], true>
+  & AsyncStoreProperty<"teamPermissionDefinitions", [], AdminTeamPermissionDefinition[], true>
+  & {
+    // NEXT_LINE_PLATFORM react-like
+    useEmailTemplates(): AdminEmailTemplate[],
+    listEmailTemplates(): Promise<AdminEmailTemplate[]>,
+    updateEmailTemplate(type: EmailTemplateType, data: AdminEmailTemplateUpdateOptions): Promise<void>,
+    resetEmailTemplate(type: EmailTemplateType): Promise<void>,
+
+    createApiKey(options: ApiKeyCreateOptions): Promise<ApiKeyFirstView>,
+
+    createTeamPermissionDefinition(data: AdminTeamPermissionDefinitionCreateOptions): Promise<AdminTeamPermission>,
+    updateTeamPermissionDefinition(permissionId: string, data: AdminTeamPermissionDefinitionUpdateOptions): Promise<void>,
+    deleteTeamPermissionDefinition(permissionId: string): Promise<void>,
+
+    // NEXT_LINE_PLATFORM react-like
+    useSvixToken(): string,
+
+    sendTestEmail(options: {
+      recipientEmail: string,
+      emailConfig: EmailConfig,
+    }): Promise<Result<undefined, { errorMessage: string }>>,
+  }
+  & StackServerApp<HasTokenStore, ProjectId>
+);
+export type StackAdminAppConstructor = {
+  new <
+    HasTokenStore extends boolean,
+    ProjectId extends string
+  >(options: StackAdminAppConstructorOptions<HasTokenStore, ProjectId>): StackAdminApp<HasTokenStore, ProjectId>,
+  new (options: StackAdminAppConstructorOptions<boolean, string>): StackAdminApp<boolean, string>,
+};
+export const StackAdminApp: StackAdminAppConstructor = _StackAdminAppImpl;
