+++ packages/template/src/lib/stack-app/apps/interfaces/server-app.ts
@@ -1,0 +1,85 @@
+import { WebAuthnError, startAuthentication, startRegistration } from "@simplewebauthn/browser";
+import { KnownErrors, StackAdminInterface, StackClientInterface, StackServerInterface } from "@stackframe/stack-shared";
+import { ProductionModeError, getProductionModeErrors } from "@stackframe/stack-shared/dist/helpers/production-mode";
+import { ApiKeyCreateCrudRequest, ApiKeyCreateCrudResponse } from "@stackframe/stack-shared/dist/interface/adminInterface";
+import { ApiKeysCrud } from "@stackframe/stack-shared/dist/interface/crud/api-keys";
+import { ContactChannelsCrud } from "@stackframe/stack-shared/dist/interface/crud/contact-channels";
+import { CurrentUserCrud } from "@stackframe/stack-shared/dist/interface/crud/current-user";
+import { EmailTemplateCrud, EmailTemplateType } from "@stackframe/stack-shared/dist/interface/crud/email-templates";
+import { InternalProjectsCrud, ProjectsCrud } from "@stackframe/stack-shared/dist/interface/crud/projects";
+import { TeamInvitationCrud } from "@stackframe/stack-shared/dist/interface/crud/team-invitation";
+import { TeamMemberProfilesCrud } from "@stackframe/stack-shared/dist/interface/crud/team-member-profiles";
+import { TeamPermissionDefinitionsCrud, TeamPermissionsCrud } from "@stackframe/stack-shared/dist/interface/crud/team-permissions";
+import { TeamsCrud } from "@stackframe/stack-shared/dist/interface/crud/teams";
+import { UsersCrud } from "@stackframe/stack-shared/dist/interface/crud/users";
+import { InternalSession } from "@stackframe/stack-shared/dist/sessions";
+import { encodeBase64 } from "@stackframe/stack-shared/dist/utils/bytes";
+import { AsyncCache } from "@stackframe/stack-shared/dist/utils/caches";
+import { scrambleDuringCompileTime } from "@stackframe/stack-shared/dist/utils/compile-time";
+import { isBrowserLike } from "@stackframe/stack-shared/dist/utils/env";
+import { StackAssertionError, concatStacktraces, throwErr } from "@stackframe/stack-shared/dist/utils/errors";
+import { ReadonlyJson } from "@stackframe/stack-shared/dist/utils/json";
+import { DependenciesMap } from "@stackframe/stack-shared/dist/utils/maps";
+import { ProviderType } from "@stackframe/stack-shared/dist/utils/oauth";
+import { deepPlainEquals, filterUndefined, omit, pick } from "@stackframe/stack-shared/dist/utils/objects";
+import { ReactPromise, neverResolve, runAsynchronously, wait } from "@stackframe/stack-shared/dist/utils/promises";
+import { suspend, suspendIfSsr } from "@stackframe/stack-shared/dist/utils/react";
+import { Result } from "@stackframe/stack-shared/dist/utils/results";
+import { Store, storeLock } from "@stackframe/stack-shared/dist/utils/stores";
+import { mergeScopeStrings } from "@stackframe/stack-shared/dist/utils/strings";
+import { getRelativePart, isRelative } from "@stackframe/stack-shared/dist/utils/urls";
+import { generateUuid } from "@stackframe/stack-shared/dist/utils/uuids";
+import * as cookie from "cookie";
+import { GetUserOptions, AsyncStoreProperty } from "../../common";
+import { ServerTeamCreateOptions, ServerTeam, ServerListUsersOptions } from "../../teams";
+import { ProjectCurrentServerUser, ServerUser, ServerUserCreateOptions } from "../../users";
+import { _StackServerAppImpl } from "../implementations/server-app-impl";
+import { StackClientAppConstructorOptions, StackClientApp } from "./client-app";
+import { constructRedirectUrl } from "../../../../utils/url";
+import { addNewOAuthProviderOrScope, callOAuthCallback, signInWithOAuth } from "../../../auth";
+import { CookieHelper, createBrowserCookieHelper, createCookieHelper, createPlaceholderCookieHelper, deleteCookieClient, getCookieClient, setOrDeleteCookie, setOrDeleteCookieClient } from "../../../cookie";
+
+
+export type StackServerAppConstructorOptions<HasTokenStore extends boolean, ProjectId extends string> = StackClientAppConstructorOptions<HasTokenStore, ProjectId> & {
+  secretServerKey?: string,
+};
+
+export type StackServerApp<HasTokenStore extends boolean = boolean, ProjectId extends string = string> = (
+  & {
+    createTeam(data: ServerTeamCreateOptions): Promise<ServerTeam>,
+    /**
+     * @deprecated use `getUser()` instead
+     */
+    getServerUser(): Promise<ProjectCurrentServerUser<ProjectId> | null>,
+
+    createUser(options: ServerUserCreateOptions): Promise<ServerUser>,
+
+    // IF_PLATFORM react-like
+    useUser(options: GetUserOptions<HasTokenStore> & { or: 'redirect' }): ProjectCurrentServerUser<ProjectId>,
+    useUser(options: GetUserOptions<HasTokenStore> & { or: 'throw' }): ProjectCurrentServerUser<ProjectId>,
+    useUser(options?: GetUserOptions<HasTokenStore>): ProjectCurrentServerUser<ProjectId> | null,
+    // END_PLATFORM
+
+    getUser(options: GetUserOptions<HasTokenStore> & { or: 'redirect' }): Promise<ProjectCurrentServerUser<ProjectId>>,
+    getUser(options: GetUserOptions<HasTokenStore> & { or: 'throw' }): Promise<ProjectCurrentServerUser<ProjectId>>,
+    getUser(options?: GetUserOptions<HasTokenStore>): Promise<ProjectCurrentServerUser<ProjectId> | null>,
+
+    // NEXT_LINE_PLATFORM react-like
+    useUsers(options?: ServerListUsersOptions): ServerUser[] & { nextCursor: string | null },
+    listUsers(options?: ServerListUsersOptions): Promise<ServerUser[] & { nextCursor: string | null }>,
+  }
+  & AsyncStoreProperty<"user", [id: string], ServerUser | null, false>
+  & Omit<AsyncStoreProperty<"users", [], ServerUser[], true>, "listUsers" | "useUsers">
+  & AsyncStoreProperty<"team", [id: string], ServerTeam | null, false>
+  & AsyncStoreProperty<"teams", [], ServerTeam[], true>
+  & StackClientApp<HasTokenStore, ProjectId>
+);
+export type StackServerAppConstructor = {
+  new <
+    TokenStoreType extends string,
+    HasTokenStore extends (TokenStoreType extends {} ? true : boolean),
+    ProjectId extends string
+  >(options: StackServerAppConstructorOptions<HasTokenStore, ProjectId>): StackServerApp<HasTokenStore, ProjectId>,
+  new (options: StackServerAppConstructorOptions<boolean, string>): StackServerApp<boolean, string>,
+};
+export const StackServerApp: StackServerAppConstructor = _StackServerAppImpl;
