+++ packages/template/src/lib/stack-app/projects/index.ts
@@ -1,0 +1,131 @@
+import { WebAuthnError, startAuthentication, startRegistration } from "@simplewebauthn/browser";
+import { KnownErrors, StackAdminInterface, StackClientInterface, StackServerInterface } from "@stackframe/stack-shared";
+import { ProductionModeError, getProductionModeErrors } from "@stackframe/stack-shared/dist/helpers/production-mode";
+import { ApiKeyCreateCrudRequest, ApiKeyCreateCrudResponse } from "@stackframe/stack-shared/dist/interface/adminInterface";
+import { ApiKeysCrud } from "@stackframe/stack-shared/dist/interface/crud/api-keys";
+import { ContactChannelsCrud } from "@stackframe/stack-shared/dist/interface/crud/contact-channels";
+import { CurrentUserCrud } from "@stackframe/stack-shared/dist/interface/crud/current-user";
+import { EmailTemplateCrud, EmailTemplateType } from "@stackframe/stack-shared/dist/interface/crud/email-templates";
+import { InternalProjectsCrud, ProjectsCrud } from "@stackframe/stack-shared/dist/interface/crud/projects";
+import { TeamInvitationCrud } from "@stackframe/stack-shared/dist/interface/crud/team-invitation";
+import { TeamMemberProfilesCrud } from "@stackframe/stack-shared/dist/interface/crud/team-member-profiles";
+import { TeamPermissionDefinitionsCrud, TeamPermissionsCrud } from "@stackframe/stack-shared/dist/interface/crud/team-permissions";
+import { TeamsCrud } from "@stackframe/stack-shared/dist/interface/crud/teams";
+import { UsersCrud } from "@stackframe/stack-shared/dist/interface/crud/users";
+import { InternalSession } from "@stackframe/stack-shared/dist/sessions";
+import { encodeBase64 } from "@stackframe/stack-shared/dist/utils/bytes";
+import { AsyncCache } from "@stackframe/stack-shared/dist/utils/caches";
+import { scrambleDuringCompileTime } from "@stackframe/stack-shared/dist/utils/compile-time";
+import { isBrowserLike } from "@stackframe/stack-shared/dist/utils/env";
+import { StackAssertionError, concatStacktraces, throwErr } from "@stackframe/stack-shared/dist/utils/errors";
+import { ReadonlyJson } from "@stackframe/stack-shared/dist/utils/json";
+import { DependenciesMap } from "@stackframe/stack-shared/dist/utils/maps";
+import { ProviderType } from "@stackframe/stack-shared/dist/utils/oauth";
+import { deepPlainEquals, filterUndefined, omit, pick } from "@stackframe/stack-shared/dist/utils/objects";
+import { ReactPromise, neverResolve, runAsynchronously, wait } from "@stackframe/stack-shared/dist/utils/promises";
+import { suspend, suspendIfSsr } from "@stackframe/stack-shared/dist/utils/react";
+import { Result } from "@stackframe/stack-shared/dist/utils/results";
+import { Store, storeLock } from "@stackframe/stack-shared/dist/utils/stores";
+import { mergeScopeStrings } from "@stackframe/stack-shared/dist/utils/strings";
+import { getRelativePart, isRelative } from "@stackframe/stack-shared/dist/utils/urls";
+import { generateUuid } from "@stackframe/stack-shared/dist/utils/uuids";
+import * as cookie from "cookie";
+
+import { constructRedirectUrl } from "../../../utils/url";
+import { addNewOAuthProviderOrScope, callOAuthCallback, signInWithOAuth } from "../../auth";
+import { CookieHelper, createBrowserCookieHelper, createCookieHelper, createPlaceholderCookieHelper, deleteCookieClient, getCookieClient, setOrDeleteCookie, setOrDeleteCookieClient } from "../../cookie";
+import { AdminProjectConfig, AdminProjectConfigUpdateOptions, ProjectConfig } from "../project-configs";
+import { StackAdminApp } from "../apps/interfaces/admin-app";
+
+
+type _______________PROJECT_______________ = never;  // this is a marker for VSCode's outline view
+
+export type Project = {
+  readonly id: string,
+  readonly displayName: string,
+  readonly config: ProjectConfig,
+};
+
+export type AdminProject = {
+  readonly id: string,
+  readonly displayName: string,
+  readonly description: string | null,
+  readonly createdAt: Date,
+  readonly userCount: number,
+  readonly isProductionMode: boolean,
+  readonly config: AdminProjectConfig,
+
+  update(this: AdminProject, update: AdminProjectUpdateOptions): Promise<void>,
+  delete(this: AdminProject): Promise<void>,
+
+  getProductionModeErrors(this: AdminProject): Promise<ProductionModeError[]>,
+  useProductionModeErrors(this: AdminProject): ProductionModeError[],
+} & Project;
+
+export type AdminOwnedProject = {
+  readonly app: StackAdminApp<false>,
+} & AdminProject;
+
+export type AdminProjectUpdateOptions = {
+  displayName?: string,
+  description?: string,
+  isProductionMode?: boolean,
+  config?: AdminProjectConfigUpdateOptions,
+};
+export function adminProjectUpdateOptionsToCrud(options: AdminProjectUpdateOptions): ProjectsCrud["Admin"]["Update"] {
+  return {
+    display_name: options.displayName,
+    description: options.description,
+    is_production_mode: options.isProductionMode,
+    config: {
+      domains: options.config?.domains?.map((d) => ({
+        domain: d.domain,
+        handler_path: d.handlerPath
+      })),
+      oauth_providers: options.config?.oauthProviders?.map((p) => ({
+        id: p.id as any,
+        enabled: p.enabled,
+        type: p.type,
+        ...(p.type === 'standard' && {
+          client_id: p.clientId,
+          client_secret: p.clientSecret,
+          facebook_config_id: p.facebookConfigId,
+          microsoft_tenant_id: p.microsoftTenantId,
+        }),
+      })),
+      email_config: options.config?.emailConfig && (
+        options.config.emailConfig.type === 'shared' ? {
+          type: 'shared',
+        } : {
+          type: 'standard',
+          host: options.config.emailConfig.host,
+          port: options.config.emailConfig.port,
+          username: options.config.emailConfig.username,
+          password: options.config.emailConfig.password,
+          sender_name: options.config.emailConfig.senderName,
+          sender_email: options.config.emailConfig.senderEmail,
+        }
+      ),
+      sign_up_enabled: options.config?.signUpEnabled,
+      credential_enabled: options.config?.credentialEnabled,
+      magic_link_enabled: options.config?.magicLinkEnabled,
+      passkey_enabled: options.config?.passkeyEnabled,
+      allow_localhost: options.config?.allowLocalhost,
+      create_team_on_sign_up: options.config?.createTeamOnSignUp,
+      client_team_creation_enabled: options.config?.clientTeamCreationEnabled,
+      client_user_deletion_enabled: options.config?.clientUserDeletionEnabled,
+      team_creator_default_permissions: options.config?.teamCreatorDefaultPermissions,
+      team_member_default_permissions: options.config?.teamMemberDefaultPermissions,
+    },
+  };
+}
+
+export type AdminProjectCreateOptions = Omit<AdminProjectUpdateOptions, 'displayName'> & {
+  displayName: string,
+};
+export function adminProjectCreateOptionsToCrud(options: AdminProjectCreateOptions): InternalProjectsCrud["Server"]["Create"] {
+  return {
+    ...adminProjectUpdateOptionsToCrud(options),
+    display_name: options.displayName,
+  };
+}
